### 2025-06-24 #8474 @kyleniemeyer
Hello @kevin-tofu, we will use this issue to assign an editor and find reviewers. Unfortunately, we don't have any editors with the relevant expertise available to handle this at the moment, so I need to put this on our waitlist.

In the meantime, can you address the DOI error/warning above? Also, any suggestions for reviewers would be welcome, though please just share their info here and do not invite/ask them directly. Thanks!
