+++ paper/paper.md
@@ -16,7 +16,7 @@ bibliography: paper.bib
 ---
 
 
-## Summary
+# Summary
 
 Topology optimization is a computational method for optimizing the shape and material layout of structures and components, typically using the finite element method (FEM). It automatically generates optimal structural configurations by adjusting the distribution of material to achieve target performance—for example, minimizing compliance under a volume constraint. Specifically, the method optimizes an objective function by adjusting design variables that represent the presence or absence of material in each region.
 
@@ -28,7 +28,7 @@ Installation and usage instructions are available in the [GitHub repository](htt
 This article describes version **v0.2.0** of *scikit-topt*.
 
 
-## Statement of Need
+# Statement of Need
 
 There is a growing demand for accessible and flexible tools for topology optimization that support unstructured meshes and can be easily integrated into Python-based workflows.
 While many existing libraries focus on specific problems with rigid architectures or heavy dependencies, Scikit-topt offers a lightweight, modular, and fully Pythonic framework suitable for both academic research and industrial applications.
@@ -40,10 +40,10 @@ Scikit-topt addresses this gap by enabling users to track and visualize the prog
 The project is under active development and welcomes community contributions.
 
 
-## Purpose and Prior Art
+# Purpose and Prior Art
 
 There are numerous open-source projects for topology optimization.
-Examples include FEniTop[@jia_fenitop_2024], TopOpt.jl[@tarek2019topoptjl], Topology optimization using PETSc[@Smit2021], DL4TO[@dl4to2023], and pytopo3d[@kim2025pytopo3d].
+Examples include FEniTop[@Jia2024], TopOpt.jl[@tarek2019topoptjl], Topology optimization using PETSc[@Smit2021], DL4TO[@Erzmann2023], and pytopo3d[@kim2025pytopo3d].
 However, many projects are specialized implementations targeted at specific problems and are not well-suited for solving general topology optimization tasks. Common limitations include:
 
 * Parameters and problem settings are hard-coded, limiting flexibility.
@@ -51,10 +51,10 @@ However, many projects are specialized implementations targeted at specific prob
 * They are not packaged properly or depend on outdated software, making installation and usage difficult.
 * They are not designed as standard experimental platforms, making it hard to compare algorithms across different parameters and case studies.
 
-Scikit-topt avoids these issues by not including compiled code and depending only on Scipy[@scipy], a widely used and stable Python library, making packaging and installation straightforward.
+Scikit-topt avoids these issues by not including compiled code and depending only on Scipy[@Virtanen2020], a widely used and stable Python library, making packaging and installation straightforward.
 For finite element matrix assembly, it uses Scikit-FEM[@skfem], which also contains no compiled components.
 
-Furthermore, Scikit-topt visualizes parameter transitions and material density evolution through Matplotlib[@hunter2007] and PyVista[@sullivan2019pyvista].
+Furthermore, Scikit-topt visualizes parameter transitions and material density evolution through Matplotlib[@Hunter2007] and PyVista[@Sullivan2019].
 
 Scikit-topt implements topology optimization algorithms based on the density method, and currently supports the following features:
 
@@ -67,8 +67,8 @@ Scikit-topt implements topology optimization algorithms based on the density met
 * Optimization under multiple load cases
 * Continuation for scheduling parameter transitions
 * Includes FEA solvers powered by:
-  - Scipy[@scipy]
-  - PyAMG[@pyamg]
+  - Scipy[@Virtanen2020]
+  - PyAMG[@Bell2023]
   - scikit-fem[@skfem]
 * Support for multiple optimization algorithms
   - Optimality Criteria Method
@@ -78,15 +78,11 @@ Scikit-topt implements topology optimization algorithms based on the density met
   - Lagrangian Method
 
 
-## Usage Example
+# Usage Example
 
[…]
