+++ scikit-topt/sktopt/core/__init__.py
@@ -8,7 +8,8 @@ from sktopt.core.optimizers.loglagrangian import LogLagrangian_Config
 from sktopt.core.optimizers.loglagrangian import LogLagrangian_Optimizer
 from sktopt.core.optimizers.linearlagrangian import LinearLagrangian_Config
 from sktopt.core.optimizers.linearlagrangian import LinearLagrangian_Optimizer
-
+# from sktopt.core.optimizers.evo import Evolutionary_Config
+# from sktopt.core.optimizers.evo import Evolutionary_Optimizer
 
 __all__ = [
     "OC_Config",
@@ -20,5 +21,7 @@ __all__ = [
     "LogLagrangian_Config",
     "LogLagrangian_Optimizer",
     "LinearLagrangian_Config",
-    "LinearLagrangian_Optimizer"
+    "LinearLagrangian_Optimizer",
+    # "Evolutionary_Config",
+    # "Evolutionary_Optimizer"
 ]
