+++ scikit-topt/sktopt/core/optimizers/common.py
@@ -18,6 +18,7 @@ from sktopt.tools.logconf import mylogger
 logger = mylogger(__name__)
 
 
+
 @dataclass
 class DensityMethodConfig():
     """
@@ -99,14 +100,6 @@ class DensityMethodConfig():
     rho_max : float
         Maximum density value (usually 1.0 for full material).
 
-    move_limit_init : float
-        Initial move limit for density update.
-    move_limit : float
-        Final move limit.
-    move_limit_step : int
-        Number of continuation steps for move limit.
-        If negative, move limit is fixed.
-
     restart : bool
         Whether to resume optimization from an existing state.
     restart_from : int
@@ -119,11 +112,6 @@ class DensityMethodConfig():
 
     design_dirichlet : bool
         If True, Dirichlet boundary elements are included in the design domain.
-
-    lambda_lower : float
-        Lower bound for the Lagrange multiplier in constrained optimization.
-    lambda_upper : float
-        Upper bound for the Lagrange multiplier.
 
     sensitivity_filter : bool
         If True, applies filtering to the sensitivity field.
@@ -154,33 +142,24 @@ class DensityMethodConfig():
     beta_curvature: float = 2.0
     beta_eta: float = 0.50
     eta: float = 0.6
-    percentile_init: float = 60
-    percentile: float = -90
-    percentile_step: int = -1
     filter_radius_init: float = 2.0
     filter_radius: float = 1.20
     filter_radius_step: int = -3
-    # mu_p: float = 200.0
     E0: float = 210e9
     E_min: float = 210e6
     rho_min: float = 1e-2
     rho_max: float = 1.0
-    move_limit_init: float = 0.3
-    move_limit: float = 0.10
-    move_limit_step: int = -3
     restart: bool = False
     restart_from: int = -1
     export_img: bool = False
     export_img_opaque: bool = False
     design_dirichlet: bool = False
-    lambda_lower: float = 1e-7
-    lambda_upper: float = 1e+7
     sensitivity_filter: bool = False
     solver_option: Literal["spsolve", "cg_pyamg"] = "spsolve"
     scaling: bool = False
 
     @classmethod
-    def from_defaults(cls, **args):
+    def from_defaults(cls, **args) -> 'DensityMethodConfig':
         sig = inspect.signature(cls)
         valid_keys = sig.parameters.keys()
         filtered_args = {k: v for k, v in args.items() if k in valid_keys}
@@ -217,6 +196,50 @@ class DensityMethodConfig():
             return f"{self.dst_path}/mesh_rho/info_{prefix}-{iter:08d}.jpg"
         else:
             return None
+
+
+@dataclass
+class DensityMethodLagrangianConfig(DensityMethodConfig):
+    """
+    Configuration class for controlling parameters in topology optimization.
+
+    This class defines optimization parameters, filtering and projection settings,
+    continuation schedules, and numerical solver options. It is designed to support
+    sensitivity-based topology optimization algorithms such as SIMP and RAMP.
+
+    Attributes
+    ----------
+
+    percentile_init : float
+        Initial percentile used to scale sensitivity fields.
+    percentile : float
+        Final percentile value. If set to a negative value,
+        percentile-based scaling is disabled, resulting in a
+        more "exact" optimization behavior.
+    percentile_step : int
+        Number of continuation steps for the percentile value.
+        If negative, fixed at percentile_init.
+    move_limit_init : float
+        Initial move limit for density update.
+    move_limit : float
+        Final move limit.
+    move_limit_step : int
+        Number of continuation steps for move limit.
+        If negative, move limit is fixed.
[…]
