+++ scikit-topt/sktopt/core/optimizers/linearlagrangian.py
@@ -7,7 +7,7 @@ logger = mylogger(__name__)
 
 
 @dataclass
-class LinearLagrangian_Config(common.DensityMethodConfig):
+class LinearLagrangian_Config(common.DensityMethodLagrangianConfig):
     """
     Configuration for Log-space Lagrangian Gradient Update method.
 
@@ -171,6 +171,7 @@ class LinearLagrangian_Optimizer(common.DensityMethod):
         assert cfg.lambda_lower < 0
         assert cfg.lambda_upper > 0
         super().__init__(cfg, tsk)
+        self.recorder = self.add_recorder(tsk)
         self.recorder.add("dC", plot_type="min-max-mean-std")
         self.recorder.add("lambda_v", ylog=False)
         self.lambda_v = cfg.lambda_v
