### 2025-07-02 b3be022
add Parallel solver with joblib
