### 2025-07-04 e792a93
add abstract class
