+++ scikit-topt/sktopt/core/misc.py
@@ -148,4 +148,8 @@ def add_common_arguments(
     parser.add_argument(
         '--solver_option', '-SO', type=str, default="spsolve", help=''
     )
+    parser.add_argument(
+        '--n_joblib', '-NJ', type=int, default=2, help=''
+    )
+    
     return parser
