+++ pyproject.toml
@@ -9,7 +9,7 @@ license = { text = "Apache-2.0" }
 readme = "README.md"
 requires-python = ">=3.11,<3.14"
 dependencies = [
-    "scikit-fem (>=10.0.2,<11.0.0)",
+    "scikit-fem (>=11.0,<12.0)",
     "matplotlib (>=3.10.1,<4.0.0)",
     "meshio (>=5.3.5,<6.0.0)",
     "numba (>=0.61.0,<0.62.0)",
