### 2025-07-11 c7f8962
remove redundant print and bugfix
