+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "scikit-topt"
-version = "0.2.2"
+version = "0.2.3"
 description = "A lightweight, flexible Python library for topology optimization built on top of Scikit Libraries"
 authors = [
     { name = "Kevin-Tofu", email = "koheitech001@gmail.com" }
