+++ README.md
@@ -62,7 +62,7 @@ If you use Scikit Topt in your research or software, please cite it as:
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.15441499},
   url          = {https://doi.org/10.5281/zenodo.15441499},
-  note         = {Version 0.2.3}
+  note         = {Version 0.2.4}
 }
 ```
 
