+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "scikit-topt"
-version = "0.2.3"
+version = "0.2.4"
 description = "A lightweight, flexible Python library for topology optimization built on top of Scikit Libraries"
 authors = [
     { name = "Kevin-Tofu", email = "koheitech001@gmail.com" }
@@ -23,7 +23,7 @@ dependencies = [
 
 [tool.poetry]
 name = "scikit-topt"
-version = "0.2.3"
+version = "0.2.4"
 description = "A lightweight, flexible Python library for topology optimization built on top of Scikit Libraries"
 authors = ["Kevin-Tofu <koheitech001@gmail.com>"]
 readme = "README.md"
