+++ scikit-topt/sktopt/core/__init__.py
@@ -2,8 +2,8 @@ from sktopt.core.optimizers.oc import OC_Config
 from sktopt.core.optimizers.oc import OC_Optimizer
 from sktopt.core.optimizers.logmoc import LogMOC_Config
 from sktopt.core.optimizers.logmoc import LogMOC_Optimizer
-# from sktopt.core.optimizers.linearmoc import LinearMOC_Config
-# from sktopt.core.optimizers.linearmoc import LinearMOC_Optimizer
+from sktopt.core.optimizers.linearmoc import LinearMOC_Config
+from sktopt.core.optimizers.linearmoc import LinearMOC_Optimizer
 from sktopt.core.optimizers.loglagrangian import LogLagrangian_Config
 from sktopt.core.optimizers.loglagrangian import LogLagrangian_Optimizer
 from sktopt.core.optimizers.linearlagrangian import LinearLagrangian_Config
@@ -16,8 +16,8 @@ __all__ = [
     "OC_Optimizer",
     "LogMOC_Config",
     "LogMOC_Optimizer",
-    # "LinearMOC_Config",
-    # "LinearMOC_Optimizer",
+    "LinearMOC_Config",
+    "LinearMOC_Optimizer",
     "LogLagrangian_Config",
     "LogLagrangian_Optimizer",
     "LinearLagrangian_Config",
