### 2025-07-23 #8474 @aslan-ng
Hi all,

I’d be happy to serve as a reviewer for this submission. I’m a Mechanical Engineer with experience in simulation, and I’m also familiar with topology optimization. Please let me know if I can be of help.

Best,
Aslan
