+++ docs/optimization.rst
@@ -17,201 +17,158 @@ Optimality Criteria
 Optimality Criteria Method (OC)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-The Optimality Criteria (OC) method is a widely used algorithm in density-based topology optimization due to its simplicity and effectiveness. It provides a closed-form update rule for the design variables based on the optimality conditions derived from the Lagrangian of the problem.
+The Optimality Criteria (OC) method is a widely used algorithm in density-based topology optimization due to its simplicity and effectiveness. 
+It provides a closed-form multiplicative update rule for the design variables, derived from the Karush–Kuhn–Tucker (KKT) optimality conditions of the compliance minimization problem.
 
 Assume the goal is to minimize compliance :math:`C(\rho)` subject to a volume constraint:
 
 .. math::
 
    \min_{\rho} \quad C(\rho) = \mathbf{f}^T \mathbf{u}(\rho) \\
-   \text{subject to} \quad \sum_e \rho_e v_e \leq V_{\text{max}}, \quad \rho_{\text{min}} \leq \rho_e \leq 1
+   \text{subject to} \quad \sum_e \rho_e v_e \leq V_{\text{max}}, 
+   \quad \rho_{\text{min}} \leq \rho_e \leq 1.
 
-The KKT conditions yield an optimality criterion involving the derivative of the objective function and a Lagrange multiplier :math:`\lambda`. Based on this, the OC method defines the update rule for each element :math:`e` as:
+The KKT conditions lead to an element-wise update rule involving the compliance sensitivity and a Lagrange multiplier :math:`\lambda`:
 
 .. math::
 
    \rho_e^{(t+1)} = \text{clip}\left(
-   \rho_e^{(t)} \cdot \left( \frac{-\partial C / \partial \rho_e}{\lambda} \right)^{\eta},\ 
+   \rho_e^{(t)} \cdot 
+   \left( \frac{-\partial C / \partial \rho_e}{\lambda} \right)^{\eta},\ 
    \rho_{\text{min}},\ 1
-   \right)
+   \right).
 
-where:
+Here:
 
 - :math:`\partial C / \partial \rho_e` is the sensitivity of compliance with respect to the design variable,
-- :math:`\lambda` is chosen to satisfy the volume constraint (e.g., by bisection),
+- :math:`\lambda` is a Lagrange multiplier chosen (typically by bisection) to satisfy the volume constraint,
 - :math:`\eta` is a numerical damping or scaling factor (commonly :math:`\eta = 1`),
-- `clip` ensures the updated value stays within allowable bounds.
+- ``clip`` ensures the updated density stays within prescribed bounds,
+- optional move limits may also be enforced to restrict the maximum per-iteration change in :math:`\rho_e`.
 
-The OC update rule is applied iteratively, and move limits may also be enforced to restrict the change in :math:`\rho_e` between iterations.
+The OC method requires only element-wise sensitivities and a scalar multiplier update, 
+and therefore avoids reliance on general-purpose gradient-based optimizers such as MMA or SQP.
 
 Advantages
 ^^^^^^^^^^
 
 - **Simplicity**: Easy to implement and understand.
 - **Efficiency**: Fast convergence in practice for compliance minimization.
-- **No gradient solver needed**: Only objective sensitivities are required, not the full gradient descent machinery.
+- **Closed-form updates**: Only sensitivities and a scalar Lagrange multiplier search are required, no external optimization solver.
 
 Disadvantages
 ^^^^^^^^^^^^^
 
-- **Not general-purpose**: Hard to extend to problems with multiple constraints or noncompliance objectives.
[…]
