+++ README.md
@@ -45,7 +45,7 @@ The currently supported features are as follows:
   - Optimality Criteria (OC) Method  
   - (Log-Space) Modified OC Method 
   - Lagrangian Method
-- able to handle multiple objectives / constraints
+- able to handle multiple force condition
 - High-performance computation using sparse matrices with Scipy and PyAMG  
 - has a function to monitor the transition of parameters.
 
