+++ docs/optimization.rst
@@ -127,48 +127,3 @@ This is not part of the classical OC method but can improve stability in practic
 - Requires careful tuning of :math:`\eta` and the handling of :math:`\lambda_v`.
 - Volume constraint is not enforced exactly unless :math:`\lambda_v` is solved consistently (e.g., via bisection).
 - Convergence may be sensitive to filtering and parameter settings.
-
-
-2. Linear-Space Update Method
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-An alternative to the multiplicative OC update is to use a **linear (additive) update** rule in the physical density domain. 
-Here, the design variables are updated by adding an explicit increment :math:`\Delta \rho` to the current density:
-
-.. math::
-
-   \rho_e^{(t+1)} = \rho_e^{(t)} + \Delta \rho_e,
-
-with the increment defined as
-
-.. math::
-
-   \Delta \rho_e = -\eta \, \left( \frac{\partial C}{\partial \rho_e} + \lambda_v \frac{\partial g}{\partial \rho_e} \right),
-
-where:
-
-- :math:`\partial C / \partial \rho_e` is the compliance sensitivity,
-- :math:`g(\rho)` is the volume constraint function,
-- :math:`\lambda_v` is a constraint-related parameter (dual variable in a Lagrangian formulation, or a penalty weight in a penalty-based approximation),
-- :math:`\eta` is a step size (learning rate).
-
-In practice, the update increment :math:`\Delta \rho_e` is often **clipped** to enforce move limits or bound constraints:
-
-.. math::
-
-   \rho_e^{(t+1)} = \text{clip}\!\left(\rho_e^{(t)} + \Delta \rho_e,\; \rho_{\min},\; 1\right).
-
-This formulation resembles a gradient-descent step and is structurally simpler than multiplicative OC, 
-which makes it more flexible for integration with additional constraint-handling or stabilization techniques.
-
-**Advantages**:
-
-- Direct control over the update magnitude through :math:`\eta`.
-- Easier to integrate with projections, filters, or move-limit strategies.
-- Conceptually simple and convenient for algorithmic experimentation.
-
-**Disadvantages**:
-
-- Positivity of :math:`\rho` is not guaranteed unless explicitly enforced by clipping.
-- Volume constraints are only approximately satisfied unless post-processing or multiplier updates are applied.
-- Sensitive to parameter tuning; damping or stabilization is usually required for robust performance.
