### 2025-09-26 443c51e
eliminate wrong optimizers
