+++ scikit-topt/sktopt/core/__init__.py
@@ -2,12 +2,12 @@ from sktopt.core.optimizers.oc import OC_Config
 from sktopt.core.optimizers.oc import OC_Optimizer
 from sktopt.core.optimizers.logmoc import LogMOC_Config
 from sktopt.core.optimizers.logmoc import LogMOC_Optimizer
-from sktopt.core.optimizers.linearmoc import LinearMOC_Config
-from sktopt.core.optimizers.linearmoc import LinearMOC_Optimizer
-from sktopt.core.optimizers.loglagrangian import LogLagrangian_Config
-from sktopt.core.optimizers.loglagrangian import LogLagrangian_Optimizer
-from sktopt.core.optimizers.linearlagrangian import LinearLagrangian_Config
-from sktopt.core.optimizers.linearlagrangian import LinearLagrangian_Optimizer
+# from sktopt.core.optimizers.linearmoc import LinearMOC_Config
+# from sktopt.core.optimizers.linearmoc import LinearMOC_Optimizer
+# from sktopt.core.optimizers.loglagrangian import LogLagrangian_Config
+# from sktopt.core.optimizers.loglagrangian import LogLagrangian_Optimizer
+# from sktopt.core.optimizers.linearlagrangian import LinearLagrangian_Config
+# from sktopt.core.optimizers.linearlagrangian import LinearLagrangian_Optimizer
 # from sktopt.core.optimizers.evo import Evolutionary_Config
 # from sktopt.core.optimizers.evo import Evolutionary_Optimizer
 
@@ -16,12 +16,12 @@ __all__ = [
     "OC_Optimizer",
     "LogMOC_Config",
     "LogMOC_Optimizer",
-    "LinearMOC_Config",
-    "LinearMOC_Optimizer",
-    "LogLagrangian_Config",
-    "LogLagrangian_Optimizer",
-    "LinearLagrangian_Config",
-    "LinearLagrangian_Optimizer",
+    # "LinearMOC_Config",
+    # "LinearMOC_Optimizer",
+    # "LogLagrangian_Config",
+    # "LogLagrangian_Optimizer",
+    # "LinearLagrangian_Config",
+    # "LinearLagrangian_Optimizer",
     # "Evolutionary_Config",
     # "Evolutionary_Optimizer"
 ]
