+++ scikit-topt/sktopt/core/optimizers/linearlagrangian.py
@@ -1,286 +1,0 @@
-from dataclasses import dataclass
-import numpy as np
-import sktopt
-from sktopt.core.optimizers import common
-from sktopt.tools.logconf import mylogger
-logger = mylogger(__name__)
-
-
-@dataclass
-class LinearLagrangian_Config(common.DensityMethodLagrangianConfig):
-    """
-    Configuration for Log-space Lagrangian Gradient Update method.
-
-    This class defines the configuration parameters for performing topology optimization
-    via gradient-based updates in the logarithmic domain. Unlike traditional Optimality Criteria (OC)
-    methods, this approach explicitly follows the gradient of the Lagrangian
-    (compliance + volume penalty) and applies the update in log-space to ensure
-    positive densities and multiplicative-like behavior.
-
-    Attributes
-    ----------
-    mu_p : float
-        Weighting factor applied to the volume constraint in the Lagrangian.
-        This term scales the influence of the volume penalty in the descent direction.
-
-    lambda_v : float
-        Initial value for the Lagrange multiplier associated with the volume constraint.
-        This is added directly to the compliance gradient to form the full Lagrangian gradient.
-
-    lambda_decay : float
-        Decay factor applied to lambda_v over iterations, allowing gradual tuning
-        of constraint strength.
-
-    lambda_lower : float
-        Minimum allowed value for the Lagrange multiplier. Can be negative in this formulation,
-        since lambda_v is added to the gradient rather than used as a ratio.
-
-    lambda_upper : float
-        Maximum allowed value for the Lagrange multiplier. Clamping helps avoid instability
-        in the update steps due to large penalties.
-
-    Notes
-    -----
-    This method is sometimes referred to as 'EUMOC', but it is mathematically distinct
-    from classical OC-based updates. It performs gradient descent on the Lagrangian
-    in log(ρ)-space, leading to multiplicative behavior while maintaining gradient fidelity.
-
-    """
-    mu_p: float = 5.0
-    lambda_v: float = 0.1
-    lambda_decay: float = 0.90
-    lambda_lower: float = -1e+7
-    lambda_upper: float = 1e+7
-
-
-# log(x) = -0.4   →   x ≈ 0.670
-# log(x) = -0.3   →   x ≈ 0.741
-# log(x) = -0.2   →   x ≈ 0.819
-# log(x) = -0.1   →   x ≈ 0.905
-# log(x) =  0.0   →   x =  1.000
-# log(x) = +0.1   →   x ≈ 1.105
-# log(x) = +0.2   →   x ≈ 1.221
-# log(x) = +0.3   →   x ≈ 1.350
-# log(x) = +0.4   →   x ≈ 1.492
-
-
-def kkt_update(
-    rho,
-    dC, lambda_v, scaling_rate,
-    eta, move_limit,
-    tmp_lower, tmp_upper,
-    rho_min: float, rho_max: float,
-    percentile: float,
-    interpolation: str
-):
-    """
-    In-place version of the modified OC update (log-space),
-    computing dL = dC + lambda_v inside the function.
-
-    Parameters:
-    - rho: np.ndarray, design variables (will be updated in-place)
-    - dC: np.ndarray, compliance sensitivity (usually negative)
-    - lambda_v: float, Lagrange multiplier for volume constraint
-    - move: float, maximum allowed change per iteration
-    - eta: float, learning rate
-    - rho_min: float, minimum density
-    - rho_max: float, maximum density
-    - tmp_lower, tmp_upper, scaling_rate: work arrays (same shape as rho)
-    """
-
-    np.copyto(scaling_rate, dC)
[…]
