### 2025-09-27 df35e81
eliminate removed module api
