+++ docs/api.rst
@@ -9,6 +9,5 @@ This page provides detailed documentation for modules, classes and functions.
    core.optimizers.common
    core.optimizers.oc
    core.optimizers.logmoc
-   core.optimizers.loglagrangian
    mesh.task
    mesh.utils
