+++ README.md
@@ -143,14 +143,14 @@ optimizer.optimize()
 
 ### Optimize Toy Problem with command line.
 ```bash
-OMP_NUM_THREADS=3 OPENBLAS_NUM_THREADS=3 MKL_NUM_THREADS=3 PYTHONPATH=./scikit-topt python ./sktopt/core/optimizer/logmoc.py \
+OMP_NUM_THREADS=3 OPENBLAS_NUM_THREADS=3  MKL_NUM_THREADS=3 PYTHONPATH=scikit-topt python ./scikit-topt/sktopt/core/optimizers/logmoc.py \
  --dst_path ./result/base_moc_down \
  --interpolation SIMP \
  --vol_frac 0.40 \
  --eta 0.8 \
  --record_times 100 \
- --max_iters 600 \
- --mu_p 2000 \
+ --max_iters 300 \
+ --mu_p 5000 \
  --task_name down_box \
  --solver_option spsolve \
  --rho_min 1e-2 \
