+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -259,9 +259,6 @@ def interpolation_funcs(cfg: DensityMethodConfig):
         raise ValueError("Interpolation method must be SIMP or RAMP.")
 
 
-
-
-
 class DensityMethodBase(ABC):
     @abstractmethod
     def add_recorder(self):
@@ -302,17 +299,17 @@ class DensityMethodBase(ABC):
     @abstractmethod
     def rho_update(
         self,
-        iter_loop: int,
-        rho_candidate: np.ndarray,
+        iter_num: int,
+        rho_design_eles: np.ndarray,
         rho_projected: np.ndarray,
-        dC_drho_ave: np.ndarray,
+        dC_drho_design_eles: np.ndarray,
         u_dofs: np.ndarray,
-        strain_energy_ave: np.ndarray,
+        strain_energy_mean: np.ndarray,
         scaling_rate: np.ndarray,
         move_limit: float,
         eta: float,
-        tmp_lower: np.ndarray,
-        tmp_upper: np.ndarray,
+        rho_clip_lower: np.ndarray,
+        rho_clip_upper: np.ndarray,
         lambda_lower: float,
         lambda_upper: float,
         percentile: float,
@@ -321,7 +318,7 @@ class DensityMethodBase(ABC):
         vol_frac: float
     ):
         pass
-    
+
 
 class DensityMethod(DensityMethodBase):
     """
@@ -558,24 +555,30 @@ class DensityMethod(DensityMethodBase):
         return rho, iter_begin, iter_end
 
     def initialize_params(self):
+        """
+        To take advantage of in-place computation and improve execution speed 
+        by avoiding repeated memory allocations,
+        memory regions are preallocated and reused across iterations
+        """
         tsk = self.tsk
         cfg = self.cfg
         rho, iter_begin, iter_end = self.initialize_density()
         rho_prev = np.zeros_like(rho)
         rho_filtered = np.zeros_like(rho)
         rho_projected = np.zeros_like(rho)
-        dH = np.empty_like(rho)
+        dH_drho = np.empty_like(rho)
         grad_filtered = np.empty_like(rho)
         dC_drho_projected = np.empty_like(rho)
-        strain_energy_ave = np.zeros_like(rho)
+        strain_energy_mean = np.zeros_like(rho)
         dC_drho_full = np.zeros_like(rho)
-        dC_drho_ave = np.zeros_like(rho[tsk.design_elements])
+        dC_drho_design_eles = np.zeros_like(rho[tsk.design_elements])
         scaling_rate = np.empty_like(rho[tsk.design_elements])
-        rho_candidate = np.empty_like(rho[tsk.design_elements])
-        tmp_lower = np.empty_like(rho[tsk.design_elements])
-        tmp_upper = np.empty_like(rho[tsk.design_elements])
-        force_list = tsk.force if isinstance(tsk.force, list) else [tsk.force]
-        u_dofs = np.zeros((tsk.basis.N, len(force_list)))
+        rho_design_eles = np.empty_like(rho[tsk.design_elements])
+        rho_clip_lower = np.empty_like(rho[tsk.design_elements])
+        rho_clip_upper = np.empty_like(rho[tsk.design_elements])
+        force_vec_list = tsk.force \
+            if isinstance(tsk.force, list) else [tsk.force]
+        u_dofs = np.zeros((tsk.basis.N, len(force_vec_list)))
         filter_radius = cfg.filter_radius_init \
             if cfg.filter_radius_step > 0 else cfg.filter_radius
         return (
@@ -585,17 +588,17 @@ class DensityMethod(DensityMethodBase):
             rho_prev,
             rho_filtered,
             rho_projected,
-            dH,
+            dH_drho,
             grad_filtered,
[…]
