### 2025-09-27 e070d45
update fraction printing
