### 2025-09-27 2b67f4e
add comment on init function
