+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -190,12 +190,12 @@ class DensityMethodConfig():
     #         data = yaml.safe_load(f)
     #     return OC_Config(**data)
 
-    def vtu_path(self, iter: int):
-        return f"{self.dst_path}/mesh_rho/info_mesh-{iter:08d}.vtu"
+    def vtu_path(self, iter_num: int):
+        return f"{self.dst_path}/mesh_rho/info_mesh-{iter_num:08d}.vtu"
 
-    def image_path(self, iter: int, prefix: str):
+    def image_path(self, iter_num: int, prefix: str):
         if self.export_img:
-            return f"{self.dst_path}/mesh_rho/info_{prefix}-{iter:08d}.jpg"
+            return f"{self.dst_path}/mesh_rho/info_{prefix}-{iter_num:08d}.jpg"
         else:
             return None
 
@@ -533,11 +533,11 @@ class DensityMethod(DensityMethodBase):
                 data = np.load(data_path)
                 iter_begin = cfg.restart_from + 1
             else:
-                iter, data_path = misc.find_latest_iter_file(
+                iter_num, data_path = misc.find_latest_iter_file(
                     f"{cfg.dst_path}/data"
                 )
                 data = np.load(data_path)
-                iter_begin = iter + 1
+                iter_begin = iter_num + 1
             iter_end = cfg.max_iters + 1
             self.recorder.import_histories()
             rho[tsk.design_elements] = data["rho_design_elements"]
@@ -556,9 +556,61 @@ class DensityMethod(DensityMethodBase):
 
     def initialize_params(self):
         """
-        To take advantage of in-place computation and improve execution speed 
-        by avoiding repeated memory allocations,
-        memory regions are preallocated and reused across iterations
+            Initialize parameters and preallocate work arrays for density-based topology 
+            optimization using the SIMP method.
+
+            This function prepares all arrays required for iterative optimization. 
+            By preallocating memory and reusing arrays in-place, repeated allocations 
+            are avoided and execution speed is improved.
+
+            Returns
+            -------
+            iter_begin : int
+                Starting iteration index.
+            iter_end : int
+                Final iteration index.
+            rho : ndarray of shape (n_elements,)
+                Current element-wise design density variables.
+            rho_prev : ndarray of shape (n_elements,)
+                Copy of densities from the previous iteration.
+            rho_filtered : ndarray of shape (n_elements,)
+                Densities after applying the density filter (e.g., Helmholtz filter).
+            rho_projected : ndarray of shape (n_elements,)
+                Densities after applying the projection function (e.g., Heaviside projection).
+            dH_drho : ndarray of shape (n_elements,)
+                Derivative of the projection function with respect to the filtered densities.
+            grad_filtered : ndarray of shape (n_elements,)
+                Gradient field after combining sensitivity with projection derivative.
+            dC_drho_projected : ndarray of shape (n_elements,)
+                Compliance sensitivities with respect to the projected densities.
+            strain_energy_mean : ndarray of shape (n_elements,)
+                Average element strain energy over all load cases.
+            dC_drho_full : ndarray of shape (n_elements,)
+                Compliance sensitivities mapped back to the full set of elements.
+            dC_drho_design_eles : ndarray of shape (n_design_elements,)
+                Compliance sensitivities restricted to design elements only.
+            scaling_rate : ndarray of shape (n_design_elements,)
+                Scaling factors used in the update scheme (e.g., MOC/OC update).
+            rho_design_eles : ndarray of shape (n_design_elements,)
+                Subset of densities corresponding to design elements only.
[…]
