+++ scikit-topt/sktopt/fea/composer.py
@@ -478,152 +478,6 @@ if __name__ == '__main__':
     import pyvista as pv
     from memory_profiler import profile
 
-    @profile
-    def test_1():
-        from sktopt.mesh import toy_problem
-
-        tsk = toy_problem.toy_msh("plate-0.2.msh")
-        rho = np.ones(tsk.all_elements.shape)
-
-        K1 = assemble_stiffness_matrix(
-            tsk.basis, rho, tsk.E0, 0.0, 1.0, tsk.nu0
-        )
-        lam, mu = lame_parameters(tsk.E0, tsk.nu0)
-
-        def C(T):
-            return 2. * mu * T + lam * eye(trace(T), T.shape[0])
-
-        @skfem.BilinearForm
-        def stiffness(u, v, w):
-            return ddot(C(sym_grad(u)), sym_grad(v))
-
-        _F = tsk.force
-        K2 = stiffness.assemble(tsk.basis)
-
-        # print("tsk.dirichlet_nodes", tsk.dirichlet_nodes)
-        K1_e, F1_e = skfem.enforce(K1, _F, D=tsk.dirichlet_nodes)
-        K2_e, F2_e = skfem.enforce(K2, _F, D=tsk.dirichlet_nodes)
-
-        U1_e = scipy.sparse.linalg.spsolve(K1_e, F1_e)
-        U2_e = scipy.sparse.linalg.spsolve(K2_e, F2_e)
-
-        print("U1_e:", np.average(U1_e))
-        print("U2_e:", np.average(U2_e))
-
-        sf = 1.0
-        m1 = tsk.mesh.translated(sf * U1_e[tsk.basis.nodal_dofs])
-        m1.save('K1.vtk')
-        m2 = tsk.mesh.translated(sf * U2_e[tsk.basis.nodal_dofs])
-        m2.save('K2.vtk')
-
-        K1_e, F1_e = skfem.enforce(K1, _F, D=tsk.dirichlet_nodes)
-        # K1_e_np = K1_e.toarray()
-        U1_e = scipy.sparse.linalg.spsolve(K1_e, F1_e)
-        u = U1_e
-        K = K1_e.toarray()
-        U_global = 0.5 * u @ (K @ u)
-        print("Global:", U_global)
-
-        print(tsk.basis.element_dofs.shape, rho.shape)
-        U_elementwise1 = strain_energy_hdcode(
-            u, tsk.basis.element_dofs,
-            tsk.basis,
-            rho,
-            tsk.E0,
-            tsk.Emin,
-            1.0,
-            tsk.nu0,
-        ).sum()
-        element_dofs = tsk.basis.element_dofs[:, tsk.design_elements]
-        rho_design = rho[tsk.design_elements]
-        print(element_dofs.shape, rho_design.shape)
-
-        t0 = time.time()
-        U_elementwise2 = strain_energy_hdcode(
-            u, element_dofs,
-            tsk.basis,
-            rho_design,
-            tsk.E0,
-            tsk.Emin,
-            1.0,
-            tsk.nu0,
-        ).sum()
-        t1 = time.time()
-
-        U_elementwise2_numba1 = strain_energy_hdcode_numba(
-            u, element_dofs,
-            tsk.mesh.p,
-            rho_design,
-            tsk.E0,
-            tsk.Emin,
-            1.0,
-            tsk.nu0,
-        ).sum()
-        t2 = time.time()
-        U_elementwise2_numba2 = strain_energy_hdcode_numba(
-            u, element_dofs,
-            tsk.mesh.p,
-            rho_design,
-            tsk.E0,
-            tsk.Emin,
-            1.0,
-            tsk.nu0,
-        ).sum()
-        t3 = time.time()
-
-        print("numpy", t1 - t0)
-        print("numba-1st", t2 - t1)
[…]
