+++ _sources/core.optimizers.loglagrangian.rst.txt
@@ -1,8 +1,0 @@
-sktopt.core.optimizers.loglagrangian
-========================================
-
-.. automodule:: sktopt.core.optimizers.loglagrangian
-   :members:
-   :no-index:
-   :undoc-members:
-   :show-inheritance:
