+++ README.md
@@ -199,6 +199,7 @@ If you use Scikit Topt in your research or software, please cite it as:
 
 ## ToDo
 - Set break point from the optimization loop
+- Add A feature to assign tags to nodes and cells
 - Add other optimizers
   - Evolutionary Algorithms
   - Level Set
