### 2025-09-27 ac4a9a5
change way to assemble force vector, use LinearForm
