+++ scikit-topt/sktopt/mesh/task.py
@@ -1,7 +1,10 @@
-from typing import Literal
+from typing import Union, List, Literal
 from dataclasses import dataclass
+
 import numpy as np
 import skfem
+
+from skfem import FacetBasis, asm, LinearForm
 from scipy.spatial import cKDTree
 import matplotlib.pyplot as plt
 from sktopt.mesh import utils
@@ -16,6 +19,67 @@ def setdiff1d(a, b):
 
 _lit_bc = Literal['u^1', 'u^2', 'u^3', 'all']
 _lit_force = Literal['u^1', 'u^2', 'u^3']
+
+
+def assemble_surface_forces(
+    basis,
+    force_facets: Union[np.ndarray, List[np.ndarray]],
+    force_dir_type: Union[str, List[str]],
+    force_value: Union[float, List[float]],
+    *,
+    treat_value_as_total_force: bool = True,
+):
+    def _to_list(x):
+        return x if isinstance(x, list) else [x]
+
+    def _dir_to_comp(s: str) -> int:
+        if not (isinstance(s, str) and s.startswith('u^') and s[2:].isdigit()):
+            raise ValueError(f"force_dir_type must be like 'u^1','u^2','u^3', got: {s}")
+        c = int(s[2:]) - 1
+        if c < 0:
+            raise ValueError(f"Invalid component index parsed from {s}")
+        return c
+
+    facets_list = _to_list(force_facets)
+    dirs_list = _to_list(force_dir_type)
+    vals_list = _to_list(force_value)
+
+    if not (len(facets_list) == len(dirs_list) == len(vals_list)):
+        raise ValueError(
+            "Lengths of force_facets, force_dir_type, and force_value must match when lists."
+        )
+
+    @LinearForm
+    def l_one(v, w):
+        return 1.0
+
+    F_list = []
+
+    for facets, dir_s, val in zip(facets_list, dirs_list, vals_list):
+        comp = _dir_to_comp(dir_s)
+        fb = FacetBasis(basis.mesh, basis.elem, facets=np.asarray(facets, dtype=int))
+
+        A_arr = asm(l_one, fb)
+        A = float(np.sum(A_arr))  # sum up all contributions to get scalar area
+
+        if A <= 0.0:
+            raise ValueError(
+                "Selected facets have zero total area; check facet indices or geometry."
+            )
+
+        if treat_value_as_total_force:
+            p = float(val) / A
+        else:
+            p = float(val)
+
+        @LinearForm
+        def l_comp(v, w):
+            return p * v[comp]
+
+        F = asm(l_comp, fb)
+        F_list.append(F)
+
+    return F_list[0] if (len(F_list) == 1) else F_list
 
 
 @dataclass
@@ -41,9 +105,6 @@ class TaskConfig():
         Degrees of freedom constrained by Dirichlet (displacement) boundary conditions.
     dirichlet_elements : np.ndarray
         Elements that contain Dirichlet boundary points.
-    force_dofs : np.ndarray or list of np.ndarray
-        Degrees of freedom where external forces are applied.
-        Can be a list for multiple load cases.
     force_elements : np.ndarray
         Elements that contain the force application points.
     force : np.ndarray or list of np.ndarray
@@ -73,7 +134,6 @@ class TaskConfig():
     dirichlet_dofs: np.ndarray
     dirichlet_elements: np.ndarray
     force_nodes: np.ndarray | list[np.ndarray] 
-    force_dofs: np.ndarray | list[np.ndarray]
     force_elements: np.ndarray
     force: np.ndarray | list[np.ndarray]
     design_elements: np.ndarray
@@ -88,7 +148,6 @@ class TaskConfig():
     def mesh(self):
[…]
