### 2025-09-29 #9092 @editorialbot
[REVIEW]: Scikit-Topt: A Python Library for Algorithm Development in Topology Optimization
**Submitting author:** @kevin-tofu (<a href="http://orcid.org/0009-0008-2278-6418">Kohei Watanabe</a>)
**Repository:** https://github.com/kevin-tofu/scikit-topt
**Branch with paper.md** (empty if default branch): master
**Version:** 0.3.8
