### 2025-09-29 53d225f
update from_facets classmethod
