+++ scikit-topt/sktopt/mesh/task.py
@@ -23,7 +23,7 @@ _lit_force = Literal['u^1', 'u^2', 'u^3']
 
 def assemble_surface_forces(
     basis,
-    force_facets: Union[np.ndarray, List[np.ndarray]],
+    force_facets_ids: Union[np.ndarray, List[np.ndarray]],
     force_dir_type: Union[str, List[str]],
     force_value: Union[float, List[float]],
     *,
@@ -40,13 +40,13 @@ def assemble_surface_forces(
             raise ValueError(f"Invalid component index parsed from {s}")
         return c
 
-    facets_list = _to_list(force_facets)
+    facets_list = _to_list(force_facets_ids)
     dirs_list = _to_list(force_dir_type)
     vals_list = _to_list(force_value)
 
     if not (len(facets_list) == len(dirs_list) == len(vals_list)):
         raise ValueError(
-            "Lengths of force_facets, force_dir_type, and force_value must match when lists."
+            "Lengths of force_facets_ids, force_dir_type, and force_value must match when lists."
         )
 
     @LinearForm
@@ -154,13 +154,79 @@ class TaskConfig():
         E: float,
         nu: float,
         basis: skfem.Basis,
-        dirichlet_nodes: np.ndarray | list[np.ndarray],
+        dirichlet_nodes: np.ndarray,
         dirichlet_dir: _lit_bc | list[_lit_bc],
-        force_facets: np.ndarray | list[np.ndarray],
-        force_dir_type: str | list[str],
-        force_value: float | list[float],
+        force_nodes: np.ndarray,
+        force: np.ndarray | list[np.ndarray],
         design_elements: np.ndarray,
-    ):
+    ) -> 'TaskConfig':
+        """
+        Create a TaskConfig from material parameters and boundary-condition
+        specifications expressed by **node-index** sets.
+
+        This constructor:
+          - Resolves Dirichlet DOFs from `dirichlet_nodes` and `dirichlet_dir`.
+          - Identifies Dirichlet/Neumann (force) **elements** that touch those nodes.
+          - Builds load vector(s) from `force` and filters out non-designable elements.
+          - Derives sets such as free DOFs/elements and fixed/design elements.
+          - Precomputes per-element volumes.
+
+        Parameters
+        ----------
+        E : float
+            Young's modulus.
+        nu : float
+            Poisson's ratio.
+        basis : skfem.Basis
+            Finite-element space (provides mesh, DOFs, etc.).
+        dirichlet_nodes : np.ndarray or list[np.ndarray]
+            Global **node indices** that define Dirichlet boundaries.
+            If a list is provided, it must align one-to-one with `dirichlet_dir`.
+        dirichlet_dir : _lit_bc or list[_lit_bc]
+            Direction specifier(s) for Dirichlet constraints.  
+            `_lit_bc = Literal['u^1', 'u^2', 'u^3', 'all']`  
+            Use `'all'` to fix all components, or `'u^1'|'u^2'|'u^3'` to fix a
+            single component (compatible with `basis.get_dofs(...).nodal[...]`).
+            When `dirichlet_nodes` is a list, this must be a list of equal length.
+        force_nodes : np.ndarray or list[np.ndarray]
+            Global **node indices** where Neumann (external) loads are applied.
+            A list denotes multiple load regions (load cases).
+        force : np.ndarray or list[np.ndarray]
+            Load vector(s) associated with `force_nodes`. Provide a list to match
+            multiple regions/cases. Each vector should already be assembled in the
+            global DOF ordering expected by the solver.
+        design_elements : np.ndarray
+            Element indices initially considered designable. Elements touching
+            `force_nodes` are removed from this set.
+
+        Returns
+        -------
+        TaskConfig
+            A fully initialized configuration containing:
+              - material parameters (`E`, `nu`)
+              - mesh/basis
[…]
