### 2025-09-29 baf8a3d
update from_facets classmethod
