+++ scikit-topt/sktopt/mesh/task.py
@@ -45,6 +45,10 @@ def assemble_surface_forces(
     vals_list = _to_list(force_value)
 
     if not (len(facets_list) == len(dirs_list) == len(vals_list)):
+        
+        print("len(facets_list) : ", len(facets_list))
+        print("len(dirs_list) : ", len(dirs_list))
+        print("len(vals_list) : ", len(vals_list))
         raise ValueError(
             "Lengths of force_facets_ids, force_dir_type, and force_value must match when lists."
         )
@@ -380,9 +384,12 @@ class TaskConfig():
                 "dirichlet_facets_ids should be list[np.ndarray] or np.ndarray"
             )
 
-        force_facets_ids = np.concatenate(force_facets_ids) \
+        print("force_facets_ids: ", len(force_facets_ids), type(force_facets_ids))
+        force_facets_ids_concat = np.concatenate(force_facets_ids) \
             if isinstance(force_facets_ids, list) else force_facets_ids
-        force_nodes = np.unique(facets[:, force_facets_ids].ravel())
+        force_nodes = np.unique(facets[:, force_facets_ids_concat].ravel())
+
+        print("force_facets_ids: ", len(force_facets_ids), type(force_facets_ids))
         force = assemble_surface_forces(
             basis,
             force_facets_ids=force_facets_ids,
