### 2025-09-30 cb43c17
fix toy problem setting
