+++ scikit-topt/sktopt/mesh/toy_problem.py
@@ -196,8 +196,8 @@ def toy2():
     force_facets_1 = basis.mesh.facets_satisfying(
         in_range_1, boundaries_only=True
     )
-    print(f"force_facets_0 : {force_facets_0}")
-    print(f"force_facets_1 : {force_facets_1}")
+    # print(f"force_facets_0 : {force_facets_0}")
+    # print(f"force_facets_1 : {force_facets_1}")
     force_dir_type = ["u^2", "u^2"]
     force_value = [-100, 100]
     design_elements = utils.get_elements_in_box(
