+++ docs/tutorial.rst
@@ -67,7 +67,7 @@ Task Configuration
 
 .. code-block:: python
 
-    dirichlet_in_range = sktopt.utils.get_facets_in_range(
+    dirichlet_in_range = sktopt.mesh.utils.get_facets_in_range(
         (0.0, 0.05), (0.0, y_len), (0.0, z_len)
     )
     dirichlet_facets = basis.mesh.facets_satisfying(
@@ -76,10 +76,10 @@ Task Configuration
     dirichlet_dir = "all"
 
     eps = mesh_size
-    in_range_0 = utils.get_facets_in_range(
+    in_range_0 = sktopt.mesh.utils.get_facets_in_range(
         (x_len, x_len), (y_len-eps, y_len), (0, z_len)
     )
-    in_range_1 = utils.get_facets_in_range(
+    in_range_1 = sktopt.mesh.utils.get_facets_in_range(
         (x_len, x_len), (0, eps), (0, z_len)
     )
     force_facets_0 = basis.mesh.facets_satisfying(
@@ -90,7 +90,7 @@ Task Configuration
     )
     force_dir_type = ["u^2", "u^2"]
     force_value = [-100, 100]
-    design_elements = utils.get_elements_in_box(
+    design_elements = sktopt.meshutils.get_elements_in_box(
         mesh,
         (0.0, x_len), (0.0, y_len), (0.0, z_len)
     )
