+++ README.md
@@ -90,7 +90,7 @@ basis = skfem.Basis(mesh, e, intorder=2)
 ```Python
 
 # Specify Dirichlet Boundary Conditions
-dirichlet_in_range = sktopt.utils.get_facets_in_range(
+dirichlet_in_range = sktopt.mesh.utils.get_facets_in_range(
     (0.0, 0.03), (0.0, y_len), (0.0, z_len)
 )
 dirichlet_facets = basis.mesh.facets_satisfying(
@@ -100,7 +100,7 @@ dirichlet_dir = "all"
 
 # Define Force Vector
 
-in_range = sktopt.utils.get_facets_in_range(
+in_range = sktopt.mesh.utils.get_facets_in_range(
   (x_len, x_len),
   (y_len*2/5, y_len*3/5),
   (z_len*2/5, z_len*3/5)
