### 2025-09-30 10a9872
add config for scheduler
