### 2025-09-30 338a1ce
add config for scheduler
