+++ scikit-topt/sktopt/tools/__init__.py
@@ -1,4 +1,5 @@
 from sktopt.tools.history import HistoriesLogger
+from sktopt.tools.scheduler import SchedulerConfig
 from sktopt.tools.scheduler import Schedulers
 from sktopt.tools.scheduler import SchedulerStepAccelerating
 from sktopt.tools.scheduler import SchedulerStep
@@ -6,6 +7,7 @@ from sktopt.tools.scheduler import SchedulerSawtoothDecay
 
 __all__ = [
     "HistoriesLogger",
+    "SchedulerConfig",
     "Schedulers",
     "SchedulerStep",
     "SchedulerStepAccelerating",
