### 2025-09-30 cf06a5a
add SchedulerConfig, and update Config for Optimizers
