### 2025-09-30 c7ba070
add SchedulerConfig, and update Config for Optimizers
