### 2025-10-01 274f4aa
update SchedulerConfig
