+++ scikit-topt/sktopt/tools/history.py
@@ -321,4 +321,3 @@ class HistoriesLogger():
 
         # Replace histories with updated ones
         self.histories = new_histories
-

[…]
