### 2025-10-02 5df55b4
add Constant in Scheduler, and classmethod
