+++ scikit-topt/sktopt/tools/scheduler.py
@@ -13,7 +13,7 @@ logger = mylogger(__name__)
 
 def schedule_exp_slowdown(
     it: int, total: int,
-    start: float = 1.0, target: float = 0.4, rate: float = 10.0
+    initial_value: float = 1.0, target_value: float = 0.4, rate: float = 10.0
 ):
     if total <= 0:
         raise ValueError("total must be positive")
@@ -22,28 +22,58 @@ def schedule_exp_slowdown(
     decay = np.exp(-rate * t)
     final_decay = np.exp(-rate)
 
-    if start > target:
-        return target + \
-            (start - target) * (decay - final_decay) / (1 - final_decay)
+    if initial_value > target_value:
+        return target_value + \
+            (initial_value - target_value) * (decay - final_decay) / (1 - final_decay)
     else:
-        return target - \
-            (target - start) * (decay - final_decay) / (1 - final_decay)
+        return target_value - \
+            (target_value - initial_value) * (decay - final_decay) / (1 - final_decay)
 
 
 def schedule_exp_accelerate(
     it: int, total: int,
-    start: float = 1.0, target: float = 0.4, rate: float = 10.0
+    initial_value: float = 1.0, target_value: float = 0.4, rate: float = 10.0
 ):
     t = it / total
-    if start > target:
-        return target + (start - target) * (1 - np.exp(rate * (t - 1)))
+    if initial_value > target_value:
+        return target_value + (initial_value - target_value) * (1 - np.exp(rate * (t - 1)))
     else:
-        return target - (target - start) * (1 - np.exp(rate * (t - 1)))
+        return target_value - (target_value - initial_value) * (1 - np.exp(rate * (t - 1)))
+
+
+def schedule_constant(
+    it: int, total: int,
+    target_value: float = 0.4,
+    **args
+):
+    """
+    Step-function scheduler where each step value is used for
+    (approximately) equal number of iterations.
+
+    Parameters
+    ----------
+    it : int
+        Current iteration index.
+    total : int
+        Total number of iterations.
+    initial_value : float
+        Starting value.
+    target_value : float
+        Final target_value value.
+    num_steps : int
+        Number of discrete step values (including initial_value and target_value).
+
+    Returns
+    -------
+    float
+        Scheduled value for the given iteration.
+    """
+    return target_value
 
 
 def schedule_step(
     it: int, total: int,
-    start: float = 1.0, target: float = 0.4,
+    initial_value: float = 1.0, target_value: float = 0.4,
     num_steps: int = 10,
     **args
 ):
@@ -57,12 +87,12 @@ def schedule_step(
         Current iteration index.
     total : int
         Total number of iterations.
-    start : float
+    initial_value : float
         Starting value.
-    target : float
-        Final target value.
+    target_value : float
+        Final target_value value.
     num_steps : int
-        Number of discrete step values (including start and target).
+        Number of discrete step values (including initial_value and target_value).
 
     Returns
     -------
@@ -72,23 +102,23 @@ def schedule_step(
     if total <= 0:
         raise ValueError("total must be positive")
     if num_steps <= 1:
-        return target
+        return target_value
 
     # Determine which step this iteration belongs to
     step_length = total / num_steps
     step_index = min(int(it // step_length), num_steps - 1)
 
-    # Linearly divide values between start and target
+    # Linearly divide values between initial_value and target_value
     alpha = step_index / (num_steps - 1)
-    value = (1 - alpha) * start + alpha * target
+    value = (1 - alpha) * initial_value + alpha * target_value
     return value
 
 
 def schedule_step_accelerating(
     it: int,
     total: int,
-    start: float = 1.0,
-    target: float = 0.4,
[…]
