+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -25,108 +25,86 @@ logger = mylogger(__name__)
 @dataclass
 class DensityMethodConfig():
     """
-    Configuration class for controlling parameters in topology optimization.
+    Configuration for density-based topology optimization (SIMP/RAMP).
 
-    This class defines optimization parameters, filtering and projection settings,
-    continuation schedules, and numerical solver options. It is designed to support
-    sensitivity-based topology optimization algorithms such as SIMP and RAMP.
+    This configuration controls interpolation, continuation schedules via
+    `SchedulerConfig`, filtering/projection thresholds, and numerical options.
+    It targets sensitivity-based methods such as SIMP and RAMP.
+
+    Notes on continuation
+    ---------------------
+    Parameters that change during optimization are represented by
+    :class:`tools.SchedulerConfig`:
+      - ``p``: penalization power (e.g., 1.0 → 3.0)
+      - ``vol_frac``: volume fraction constraint (e.g., 0.8 → 0.4)
+      - ``beta``: Heaviside projection sharpness (e.g., 1.0 → 2.0)
+      - ``filter_radius``: filter radius (often constant)
+
+    Each scheduler defines how a value evolves (e.g., "Step", "StepAccelerating",
+    "Constant") and may include extra shape parameters (e.g., ``curvature`` for
+    accelerating schedules).
 
     Attributes
     ----------
     dst_path : str
-        Directory to which results and logs are saved.
+        Output directory for results and logs.
     interpolation : {"SIMP", "RAMP"}
-        Interpolation method for material penalization.
+        Material interpolation model for penalization.
     record_times : int
-        Number of snapshots to record during optimization.
+        Number of snapshots recorded during optimization.
     max_iters : int
         Maximum number of optimization iterations.
 
-    p_init : float
-        Initial penalization power in SIMP/RAMP interpolation.
-    p : float
-        Final value of the penalization power.
-    p_step : int
-        Number of steps to continue p from p_init to p.
-        If negative, p is fixed at p_init.
-
-    vol_frac_init : float
-        Initial volume fraction for continuation.
-    vol_frac : float
-        Final volume fraction constraint.
-    vol_frac_step : int
-        Number of continuation steps from vol_frac_init to vol_frac.
-        If negative, vol_frac is fixed at vol_frac_init.
-
-    beta_init : float
-        Initial sharpness parameter for Heaviside projection.
-    beta : float
-        Final beta value for projection.
-    beta_step : int
-        Number of continuation steps from beta_init to beta.
-        If negative, beta is fixed at beta_init.
-
-    beta_curvature : float
-        Controls the curvature of the Heaviside projection function.
     beta_eta : float
-        Threshold parameter for Heaviside projection; determines where
-        the projection function switches rapidly between 0 and 1.
-
+        Threshold parameter (η) for Heaviside projection; the density at which
+        the projection transitions sharply between 0 and 1.
     eta : float
-        Threshold density used in projection. Often used to define
-        intermediate region in Heaviside or filtering.
+        Threshold density used in projection or post-processing.
 
-    percentile_init : float
-        Initial percentile used to scale sensitivity fields.
-    percentile : float
-        Final percentile value. If set to a negative value,
-        percentile-based scaling is disabled, resulting in a
-        more "exact" optimization behavior.
-    percentile_step : int
-        Number of continuation steps for the percentile value.
-        If negative, fixed at percentile_init.
-
-    filter_radius_init : float
-        Initial radius of the sensitivity or density filter.
-    filter_radius : float
-        Final filter radius.
-    filter_radius_step : int
[…]
