### 2025-10-02 ab6609e
add factories and exception handlings
