+++ scikit-topt/sktopt/tools/scheduler.py
@@ -212,38 +212,168 @@ def schedule_sawtooth_decay(
 
 
 _lit_schedulers = Literal[
-    'Constant', 'Step', 'StepAccelerating', 'SawtoothDecay'
+    'Constant',
+    'Step', 'StepAccelerating', 'SawtoothDecay',
+    'None'
 ]
 
 
 @dataclass
-class SchedulerConfig():
-    name: str | None = None
-    init_value: float | None = None
-    target_value: float | None = None
-    num_steps: int | None = None
-    iters_max: int | None = None
-    curvature: float | None = None
+class SchedulerConfig:
+    """
+    Configuration for continuation and parameter scheduling.
+
+    Defines how a scalar parameter (e.g., penalization ``p``, projection
+    sharpness ``beta``, or volume fraction ``vol_frac``) evolves across
+    optimization iterations. Supports several scheduling strategies.
+
+    Attributes
+    ----------
+    name : str, optional
+        Identifier for the schedule (e.g., "p", "vol_frac").
+    init_value : float, optional
+        Starting value of the parameter.
+        - Required for Step, StepAccelerating, SawtoothDecay.
+        - Ignored if Constant.
+    target_value : float, optional
+        Final value of the parameter.
+        - Required for Step, StepAccelerating, SawtoothDecay.
+        - Used as fixed value if Constant.
+    num_steps : int, optional
+        Number of continuation steps or cycles.
+        - Step / StepAccelerating: number of increments between init and target.
+        - SawtoothDecay: number of sawtooth cycles (restarts).
+        - Ignored for Constant.
+    iters_max : int, optional
+        Maximum total number of iterations for which the schedule is defined.
+        - Used in SawtoothDecay to partition iterations into cycles.
+        - Typically equals the outer optimizer's max_iters.
+    curvature : float, optional
+        Shape parameter used in StepAccelerating.
+        - Example: curvature=2.0 accelerates change near the end.
+        - Ignored in other schedulers.
+    scheduler_type : {"Constant", "Step", "StepAccelerating", "SawtoothDecay"}
+        Scheduling strategy:
+          - **Constant**: fixed at ``target_value``.
+          - **Step**: discrete continuation from ``init_value`` → ``target_value`` in
+            ``num_steps`` increments.
+          - **StepAccelerating**: like Step but transition rate controlled by ``curvature``.
+          - **SawtoothDecay**: parameter decays linearly from ``init_value`` → ``target_value``
+            within each cycle (cycle length = iters_max/num_steps), then resets to
+            ``init_value`` at the start of the next cycle.
+    """
+
+    name: Optional[str] = None
+    init_value: Optional[float] = None
+    target_value: Optional[float] = None
+    num_steps: Optional[int] = None
+    iters_max: Optional[int] = None
+    curvature: Optional[float] = None
     scheduler_type: _lit_schedulers = "Constant"
-    # func: Callable = schedule_step
 
     @classmethod
     def from_defaults(
         cls,
-        name: str | None = None,
-        init_value: float | None = None,
-        target_value: float | None = None,
-        num_steps: int | None = None,
-        iters_max: int | None = None,
-        curvature: float | None = None,
+        name: Optional[str] = None,
+        init_value: Optional[float] = None,
+        target_value: Optional[float] = None,
+        num_steps: Optional[int] = None,
+        iters_max: Optional[int] = None,
+        curvature: Optional[float] = None,
         scheduler_type: _lit_schedulers = "Constant"
     ) -> 'SchedulerConfig':
+        """
+        Construct a :class:`SchedulerConfig` with validated defaults.
+
+        This helper ensures that each scheduler type receives the proper
+        parameters and fills in sensible defaults where possible.
+
+        Parameters
[…]
