### 2025-10-09 9540699
fix integration for area, wrong function
