### 2025-10-11 cd0c2b6
add from_mesh_tags classmethod
