+++ scikit-topt/sktopt/mesh/toy_problem.py
@@ -43,55 +43,9 @@ def create_box_tet(x_len, y_len, z_len, mesh_size):
     mesh = MeshTet().refined(n_refine)
     scale = np.array([x_len, y_len, z_len])
     mesh = mesh.scaled(scale)
-
     t_fixed = utils.fix_tetrahedron_orientation(mesh.t, mesh.p)
     mesh_fixed = MeshTet(mesh.p, t_fixed)
     return mesh_fixed
-
-
-def toy_base0(
-    mesh_size: float,
-    intorder: int = 2
-):
-    x_len = 8.0
-    y_len = 6.0
-    z_len = 4.0
-    eps = 1.2
-
-    # if True:
-    if False:
-        mesh = create_box_tet(x_len, y_len, z_len, mesh_size)
-        e = skfem.ElementVector(skfem.ElementTetP1())
-    else:
-        mesh = create_box_hex(x_len, y_len, z_len, mesh_size)
-        e = skfem.ElementVector(skfem.ElementHex1())
-    basis = skfem.Basis(mesh, e, intorder=intorder)
-    dirichlet_nodes = utils.get_nodes_indices_in_range(
-        basis.mesh, (0.0, 0.03), (0.0, y_len), (0.0, z_len)
-    )
-
-    F_nodes = utils.get_nodes_indices_in_range(
-        basis.mesh,
-        (x_len - eps, x_len+0.1), (y_len*2/5, y_len*3/5), (z_len-eps, z_len)
-    )
-    F_dofs = basis.get_dofs(nodes=F_nodes).nodal['u^3']
-    design_elements = utils.get_elements_in_box(
-        mesh,
-        (0.0, x_len), (0.0, y_len), (0.0, z_len)
-    )
-    E0 = 210e9
-    F = -100.0
-    return task.TaskConfig.from_defaults(
-        E0,
-        0.30,
-        basis,
-        dirichlet_nodes,
-        "all",
-        F_nodes,
-        F_dofs,
-        F,
-        design_elements
-    )
 
 
 def toy_base(
@@ -102,26 +56,26 @@ def toy_base(
     y_len = 6.0
     z_len = 4.0
     eps = 1.2
-
-    mesh = create_box_hex(x_len, y_len, z_len, mesh_size)
+    if False:
+        mesh = create_box_tet(x_len, y_len, z_len, mesh_size)
+        e = skfem.ElementVector(skfem.ElementTetP1())
+    else:
+        mesh = create_box_hex(x_len, y_len, z_len, mesh_size)
+        e = skfem.ElementVector(skfem.ElementHex1())
     e = skfem.ElementVector(skfem.ElementHex1())
     basis = skfem.Basis(mesh, e, intorder=intorder)
-    dirichlet_in_range = utils.get_facets_in_range(
+    dirichlet_in_range = utils.get_points_in_range(
         (0.0, 0.03), (0.0, y_len), (0.0, z_len)
     )
     dirichlet_facets = basis.mesh.facets_satisfying(dirichlet_in_range)
-    # dirichlet_nodes = utils.get_nodes_indices_in_range(
-    #     basis.mesh, (0.0, 0.03), (0.0, y_len), (0.0, z_len)
-    # )
-    in_range = utils.get_facets_in_range(
+    force_in_range = utils.get_points_in_range(
         (x_len - eps, x_len+0.1), (y_len*2/5, y_len*3/5), (z_len-eps, z_len)
     )
-    force_facets = basis.mesh.facets_satisfying(in_range)
-
-    design_elements = utils.get_elements_in_box(
-        mesh,
+    force_facets = basis.mesh.facets_satisfying(force_in_range)
+    desing_in_range = utils.get_points_in_range(
         (0.0, x_len), (0.0, y_len), (0.0, z_len)
[…]
