### 2025-10-11 6298b6f
update how to deal with force key
