### 2025-10-11 58108a1
update how to deal with force key
