### 2025-10-11 00b2a49
update from_mesh_tags classmethod
