### 2025-10-11 ff00624
update from_mesh_tags classmethod
