### 2025-10-16 #9092 @aslan-ng
Hi Kohei (@kevin-tofu),

Fantastic work on Scikit-Topt! The tool addresses an important challenge in topology optimization workflows outside of traditional CAD environments, and using only Python. I found the codebase clean, modular, and easy to follow. It’s evident that a lot of thought and care went into its architecture and documentation.
The integration of visualization, flexible mesh handling, and multiple optimization routines makes it a powerful and accessible platform for both research and education. It’s already a valuable toolkit, and I’m looking forward to seeing it grow and gain more adoption within the design and engineering community.

Best,
Aslan (@aslan-ng)
