### 2025-10-16 #9092 @aslan-ng
# Comments

1. A small enhancement [here](https://scikit-topt.readthedocs.io/en/latest/optimization.html#log-space-update-method) in docs:
“1. Log-space Update Method” -> “Log-space Update Method”

That “1. “ seems extra since there are no other numbers available. In the alternative method (Optimality Criteria Method) discussed [here](https://scikit-topt.readthedocs.io/en/latest/optimization.html#optimality-criteria-method-oc) no numbering has been used either.

2. A typo in README:
“Examples and Fieatures” -> “Examples and Features”

3. Image URLs in PyPI broken:
Currently, the images in the README uses a relative URL inside the img tag, e.g., `src="assets/ex-progress-report.jpg"` and `src="assets/ex-pull-down-0.gif”`. While this renders correctly on GitHub, relative paths will not resolve once the package is displayed on PyPI, leading to a broken image urls.

I recommend updating the link to use an absolute URL pointing to the raw file in the repository, like:
```
src=“https://media.githubusercontent.com/media/kevin-tofu/scikit-topt/master/assets/ex-progress-report.jpg”
```
And,
```
src=“https://media.githubusercontent.com/media/kevin-tofu/scikit-topt/master/assets/ex-pull-down-0.gif”
```

This small change will ensure the images displays correctly both on GitHub and on PyPI. That’s why, using absolute URLs in GitHub is generally encouraged.

4. In the installation section, the README lists:
```
pip install scikit-topt
poetry add scikit-topt
```
This could be interpreted as two consecutive commands to run, while they are actually alternative installation methods. To improve clarity, consider explicitly indicating that users should choose one of these methods. (It is properly described in Docs though)

5. The repository already includes some helpful examples in the README.md. At a minimum, I recommend converting those snippets into runnable scripts (or notebooks) collected in an `examples/` folder. This would let new users try things out directly, without needing to copy from the README, and would also make it easier to test the examples for reproducibility. Beyond that minimum, it would be even better if you could expand the `examples/` folder with a few additional demonstrations that highlight different use cases. For example, using different mesh types, different algorithms, …
This would provide quick, accessible entry points for first-time users while also showcasing the broader capabilities of the package. It is even possible to write a separate README for each example, e.g., `examples/example_1/README.md` to show the results, discuss the code, etc., if you wish.

6. I noticed the currently tests (in `tests/test_readme.py`) are structured as scripts with no assertions. To make them effective for CI and regression checking, please wrap them in an automated framework such as pytest or even unittest. This will make failures visible and reproducible, and aligns with JOSS requirements for automated testing.

7. When running `scikit-topt/tests/test_readme.py` file with Python 3.11, it ends with Traceback error as follows:
```
2025-10-16 11:08:53,621 [cell_basis.py:80] INFO     Initializing CellBasis(MeshHex1, ElementVector)
2025-10-16 11:08:53,624 [cell_basis.py:106] INFO     Initializing finished.
2025-10-16 11:08:53,624 [facet_basis.py:64] INFO     Initializing FacetBasis(MeshHex1, ElementVector)
2025-10-16 11:08:53,628 [facet_basis.py:116] INFO     Initializing finished.
2025-10-16 11:08:53,628 [__init__.py:91] INFO     Assembling 'l_one'.
2025-10-16 11:08:53,628 [__init__.py:96] INFO     Assembling finished.
2025-10-16 11:08:53,628 [__init__.py:91] INFO     Assembling 'l_comp'.
2025-10-16 11:08:53,629 [__init__.py:96] INFO     Assembling finished.
all_elements: (192,) design_elements: (176,) fixed_elements: (16,) dirichlet_force_elements: (40,) force_elements: [133 134 135 136 139 140 141 142 181 182 183 184 187 188 189 190]
preprocess : cg_pyamg
cg_pyamg
2025-10-16 11:08:53,916 [common_density.py:629] INFO     dst_path : ./result/pytests
./result/pytests/schedule-0-progress.jpg
2025-10-16 11:08:54,201 [common_density.py:661] INFO     iterations: 1 / 1
2025-10-16 11:08:54,201 [scheduler.py:654] INFO     p 3.000000
2025-10-16 11:08:54,201 [scheduler.py:654] INFO     vol_frac 0.800000
2025-10-16 11:08:54,201 [scheduler.py:654] INFO     beta 2.000000
2025-10-16 11:08:54,201 [scheduler.py:654] INFO     move_limit 0.100000
2025-10-16 11:08:54,201 [scheduler.py:654] INFO     eta 0.500000
2025-10-16 11:08:54,201 [scheduler.py:654] INFO     percentile None
2025-10-16 11:08:54,201 [scheduler.py:654] INFO     filter_radius 1.200000
2025-10-16 11:08:54,201 [common_density.py:684] INFO     --- project and filter ---
2025-10-16 11:08:54,201 [common_density.py:690] INFO     --- compute compliance ---
2025-10-16 11:08:54,228 [__init__.py:91] INFO     Assembling 'stiffness_form'.
2025-10-16 11:08:54,250 [__init__.py:96] INFO     Assembling finished.
2025-10-16 11:08:54,260 [common_density.py:737] INFO     --- update density ---
2025-10-16 11:08:54,260 [oc.py:205] INFO     λ: 4.0770e+02, vol_error: -0.0001, mean(rho): 0.7999
2025-10-16 11:08:54,260 [common_density.py:776] INFO     Saving at iteration 1
2025-10-16 11:08:54,260 [history.py:56] INFO     rho_projected: min=0.85258252, mean=0.85258252, max=0.85258252
2025-10-16 11:08:54,260 [history.py:56] INFO     strain_energy: min=0.00000000, mean=0.00000000, max=0.00000000
2025-10-16 11:08:54,260 [history.py:60] INFO     vol_error: -0.00009957
2025-10-16 11:08:54,260 [history.py:60] INFO     u_max: 0.00000000
2025-10-16 11:08:54,260 [history.py:60] INFO     compliance: 0.00000035
2025-10-16 11:08:54,260 [history.py:56] INFO     scaling_rate: min=0.80000000, mean=0.91462256, max=1.20000000
2025-10-16 11:08:54,260 [history.py:56] INFO     -dC: min=5.23432131, mean=366.58366259, max=1255.05298043
2025-10-16 11:08:54,260 [history.py:60] INFO     lmid: 407.69577036
Traceback (most recent call last):
  File "/Users/aslan/Projects/scikit-topt/scikit-topt/tests/test_readme.py", line 89, in <module>
    test_optimizers()
  File "/Users/aslan/Projects/scikit-topt/scikit-topt/tests/test_readme.py", line 79, in test_optimizers
    oc_optimize(tsk1)
  File "/Users/aslan/Projects/scikit-topt/scikit-topt/tests/test_readme.py", line 56, in oc_optimize
    optimizer.optimize()
  File "/Users/aslan/Projects/scikit-topt/scikit-topt/sktopt/core/optimizers/common_density.py", line 786, in optimize
    visualization.write_mesh_with_info_as_image(
  File "/Users/aslan/Projects/scikit-topt/scikit-topt/sktopt/core/visualization.py", line 112, in write_mesh_with_info_as_image
    pv.start_xvfb()
  File "/Users/aslan/.pyenv/versions/scikit-topt/lib/python3.11/site-packages/pyvista/plotting/utilities/xvfb.py", line 50, in start_xvfb
    raise OSError(XVFB_INSTALL_NOTES)
OSError: Please install Xvfb with:

Debian
$ sudo apt install libgl1-mesa-glx xvfb

CentOS / RHL
$ sudo yum install libgl1-mesa-glx xvfb
```

8. Python version metadata mismatch:
In `pyproject.toml`, you declare:
```
Requires: Python <3.14, >=3.11
```
That renders like this on PyPI:

But then list this classifier:
```
"Programming Language :: Python :: 3.11”
"Programming Language :: Python :: 3.12”
```
It renders like this on PyPI:

The classifiers above also render a badge in README.md that seems ike this:

However, the docs installation page states:
“sktopt is a Python package that can be installed using pip or poetry. It is compatible with Python 3.11 and later versions.”

This inconsistency may confuse users about which Python versions are actually supported.

9. Paper states: “The project is under active development and welcomes community contributions.”
This is excellent! You might also mention explicitly in the README or documentation that the project is open for pull requests and provide a short note on how contributors can get involved (e.g., “We welcome pull requests and discussions via GitHub Issues”). This small addition would make the project’s openness more visible and actionable for potential contributors. You can also remove it from the paper if you wish.

Please feel free to ask if you have any questions or need clarification on any of the comments.
