### 2025-10-17 7d2e272
fix documents/readme, and add test code for Python version 3.11-3.13
