+++ pyproject.toml
@@ -8,6 +8,12 @@ authors = [
 license = { text = "Apache-2.0" }
 readme = "README.md"
 requires-python = ">=3.11,<3.14"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+]
 dependencies = [
     "scikit-fem (>=11.0,<12.0)",
     "matplotlib (>=3.10.1,<4.0.0)",
