### 2025-10-17 b55d996
Exclude Python 3.13 from CI due to vtk installation failure
