+++ .github/workflows/python-tests.yml
@@ -8,7 +8,8 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ["3.11", "3.12", "3.13"]
+        python-version: ["3.11", "3.12"]
+        # python-version: ["3.11", "3.12", "3.13"]
 
     steps:
       - uses: actions/checkout@v3
