+++ scikit-topt/sktopt/mesh/toy_problem.py
@@ -107,7 +107,7 @@ def toy2():
     x_len = 8.0
     y_len = 8.0
     z_len = 1.0
-    mesh_size = 0.2
+    mesh_size = 0.5
     mesh = create_box_tet(x_len, y_len, z_len, mesh_size)
     dirichlet_in_range = utils.get_points_in_range(
         (0.0, 0.05), (0.0, y_len), (0.0, z_len)
@@ -129,7 +129,7 @@ def toy2():
     mesh = mesh.with_boundaries(boundaries)
     subdomains = {"design": np.array(range(mesh.nelements))}
     mesh = mesh.with_subdomains(subdomains)
-    e = skfem.ElementVector(skfem.ElementHex1())
+    e = skfem.ElementVector(skfem.ElementTetP1())
     basis = skfem.Basis(mesh, e, intorder=2)
     E0 = 210e9
     return task.TaskConfig.from_mesh_tags(
