+++ pyproject.toml
@@ -38,7 +38,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.11,<3.14"
+python = ">=3.11,<3.13"
 scikit-fem = ">=10.0.2,<11.0.0"
 matplotlib = ">=3.10.1,<4.0.0"
 meshio = ">=5.3.5,<6.0.0"
