### 2025-10-19 81ca173
add safe off-screen rendering with Xvfb warning and fallback
