### 2025-10-19 fd16bdb
add examples as tutorial
