### 2025-10-19 0f12e2a
add examples as tutorial
