### 2025-10-19 d99442b
add examples as tutorial
