### 2025-10-19 43efdba
add assertion in test with pytest
