+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -406,6 +406,9 @@ class DensityMethod(DensityMethodBase):
         recorder.add("scaling_rate", plot_type="min-max-mean-std")
         return recorder
 
+    def params_latest(self):
+        return self.recorder.as_object_latest()
+
     def scale(self):
         bbox = np.ptp(self.tsk.mesh.p, axis=1)
         L_max = np.max(bbox)
