+++ scikit-topt/tests/test_readme.py
@@ -47,16 +47,17 @@ def define_task() -> sktopt.mesh.task.TaskConfig:
     )
 
 
-def oc_optimize(tsk):
+def oc_optimize(tsk) -> sktopt.core.OC_Optimizer:
     cfg = sktopt.core.optimizers.OC_Config()
     cfg.max_iters = 1
     cfg.record_times = 1
     optimizer = sktopt.core.OC_Optimizer(cfg, tsk)
     optimizer.parameterize()
     optimizer.optimize()
+    return optimizer
 
 
-def logmoc_optimize(tsk):
+def logmoc_optimize(tsk) -> sktopt.core.LogMOC_Optimizer:
     cfg = sktopt.core.optimizers.LogMOC_Config(
         p=sktopt.tools.SchedulerConfig(
             init_value=1.0, target_value=3.0,
@@ -72,14 +73,24 @@ def logmoc_optimize(tsk):
     optimizer = sktopt.core.LogMOC_Optimizer(cfg, tsk)
     optimizer.parameterize()
     optimizer.optimize()
+    return optimizer
 
 
 def test_optimizers():
+    import numpy as np
+
     tsk1 = sktopt.mesh.toy_problem.toy_test()
-    oc_optimize(tsk1)
+    oc_optimizer = oc_optimize(tsk1)
+    result1 = oc_optimizer.recorder.as_object_latest()
+    result1 = oc_optimizer.recorder.as_object()
+    assert oc_optimizer is not None, "OC optimizer returned None"
+    assert np.isfinite(result1.compliance), "OC compliance must be finite"
 
     tsk2 = sktopt.mesh.toy_problem.toy2()
-    logmoc_optimize(tsk2)
+    moc_optimizer = logmoc_optimize(tsk2)
+    result2 = moc_optimizer.recorder.as_object()
+    assert moc_optimizer is not None, "LogMOC optimizer returned None"
+    assert np.isfinite(result2.compliance), "LogMOC compliance must be finite"
 
 
 if __name__ == "__main__":
