### 2025-10-20 a3a4e86
fix name of the link to license, lincense.md -> license
