### 2025-10-21 2eca2de
add LinearElastisicity for task config, inheritting TaskConfig
