+++ scikit-topt/sktopt/mesh/toy_problem.py
@@ -78,7 +78,7 @@ def toy_base(
     F = -100.0
     e = skfem.ElementVector(skfem.ElementHex1())
     basis = skfem.Basis(mesh, e, intorder=2)
-    return task.TaskConfig.from_facets(
+    return task.LinearElastisicity.from_facets(
         E0,
         0.30,
         basis,
@@ -132,7 +132,7 @@ def toy2():
     e = skfem.ElementVector(skfem.ElementTetP1())
     basis = skfem.Basis(mesh, e, intorder=2)
     E0 = 210e9
-    return task.TaskConfig.from_mesh_tags(
+    return task.LinearElastisicity.from_mesh_tags(
         E0,
         0.30,
         basis,
@@ -239,7 +239,7 @@ def toy_msh(
     design_elements = basis.mesh.elements_satisfying(desing_in_range)
     print("generate config")
     E0 = 210e9
-    return task.TaskConfig.from_facets(
+    return task.LinearElastisicity.from_facets(
         E0,
         0.30,
         basis,
