### 2025-10-22 24d677c
add files for feature, which is helmholtz for nodal
