+++ README.md
@@ -194,10 +194,11 @@ If you use Scikit Topt in your research or software, please cite it as:
 ## ToDo
 - Set break point from the optimization loop
 - Add A feature to assign tags to nodes and cells
+- Add Level Set
 - Add other optimizers
   - Evolutionary Algorithms
-  - Level Set
   - Phase Field
   - MMA
 - Add Multiple BC Conditions
 - Add Constraint for Rotation
+- Add Helmholtz filter for Nodal and distance filter
