### 2025-10-22 a63ab11
add dJ as test
