+++ README.md
@@ -194,9 +194,9 @@ If you use Scikit Topt in your research or software, please cite it as:
 ## ToDo
 - Set break point from the optimization loop
 - Add A feature to assign tags to nodes and cells
+- Add Level Set
 - Add other optimizers
   - Evolutionary Algorithms
-  - Level Set
   - Phase Field
   - MMA
 - Add Multiple BC Conditions
