### 2025-10-22 fb53cda
add dVol assembler
