+++ scikit-topt/sktopt/core/optimizers/common_levelset.py
@@ -7,31 +7,74 @@ from skfem import *
 from skfem.helpers import ddot, sym_grad, trace, eye, dot, identity, grad
 from skfem.models.elasticity import linear_elasticity, lame_parameters
 
+import sktopt
 
 
 def get_shape_derivative_assembler(lam: float, mu: float) -> Callable:
 
+    # dJ(q,u,φ) = ∫((C ⊙ ε(u) ⊙ ε(u))*q*(DH ∘ φ)*(norm ∘ ∇(φ)))dΩ
     @skfem.LinearForm
     def shape_derivative_assembler(v, w):
         εu = sym_grad(w.u_interp)
         σ = lam * trace(εu) * identity(εu) + 2 * mu * εu
         energy = dot(εu, σ)
         energy = np.sum(energy, axis=0)
+        # energy = np.einsum("ijxy,ijxy->xy", σ, εu)
         φ = w.φ_interp.value
         φ_grad = grad(w.φ_interp)
         φ_grad_norm = np.linalg.norm(φ_grad, axis=0)
+        n = φ_grad / (φ_grad_norm + 1e-12)
 
         q = w.q_interp.value
-        c = energy * q * φ * φ_grad_norm
+        q_n = np.einsum("ixy,ixy->xy", q, n)
+        # q_n = np.linalg.norm(q, axis=0)
+
+        # The derivative of Heaviside for smoothing
+        # beta = 0.1
+        # DH = (1 / (np.pi * beta)) * (1 / (1 + (φ / beta) ** 2))
+        DH = φ
+        c = energy * q_n * DH * φ_grad_norm
+
         return c * v
 
     return shape_derivative_assembler
 
 
-def test():
-    import skfem
-    import sktopt
+def get_dVol_assembler() -> Callable:
 
+    # dVol(q,u,φ) = ∫(-1/vol_D*q*(DH ∘ φ)*(norm ∘ ∇(φ)))dΩ
+    @skfem.LinearForm
+    def dVol_assembler(v, w):
+        φ_val = w.φ_interp.value                      # (nqp, nelems)
+        φ_grad = grad(w.φ_interp)                     # (dim, nqp, nelems)
+        φ_grad_norm = np.linalg.norm(φ_grad, axis=0)  # (nqp, nelems)
+        n = φ_grad / (φ_grad_norm + 1e-12)            # (dim, nqp, nelems)
+
+        q_val = w.q_interp.value                      # (dim, nqp, nelems)
+        q_n = np.einsum("ixy,ixy->xy", q_val, n)      # q·n → (nqp, nelems)
+
+        # The derivative of Heaviside for smoothing
+        beta = 0.1
+        DH = (1 / (np.pi * beta)) * (1 / (1 + (φ_val / beta)**2))  # (nqp, nelems)
+
+        # integrand
+        c = -1.0 / w.vol * q_n * DH * φ_grad_norm    # (nqp, nelems)
+        return c * v                                 # LinearForm expects (nqp, nelems)
+
+
+    return dVol_assembler
+
+
+def heaviside(phi, beta=0.1):
+    return 0.5 * (1 + (2 / np.pi) * np.arctan(phi / beta))
+
+
+def test():
+
+    # J(u,φ) = ∫((I ∘ φ)*(C ⊙ ε(u) ⊙ ε(u)))dΩ
+    # dJ(q,u,φ) = ∫((C ⊙ ε(u) ⊙ ε(u))*q*(DH ∘ φ)*(norm ∘ ∇(φ)))dΩ
+    # Vol(u,φ) = ∫(((ρ ∘ φ) - vf)/vol_D)dΩ
[…]
