### 2025-10-22 d2a970d
add weno_like functions
