+++ scikit-topt/sktopt/core/optimizers/common_levelset.py
@@ -69,12 +69,7 @@ def heaviside(phi, beta=0.1):
     return 0.5 * (1 + (2 / np.pi) * np.arctan(phi / beta))
 
 
-def test():
-
-    # J(u,φ) = ∫((I ∘ φ)*(C ⊙ ε(u) ⊙ ε(u)))dΩ
-    # dJ(q,u,φ) = ∫((C ⊙ ε(u) ⊙ ε(u))*q*(DH ∘ φ)*(norm ∘ ∇(φ)))dΩ
-    # Vol(u,φ) = ∫(((ρ ∘ φ) - vf)/vol_D)dΩ
-    # dVol(q,u,φ) = ∫(-1/vol_D*q*(DH ∘ φ)*(norm ∘ ∇(φ)))dΩ
+def test_fem():
     x_len, y_len, z_len = 8.0, 1.0, 1.0
     element_size = 0.1
     e_v = skfem.ElementVector(skfem.ElementHex1())
@@ -84,7 +79,7 @@ def test():
     )
     basis_v = skfem.Basis(mesh, e_v, intorder=2)
     basis_s = skfem.Basis(mesh, e_s, intorder=2)
-
+    
     def is_dirichlet_surface(x):
         return x[0] == 0.0
 
@@ -95,9 +90,6 @@ def test():
     E = 200e9
     nu = 0.3
     lam, mu = lame_parameters(E, nu)
-
-    shape_derivative_assembler = get_shape_derivative_assembler(lam, mu)
-    dVol_assembler = get_dVol_assembler()
 
     # --- Assemble stiffness matrix ---
     K = skfem.asm(linear_elasticity(lam, mu), basis_v)
@@ -124,10 +116,19 @@ def test():
 
     Kc, Fc, uc, I = skfem.condense(K, F, D=D)
     u = skfem.solve(Kc, Fc, uc, I)
+    return lam, mu, mesh, basis_v, basis_s, u
 
-    q = -u.copy()
 
-    φ = np.zeros(u.shape[0] // 3)
+def test_assembling(lam, mu, mesh, basis_v, basis_s, u, φ, q):
+
+    # J(u,φ) = ∫((I ∘ φ)*(C ⊙ ε(u) ⊙ ε(u)))dΩ
+    # dJ(q,u,φ) = ∫((C ⊙ ε(u) ⊙ ε(u))*q*(DH ∘ φ)*(norm ∘ ∇(φ)))dΩ
+    # Vol(u,φ) = ∫(((ρ ∘ φ) - vf)/vol_D)dΩ
+    # dVol(q,u,φ) = ∫(-1/vol_D*q*(DH ∘ φ)*(norm ∘ ∇(φ)))dΩ
+    shape_derivative_assembler = get_shape_derivative_assembler(lam, mu)
+    dVol_assembler = get_dVol_assembler()
+
+    
     params = dict(
         u_interp=basis_v.interpolate(u),
         φ_interp=basis_s.interpolate(φ),
@@ -141,5 +142,84 @@ def test():
     print(f"dVol: {dVol.shape}")
 
 
+def test_HJ_WENO_like(basis_s, φ, V):
+
+    # add non-linear diffusion on Hamilton-Jacobi
+    # ∂t∂ϕ​+V∣∇ϕ∣ = ∇⋅(D(∇ϕ)∇ϕ)
+    # D(∇ϕ)=α/(1+∣∇ϕ∣^2)
+    # (M+ΔtK(D(∇φ)))φn+1=Mφn−ΔtM(V∣∇φ∣)
+
+    alpha = 0.1
+
+    @skfem.BilinearForm
+    def a_weno_like(u, v, w):
+        φ_grad = grad(w.φ)
+        D = alpha / (1.0 + np.sum(φ_grad**2, axis=0))
[…]
