### 2025-10-23 0114e27
add RK4 for Hamilton-Jacobi and Reiniticalization
