+++ scikit-topt/sktopt/core/optimizers/common_levelset.py
@@ -61,7 +61,6 @@ def get_dVol_assembler() -> Callable:
         c = -1.0 / w.vol * q_n * DH * φ_grad_norm    # (nqp, nelems)
         return c * v                                 # LinearForm expects (nqp, nelems)
 
-
     return dVol_assembler
 
 
@@ -69,30 +68,48 @@ def heaviside(phi, beta=0.1):
     return 0.5 * (1 + (2 / np.pi) * np.arctan(phi / beta))
 
 
-def test_fem():
-    x_len, y_len, z_len = 8.0, 1.0, 1.0
-    element_size = 0.1
-    e_v = skfem.ElementVector(skfem.ElementHex1())
-    e_s = skfem.ElementHex1()
-    mesh = sktopt.mesh.toy_problem.create_box_hex(
-        x_len, y_len, z_len, element_size
-    )
-    basis_v = skfem.Basis(mesh, e_v, intorder=2)
-    basis_s = skfem.Basis(mesh, e_s, intorder=2)
-    
-    def is_dirichlet_surface(x):
-        return x[0] == 0.0
+def assemble_stiffness_matrix_with_φ(
+    basis_v: skfem.Basis,
+    basis_s: skfem.Basis,
+    φ: np.ndarray,
+    Emax: float, Emin: float, nu: float
+):
+    print("φ.shape =", φ.shape)
+    print("basis.N =", basis_s.N)
+    print("mesh.p.shape =", basis_s.mesh.p.shape)
 
-    def is_force_surface(x):
-        return x[0] == x_len
+    φ_elem = basis_s.interpolate(φ).value
+    E_elem = Emin + (Emax - Emin) * heaviside(φ_elem)
+    lam = (nu * E_elem) / ((1.0 + nu) * (1.0 - 2.0 * nu))
+    mu = E_elem / (2.0 * (1.0 + nu))
 
-    # --- Material parameters ---
-    E = 200e9
-    nu = 0.3
-    lam, mu = lame_parameters(E, nu)
+    # 3. BilinearForm
+    @skfem.BilinearForm
+    def stiffness_form(u, v, w):
+        strain_u = sym_grad(u)
+        strain_v = sym_grad(v)
+        volume_term = lam * trace(strain_u) * trace(strain_v)
+        return volume_term + 2.0 * mu * ddot(strain_u, strain_v)
 
+    K = skfem.asm(stiffness_form, basis_v)
+    return K
+
+
+def test_fem(
+    mesh: skfem.Mesh,
+    basis_v: skfem.Basis, basis_s: skfem.Basis,
+    is_dirichlet_surface: Callable,
+    is_force_surface: Callable,
+    φ: np.ndarray,
+    Emax=200e9,
+    Emin=200e4,
+    nu=0.3,
+    F_total=100.0
+):
     # --- Assemble stiffness matrix ---
-    K = skfem.asm(linear_elasticity(lam, mu), basis_v)
+    # lam, mu = lame_parameters(E, nu)
+    # K = skfem.asm(linear_elasticity(lam, mu), basis_v)
+    K = assemble_stiffness_matrix_with_φ(basis_v, basis_s, φ, Emax, Emin, nu)
 
     # --- Boundary conditions ---
     left = mesh.facets_satisfying(is_dirichlet_surface)
@@ -100,46 +117,63 @@ def test_fem():
 
[…]
