### 2025-10-23 da8924a
fix ci flow for pytest
