+++ .github/workflows/python-tests.yml
@@ -1,6 +1,14 @@
 name: Python tests
 
-on: [push, pull_request]
+on:
+  push:
+    branches:
+      - master
+      - dev
+  pull_request:
+    branches:
+      - master
+      - dev
 
 jobs:
   test:
