### 2025-10-23 d35f505
create a base of spacial filter
