+++ scikit-topt/sktopt/filter/spacial_filter.py
@@ -1,0 +1,162 @@
+from dataclasses import dataclass
+from typing import Callable, Optional, Tuple, Literal
+
+import numpy as np
+
+from scipy.spatial import cKDTree
+from scipy.sparse import coo_matrix, csr_matrix, diags
+
+import skfem
+
+
+def get_element_centers(mesh: skfem.Mesh) -> np.ndarray:
+    coords = mesh.p[:, mesh.t]
+    centers = np.mean(coords, axis=1)
+    return centers
+
+
+def make_kernel(
+    kind: Literal["linear", "quadratic", "gaussian"],
+    r_min: float, 
+    sigma: Optional[float] = None
+) -> Tuple[Callable[[np.ndarray], np.ndarray], float]:
+    if kind == "linear":
+        def kernel(d):
+            return np.maximum(0.0, r_min - d)
+        support = r_min
+
+    elif kind == "quadratic":
+        def kernel(d):
+            x = d / r_min
+            w = 1.0 - x * x
+            w[d >= r_min] = 0.0
+            return w
+        support = r_min
+
+    elif kind == "gaussian":
+        s = (r_min / 3.0) if sigma is None else float(sigma)
+        cutoff = 3.0 * s
+
+        def kernel(d):
+            return np.exp(-0.5 * (d / s) ** 2)
+        support = cutoff
+
+    else:
+        raise ValueError("unknown kernel kind")
+
+    return kernel, support
+
+
+def build_filter_matrix(
+    element_centers: np.ndarray,
+    kernel: Callable[[np.ndarray], np.ndarray],
+    support_radius: float,
+    elem_volume: Optional[np.ndarray] = None,
+    volume_correction: bool = True,
+) -> csr_matrix:
+    c = element_centers
+    n = c.shape[1]
+    tree = cKDTree(c.T)
+
+    nbrs = tree.query_ball_tree(tree, support_radius)
+
+    rows, cols, data = [], [], []
+    for i in range(n):
+        js = np.asarray(nbrs[i], dtype=int)
+        if js.size == 0:
+            continue
+        dij = np.linalg.norm(c[:, js] - c[:, [i]], axis=0)
+        w = kernel(dij)
+
+        if volume_correction and elem_volume is not None:
+            w = w * elem_volume[js]
+        rows.append(np.full(js.size, i, dtype=int))
+        cols.append(js)
+        data.append(w)
+
+    if len(data) == 0:
+        raise RuntimeError("No neighbor relations found; check support_radius.")
+
+    rows = np.concatenate(rows)
+    cols = np.concatenate(cols)
+    data = np.concatenate(data)
+
+    W = coo_matrix((data, (rows, cols)), shape=(n, n)).tocsr()
+    row_sum = np.array(W.sum(axis=1)).ravel()
+    row_sum[row_sum == 0.0] = 1.0
+    Dinv = diags(1.0 / row_sum)
+    W = Dinv @ W
+    return W
+
+
+@dataclass
+class SpacialFilter():
+    mesh: skfem.Mesh
+    radius: float
+    element_centers: np.ndarray
+    design_mask: Optional[np.ndarray] = None
+    dst_path: Optional[str] = None
+
+    def update_radius(
+        self,
+        radius: float,
+        **args
+    ):
+        self.radius = radius
+
+
+    @classmethod
+    def from_defaults(
+        cls,
+        mesh: skfem.Mesh,
+        radius: float = 1.0,
+        dst_path: Optional[str] = None,
+        design_mask: Optional[np.ndarray] = None,
+    ) -> 'SpacialFilter':
+
+        element_centers = get_element_centers(mesh)
+        return cls(
+            mesh=mesh,
+            radius=radius,
+            element_centers=element_centers,
[…]
