### 2025-10-23 f03ed95
add design_mask to spacial filter
