### 2025-10-23 3f39fa9
add gradient in SpacialFIlter, change module name filter->filters
