+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -130,9 +130,10 @@ class DensityMethodConfig():
             curvature=2.0
         )
     )
+    filter_type: Literal["spacial", "helmholtz"] = "spacial"
     filter_radius: tools.SchedulerConfig = field(
         default_factory=lambda: tools.SchedulerConfig.constant(
-            target_value=1.2
+            target_value=0.2
         )
     )
     E0: float = 210e9
@@ -446,38 +447,35 @@ class DensityMethod(DensityMethodBase):
                 -1,
                 self.cfg.max_iters
             )
-        # if "eta_init" in cfg.__dataclass_fields__:
-        #     self.schedulers.add(
-        #         "eta",
-        #         self.cfg.eta_init,
-        #         self.cfg.eta,
-        #         self.cfg.eta_step,
-        #         self.cfg.max_iters
-        #     )
-        # else:
-        #     self.schedulers.add(
-        #         "eta",
-        #         self.cfg.eta,
-        #         self.cfg.eta,
-        #         -1,
-        #         self.cfg.max_iters
-        #     )
         self.schedulers.set_iters_max(cfg.max_iters)
 
         if export:
             self.schedulers.export()
 
     def parameterize(self):
-        self.helmholz_solver = filter.HelmholtzFilter.from_defaults(
-            self.tsk.mesh,
-            self.cfg.filter_radius.init_value,
-            solver_option="cg_pyamg",
-            # solver_option=self.cfg.solver_option,
-            dst_path=f"{self.cfg.dst_path}/data",
-        )
+
+        if self.cfg.filter_type == "spacial":
+            self.filter = filters.SpacialFilter.from_defaults(
+                self.tsk.mesh,
+                self.cfg.filter_radius.init_value,
+                design_mask=self.tsk.design_mask,
+                dst_path=f"{self.cfg.dst_path}/data",
+            )
+        elif self.cfg.filter_type == "helmholtz":
+            # helmholz_solver
+            self.filter = filters.HelmholtzFilter.from_defaults(
+                self.tsk.mesh,
+                self.cfg.filter_radius.init_value,
+                design_mask=self.tsk.design_mask,
+                solver_option="cg_pyamg",
+                # solver_option=self.cfg.solver_option,
+                dst_path=f"{self.cfg.dst_path}/data",
+            )
+        else:
+            raise ValueError("should be spacial or helmholtz")
 
     def load_parameters(self):
-        self.helmholz_solver = filter.HelmholtzFilter.from_file(
+        self.filter = filters.HelmholtzFilter.from_file(
             f"{self.cfg.dst_path}/data"
         )
 
@@ -652,9 +650,7 @@ class DensityMethod(DensityMethodBase):
         ) = self.initialize_params()
         elements_volume_design = tsk.elements_volume[tsk.design_elements]
         elements_volume_design_sum = np.sum(elements_volume_design)
-        self.helmholz_solver.update_radius(
-            tsk.mesh, filter_radius, solver_option="cg_pyamg"
-        )
+        self.filter.update_radius(filter_radius)
 
         # Loop 1 - N
         for iter_num in range(iter_begin, iter_end):
@@ -675,15 +671,13 @@ class DensityMethod(DensityMethodBase):
                 precision=6
             )
 
-            if filter_radius != self.helmholz_solver.radius:
+            if filter_radius != self.filter.radius:
                 logger.info("--- Filter Update ---")
-                self.helmholz_solver.update_radius(
-                    tsk.mesh, filter_radius, "cg_pyamg"
-                )
+                self.filter.update_radius(filter_radius)
 
             logger.info("--- project and filter ---")
             rho_prev[:] = rho[:]
[…]
