+++ scikit-topt/sktopt/mesh/task.py
@@ -164,6 +164,11 @@ class TaskConfig():
     fixed_elements: np.ndarray
     dirichlet_force_elements: np.ndarray
     elements_volume: np.ndarray
+    
+    
+    @property
+    def design_mask(self):
+        return np.isin(self.all_elements, self.design_elements)
 
     @property
     def mesh(self):
@@ -428,7 +433,7 @@ class TaskConfig():
 
     @classmethod
     def from_mesh_tags(
-        cls, 
+        cls,
         E: float,
         nu: float,
         basis: skfem.Basis,
