+++ scikit-topt/sktopt/filters/__init__.py
@@ -1,0 +1,8 @@
+from sktopt.filter.helmholtz import SpacialFilter
+from sktopt.filter.helmholtz import HelmholtzFilter
+
+
+__all__ = [
+    "SpacialFilter",
+    "HelmholtzFilter",
+]
