### 2025-10-23 ec408a5
add helmholtz nodal
