+++ scikit-topt/sktopt/filters/__init__.py
@@ -1,8 +1,8 @@
 from sktopt.filters.spacial import SpacialFilter
-from sktopt.filters.helmholtz import HelmholtzFilter
+from sktopt.filters.helmholtz_ele import HelmholtzFilter_ele
 
 
 __all__ = [
     "SpacialFilter",
-    "HelmholtzFilter",
+    "HelmholtzFilter_ele",
 ]
