+++ scikit-topt/sktopt/filters/helmholtz_nodal.py
@@ -1,0 +1,184 @@
+from typing import Optional
+from dataclasses import dataclass
+
+import numpy as np
+from scipy.sparse.linalg import spsolve
+
+import skfem
+from skfem.helpers import dot, grad
+
+from sktopt.filters.base import BaseFilter
+
+
+def infer_element_from_mesh(mesh: skfem.Mesh):
+    if isinstance(mesh, skfem.MeshTri):
+        return skfem.ElementTriP1()
+    elif isinstance(mesh, skfem.MeshQuad):
+        return skfem.ElementQuad1()
+    elif isinstance(mesh, skfem.MeshTet):
+        return skfem.ElementTetP1()
+    elif isinstance(mesh, skfem.MeshHex):
+        return skfem.ElementHex1()
+    else:
+        raise ValueError(f"Unknown mesh type: {type(mesh)}")
+
+
+def node_to_element_density(mesh, rho_node):
+    return np.mean(rho_node[mesh.t], axis=0)
+
+
+def element_to_node_density_averaging(
+    mesh: skfem.Mesh,
+    elements_volume: np.ndarray,
+    rho_elem: np.ndarray,
+    design_mask: np.ndarray | None = None,
+    weighted: bool = True,
+) -> np.ndarray:
+    n_nodes = mesh.p.shape[1]
+    rho_node = np.zeros(n_nodes)
+    wsum = np.zeros(n_nodes)
+
+    t = mesh.t.T
+
+    if design_mask is None:
+        design_mask = np.ones(len(t), dtype=bool)
+
+    for e, nodes in enumerate(t):
+        rho_val = rho_elem[e] if design_mask[e] else 1.0
+        w = elements_volume[e] if weighted else 1.0
+        rho_node[nodes] += w * rho_val
+        wsum[nodes] += w
+
+    wsum[wsum == 0.0] = 1.0
+    rho_node /= wsum
+    return rho_node
+
+
+#
+# Need to be fixed
+#
+# def project_element_density_to_nodes(mesh, rho_elem, design_mask=None):
+#     basis = skfem.Basis(mesh, infer_element_from_mesh(mesh))
+#     n_elem = mesh.t.shape[1]
+#     if design_mask is None:
+#         design_mask = np.ones(n_elem, dtype=bool)
+
+#     @skfem.BilinearForm
+#     def mass(u, v, w):
+#         return dot(u, v)
+
+#     @skfem.LinearForm
+#     def rhs_design(v, w):
+#         return rho_elem[w.idx] * v
+
+#     @skfem.LinearForm
+#     def rhs_fixed(v, w):
+#         return 1.0 * v
+
+#     M = skfem.asm(mass, basis)
+#     F_design = skfem.asm(rhs_design, basis.with_elements(np.nonzero(design_mask)[0]))
+#     F_fixed = skfem.asm(rhs_fixed, basis.with_elements(np.nonzero(~design_mask)[0]))
+
+#     F = F_design + F_fixed
+#     rho_node = spsolve(M, F)
+#     return rho_node
+
+
+def solve_helmholtz(mesh, rho_node, r_min, design_mask=None):
+    basis = skfem.Basis(mesh, infer_element_from_mesh(mesh))
+    elements = mesh.t.shape[1]
+
+    if design_mask is None:
+        design_mask = np.ones(elements, dtype=bool)
+
+    rho_field = basis.interpolate(rho_node)
+
+    @skfem.BilinearForm
+    def a(u, v, w):
+        return u * v + (r_min ** 2) * dot(grad(u), grad(v))
+
+    @skfem.LinearForm
+    def L(v, w):
+        mask = design_mask[w.idx]
+        val = rho_field.value.copy()
[…]

[…]
