+++ scikit-topt/sktopt/filters/__init__.py
@@ -1,4 +1,5 @@
 from sktopt.filters.spacial import SpacialFilter
+# from sktopt.filters.helmholtz_nodal import HelmholtzFilter_nodal
 from sktopt.filters.helmholtz_ele import HelmholtzFilter_ele
 
 
