### 2025-10-24 d103e4e
implemented helmholtz filter nodal
