### 2025-10-24 acff43f
add dirichlet BCs for helmholtz filter
