### 2025-10-24 ff0a898
comment out helmhotlz element
