### 2025-10-24 8d321e7
remove item from ToDo
