+++ _sources/installation.rst.txt
@@ -2,7 +2,7 @@ Installation
 ============
 
 sktopt is a Python package that can be installed using pip or poetry.  
-It is compatible with Python 3.11 and later versions. To install sktopt, follow these steps:
+It is compatible with Python 3.11 - 3.12. To install sktopt, follow these steps:
 
 **Install sktopt**:  
 Open your terminal or command prompt and run one of the following commands:
