+++ README.md
@@ -24,8 +24,8 @@
 
 ### Example 2 : Multiple Load Condition
 <p align="center">
-  <img src="https://media.githubusercontent.com/media/kevin-tofu/scikit-topt/master/assets/ex-multi-load-condition.jpg" alt="multi-load-condition" width="400" style="margin-right: 20px;">
-  <img src="https://media.githubusercontent.com/media/kevin-tofu/scikit-topt/master/assets/ex-multi-load-v-50.jpg" alt="multi-load-condition-distribution" width="400">
+  <img src="https://raw.githubusercontent.com/kevin-tofu/scikit-topt/refs/heads/master/assets/ex-multi-load-condition.jpg" alt="multi-load-condition" width="400" style="margin-right: 20px;">
+  <img src="https://raw.githubusercontent.com/kevin-tofu/scikit-topt/refs/heads/master/assets/ex-multi-load-v-50.jpg" alt="multi-load-condition-distribution" width="400">
 </p>
 
 ### Progress Report
