+++ scikit-topt/sktopt/core/misc.py
@@ -118,10 +118,7 @@ def add_common_arguments(
         '--rho_min', '-RhM', type=float, default=1e-1, help=''
     )
     parser.add_argument(
-        '--E0', '-E', type=float, default=1.0, help=''
-    )
-    parser.add_argument(
-        '--E_min', '-EM', type=float, default=1e-6, help=''
+        '--E_min_coeff', '-EM', type=float, default=1e-3, help=''
     )
     parser.add_argument(
         '--eta', '-ET', type=float, default=0.3, help=''
