### 2025-11-03 1d16814
there is a bug in gradient, helmholtz
