### 2025-11-04 0a0f737
update gradint in helmholtz, enforce negative gradient
