+++ scikit-topt/sktopt/filters/helmholtz_filter_nodal.py
@@ -33,6 +33,7 @@ def element_to_node_density_averaging(
     rho_elem: np.ndarray,
     design_mask: np.ndarray | None = None,
     weighted: bool = True,
+    fixed_value_for_design: float = 1.0
 ) -> np.ndarray:
     n_nodes = mesh.p.shape[1]
     rho_node = np.zeros(n_nodes)
@@ -44,7 +45,8 @@ def element_to_node_density_averaging(
         design_mask = np.ones(len(t), dtype=bool)
 
     for e, nodes in enumerate(t):
-        rho_val = rho_elem[e] if design_mask[e] else 1.0
+        rho_val = rho_elem[e] if design_mask[e] else fixed_value_for_design
+        # rho_val = rho_elem[e] if design_mask[e] else 0.0
         w = elements_volume[e] if weighted else 1.0
         rho_node[nodes] += w * rho_val
         wsum[nodes] += w
@@ -84,17 +86,50 @@ def element_to_node_density_averaging(
 #     return rho_node
 
 
+# def solve_helmholtz(
+#     case: Literal["forward", "gradient"],
+#     mesh: skfem.Mesh, rho_node: np.ndarray,
+#     r_min: float, design_mask: Optional[np.ndarray] = None,
+# ):
+#     const_value = 1.0 if case == "forward" else -1e-15
+#     basis = skfem.Basis(mesh, infer_element_from_mesh(mesh))
+#     elements = mesh.t.shape[1]
+
+#     if design_mask is None:
+#         design_mask = np.ones(elements, dtype=bool)
+
+#     rho_field = basis.interpolate(rho_node)
+
+#     @skfem.BilinearForm
+#     def a(u, v, w):
+#         return u * v + (r_min ** 2) * dot(grad(u), grad(v))
+
+#     @skfem.LinearForm
+#     def L(v, w):
+#         return rho_field * v
+
+#     A = skfem.asm(a, basis)
+#     b = skfem.asm(L, basis)
+#     fixed_nodes = np.unique(mesh.t[:, ~design_mask].ravel())
+#     D = fixed_nodes
+#     x0 = np.zeros(A.shape[0])
+#     x0[fixed_nodes] = const_value
+#     rho_filtered = skfem.solve(*skfem.condense(A, b, D=D, x=x0))
+#     return rho_filtered
+
+
 def solve_helmholtz(
     case: Literal["forward", "gradient"],
-    mesh: skfem.Mesh, rho_node: np.ndarray,
-    r_min: float, design_mask: Optional[np.ndarray] = None,
+    mesh: skfem.Mesh,
+    rho_node: np.ndarray,
+    r_min: float,
+    design_mask: Optional[np.ndarray] = None,
 ):
-    const_value = 1.0 if case == "forward" else -1e-15
     basis = skfem.Basis(mesh, infer_element_from_mesh(mesh))
-    elements = mesh.t.shape[1]
-
     if design_mask is None:
-        design_mask = np.ones(elements, dtype=bool)
+        fixed_nodes = np.array([], dtype=int)
+    else:
+        fixed_nodes = np.unique(mesh.t[:, ~design_mask].ravel())
 
     rho_field = basis.interpolate(rho_node)
 
@@ -108,12 +143,19 @@ def solve_helmholtz(
 
     A = skfem.asm(a, basis)
     b = skfem.asm(L, basis)
-    fixed_nodes = np.unique(mesh.t[:, ~design_mask].ravel())
-    D = fixed_nodes
-    x0 = np.zeros(A.shape[0])
-    x0[fixed_nodes] = const_value
-    rho_filtered = skfem.solve(*skfem.condense(A, b, D=D, x=x0))
[…]
