### 2025-11-05 749044a
divide structual task into base and itself, force->neumann
