+++ scikit-topt/sktopt/core/optimizers/logmoc.py
@@ -153,7 +153,7 @@ class LogMOC_Optimizer(common_density.DensityMethod):
     def __init__(
         self,
         cfg: LogMOC_Config,
-        tsk: sktopt.mesh.TaskConfig,
+        tsk: sktopt.mesh.FEMDomain,
     ):
         assert cfg.lambda_lower < cfg.lambda_upper
         super().__init__(cfg, tsk)
