+++ scikit-topt/sktopt/mesh/__init__.py
@@ -1,10 +1,10 @@
 
-from sktopt.mesh.task import TaskConfig, LinearElastisicity
+from sktopt.mesh.task import FEMDomain, LinearElastisicity
 from sktopt.mesh import toy_problem
 from sktopt.mesh import loader
 
 __all__ = [
-    "TaskConfig",
+    "FEMDomain",
     "LinearElastisicity",
     "toy_problem",
     "loader"

[…]
