+++ scikit-topt/sktopt/mesh/__init__.py
@@ -1,5 +1,6 @@
 
-from sktopt.mesh.task import FEMDomain, LinearElastisicity
+from sktopt.mesh.task_common import FEMDomain
+from sktopt.mesh.task_elastic import LinearElastisicity
 from sktopt.mesh import toy_problem
 from sktopt.mesh import loader
 
