### 2025-11-06 0fd678e
add variables for robin cond
