### 2025-11-06 ce7f11e
add task for heat conduction
