### 2025-11-06 f9cd7fe
add task for heat conduction
