+++ scikit-topt/sktopt/mesh/task_heat.py
@@ -16,93 +16,56 @@ def setdiff1d(a, b):
     return np.ascontiguousarray(a)
 
 
-
-
-
 _lit_bc = Literal['u^1', 'u^2', 'u^3', 'all']
-_lit_neumann = Literal['u^1', 'u^2', 'u^3']
+# _lit_robin = Literal['u^1', 'u^2', 'u^3']
 
 
-def assemble_surface_neumanns(
+def assemble_surface_robin(
     basis,
-    neumann_facets_ids: Union[np.ndarray, List[np.ndarray]],
-    neumann_dir_type: Union[str, List[str]],
-    neumann_value: Union[float, List[float]],
-    *,
-    treat_value_as_total_neumann: bool = True,
+    robin_facets_ids: Union[np.ndarray, List[np.ndarray]],
+    robin_coefficient: Union[float, List[float]],
+    robin_bc_value: Union[float, List[float]]
 ):
     def _to_list(x):
         return x if isinstance(x, list) else [x]
 
-    def _dir_to_comp(s: str) -> int:
-        if not (isinstance(s, str) and s.startswith('u^') and s[2:].isdigit()):
-            raise ValueError(
-                f"neumann_dir_type must be like 'u^1','u^2','u^3', got: {s}"
-            )
-        c = int(s[2:]) - 1
-        if c < 0:
-            raise ValueError(f"Invalid component index parsed from {s}")
-        return c
+    facets_list = _to_list(robin_facets_ids)
+    robin_h_list = _to_list(robin_coefficient)
+    robin_Tenv_list = _to_list(robin_bc_value)
 
-    facets_list = _to_list(neumann_facets_ids)
-    dirs_list = _to_list(neumann_dir_type)
-    vals_list = _to_list(neumann_value)
-
-    if not (len(facets_list) == len(dirs_list) == len(vals_list)):
-        # print("len(facets_list) : ", len(facets_list))
-        # print("len(dirs_list) : ", len(dirs_list))
-        # print("len(vals_list) : ", len(vals_list))
+    if not (len(facets_list) == len(robin_h_list) == len(robin_Tenv_list)):
         raise ValueError(
-            "Lengths of neumann_facets_ids, neumann_dir_type, and neumann_value\
+            "Lengths of robin_facets_ids and robin_value\
                 must match when lists."
         )
 
-    @skfem.Functional
-    def l_one(w):
-        return 1.0
-
-    F_list = list()
-    for facets, dir_s, val in zip(facets_list, dirs_list, vals_list):
-        comp = _dir_to_comp(dir_s)
+    linear_list = list()
+    bilinear_list = list()
+    for facets, h, Tenv in zip(
+        facets_list, robin_h_list, robin_Tenv_list
+    ):
         fb = FacetBasis(
             basis.mesh, basis.elem,
             facets=np.asarray(facets, dtype=int)
         )
 
-        A = asm(l_one, fb)
-        if A <= 0.0:
-            raise ValueError(
-                "Selected facets have zero total area; check facet indices or geometry."
-            )
+        @skfem.BilinearForm
+        def robin_form(u, v, w):
+            # Surface integral (heat radiation term)
+            return h * u * v
 
-        if treat_value_as_total_neumann:
-            pressure = float(val) / A
-        else:
-            pressure = float(val)
+        @skfem.LinearForm
+        def robin_load(v, w):
+            # Temperature difference from outside temperature
+            return h * Tenv * v
 
-        @LinearForm
-        def l_comp(v, w):
-            # print(f"v.shape {v.shape}")
-            # print(f"w.n.shape {w.n.shape}")
-            # v.shape (3, 46, 16)
-            # w.n.shape (3, 46, 16)
-            # return pressure * skfem.helpers.dot(w.n, v)
-            return pressure * v[comp]
[…]

[…]
