### 2025-11-08 963f6e8
add solver for heat coundction
