+++ scikit-topt/sktopt/fea/solver_heat.py
@@ -1,0 +1,228 @@
+from typing import Callable, Literal
+
+import numpy as np
+import scipy
+from scipy.sparse.linalg import LinearOperator
+
+
+import skfem
+from skfem.models.elasticity import lame_parameters
+from skfem import Functional
+from skfem.helpers import ddot, sym_grad, trace, eye
+from skfem.helpers import transpose
+import pyamg
+
+from sktopt.mesh import LinearHeatConduction
+from sktopt.fea import composer
+from sktopt.tools.logconf import mylogger
+logger = mylogger(__name__)
+
+
+def solve_multi_load(
+    basis: skfem.CellBasis,
+    free_dofs: np.ndarray,
+    dirichlet_dofs: np.ndarray,
+    heat_source_list: list[np.ndarray],
+    k0: float, kmin: float, p: float,
+    rho: np.ndarray,
+    u_all: np.ndarray,
+    solver: Literal['auto', 'cg_jacobi', 'spsolve', 'cg_pyamg'] = 'auto',
+    elem_func: Callable = composer.simp_interpolation,
+    rtol: float = 1e-5,
+    maxiter: int = None,
+    n_joblib: int = 1
+) -> float:
+    solver = 'spsolve' if solver == 'auto' else solver
+    n_dof = basis.N
+    assert u_all.shape == (n_dof, len(heat_source_list))
+
+    K = composer.assemble_conduction_matrix(
+        basis, rho, k0, kmin, p, elem_func
+    )
+    _maxiter = min(1000, max(300, n_dof // 5)) if maxiter is None else maxiter
+    K_csr = K.tocsr()
+    K_e, _ = skfem.enforce(K_csr, heat_source_list[0], D=dirichlet_dofs)
+    heat_source_stack = np.column_stack([
+        skfem.enforce(K_csr, f, D=dirichlet_dofs)[1] for f in heat_source_list
+    ])
+    # compliance_total = 0.0
+    u_all[:, :] = 0.0
+    if solver == 'spsolve':
+        if n_joblib > 1:
+            from joblib import Parallel, delayed, parallel_backend
+            lu = scipy.sparse.linalg.splu(K_e.tocsc())
+
+            def solve_system(heat_source_stack):
+                return lu.solve(heat_source_stack)
+
+            with parallel_backend("threading"):
+                u_all[:, :] = np.column_stack(
+                    Parallel(n_jobs=n_joblib)(
+                        delayed(solve_system)(heat_source_stack[:, i]) for i in range(
+                            heat_source_stack.shape[1]
+                        )
+                    )
+                )
+
+        else:
+            lu = scipy.sparse.linalg.splu(K_e.tocsc())
+            u_all[:, :] = np.column_stack(
+                [lu.solve(heat_source_stack[:, i]) for i in range(heat_source_stack.shape[1])]
+            )
+
+    else:
+        # choose preconditioner if needed
+        if solver == 'cg_jacobi':
+            M_diag = K_e.diagonal()
+            M_inv = 1.0 / M_diag
+            M = LinearOperator(K_e.shape, matvec=lambda x: M_inv * x)
+        elif solver == 'cg_pyamg':
+            ml = pyamg.smoothed_aggregation_solver(K_e)
+            M = ml.aspreconditioner()
+        else:
+            raise ValueError(f"Unknown solver: {solver}")
+
+        for i, _ in enumerate(heat_source_list):
+            F_e = heat_source_stack[:, i]
+            # _, F_e = skfem.enforce(K_csr, force, D=dirichlet_dofs)
+            u_e, info = scipy.sparse.linalg.cg(
+                K_e, F_e, M=M, rtol=rtol, maxiter=_maxiter
+            )
[…]
