### 2025-11-08 3be996e
unified variable names
