+++ scikit-topt/sktopt/fea/solver_elastic.py
@@ -337,7 +337,7 @@ class FEM_SimpLinearElastisicity():
         self.solver_option = solver_option
         self.n_joblib = n_joblib
 
-    def compute_compliance_multi_load(
+    def compliance_multi_load(
         self,
         rho: np.ndarray, p: float,
         u_dofs: np.ndarray
@@ -355,7 +355,7 @@ class FEM_SimpLinearElastisicity():
         )
         return compliance_array
 
-    def strain_energy_skfem_multi_load(
+    def energy_multi_load(
         self,
         rho: np.ndarray, p: float, u_dofs: np.ndarray
     ) -> np.ndarray:
