+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -297,7 +297,7 @@ class DensityMethodBase(ABC):
         rho_projected: np.ndarray,
         dC_drho_design_eles: np.ndarray,
         u_dofs: np.ndarray,
-        strain_energy_mean: np.ndarray,
+        energy_mean: np.ndarray,
         scaling_rate: np.ndarray,
         move_limit: float,
         eta: float,
@@ -404,7 +404,7 @@ class DensityMethod(DensityMethodBase):
     ) -> tools.HistoriesLogger:
         recorder = tools.HistoriesLogger(self.cfg.dst_path)
         recorder.add("rho_projected", plot_type="min-max-mean-std")
-        recorder.add("strain_energy", plot_type="min-max-mean-std")
+        recorder.add("energy", plot_type="min-max-mean-std")
         recorder.add("vol_error")
         if isinstance(tsk.force, list):
             recorder.add("u_max", plot_type="min-max-mean-std")
@@ -557,7 +557,7 @@ class DensityMethod(DensityMethodBase):
                 Gradient field after combining sensitivity with projection derivative.
             dC_drho_projected : ndarray of shape (n_elements,)
                 Compliance sensitivities with respect to the projected densities.
-            strain_energy_mean : ndarray of shape (n_elements,)
+            energy_mean : ndarray of shape (n_elements,)
                 Average element strain energy over all load cases.
             dC_drho_full : ndarray of shape (n_elements,)
                 Compliance sensitivities mapped back to the full set of elements.
@@ -593,7 +593,7 @@ class DensityMethod(DensityMethodBase):
         dH_drho = np.empty_like(rho)
         grad_filtered = np.empty_like(rho)
         dC_drho_projected = np.empty_like(rho)
-        strain_energy_mean = np.zeros_like(rho)
+        energy_mean = np.zeros_like(rho)
         dC_drho_full = np.zeros_like(rho)
         dC_drho_design_eles = np.zeros_like(rho[tsk.design_elements])
         scaling_rate = np.empty_like(rho[tsk.design_elements])
@@ -619,7 +619,7 @@ class DensityMethod(DensityMethodBase):
             dH_drho,
             grad_filtered,
             dC_drho_projected,
-            strain_energy_mean,
+            energy_mean,
             dC_drho_full,
             dC_drho_design_eles,
             scaling_rate,
@@ -647,7 +647,7 @@ class DensityMethod(DensityMethodBase):
             dH_drho,
             grad_filtered,
             dC_drho_projected,
-            strain_energy_mean,
+            energy_mean,
             dC_drho_full,
             dC_drho_design_eles,
             scaling_rate,
@@ -693,16 +693,16 @@ class DensityMethod(DensityMethodBase):
             logger.info("--- compute compliance ---")
             dC_drho_design_eles[:] = 0.0
             dC_drho_full[:] = 0.0
-            strain_energy_mean[:] = 0.0
+            energy_mean[:] = 0.0
             u_max = list()
 
-            compliance_avg = self.fem.compute_compliance_multi_load(
+            compliance_avg = self.fem.compliance_multi_load(
                 rho_projected, p, u_dofs
             ).mean()
-            strain_energy = self.fem.strain_energy_skfem_multi_load(
+            energy = self.fem.energy_multi_load(
                 rho_projected, p, u_dofs,
             )
-            strain_energy_mean = strain_energy.mean(axis=1)
+            energy_mean = energy.mean(axis=1)
             for task_loop in range(self.tsk.n_tasks):
 
                 u_max.append(np.abs(u_dofs[:, task_loop]).max())
@@ -711,7 +711,7 @@ class DensityMethod(DensityMethodBase):
                     dC_drho_projected,
                     dC_drho_func(
                         rho_projected,
[…]
