### 2025-11-09 7307201
add solver for heat conduction
