+++ scikit-topt/sktopt/mesh/task_common.py
@@ -40,7 +40,7 @@ class FEMDomain():
         Degrees of freedom constrained by Dirichlet (displacement) boundary conditions.
     dirichlet_elements : np.ndarray
         Elements that contain Dirichlet boundary points.
-    force_elements : np.ndarray
+    neumann_elements : np.ndarray
         Elements that contain the force application points.
     force : np.ndarray or list of np.ndarray
         External force vector(s) applied to the system.
@@ -55,26 +55,28 @@ class FEMDomain():
         Array of all element indices in the mesh.
     fixed_elements : np.ndarray
         Elements excluded from the design domain.
-    dirichlet_force_elements : np.ndarray
+    dirichlet_neumann_elements : np.ndarray
         Union of Dirichlet and force elements.
         Useful for identifying constrained and loaded regions.
     elements_volume : np.ndarray
         Volume of each finite element, used in volume constraints and integration.
     """
     basis: skfem.Basis
-    dirichlet_nodes: np.ndarray
-    dirichlet_dofs: np.ndarray
-    dirichlet_elements: np.ndarray
-    neumann_nodes: Optional[np.ndarray | list[np.ndarray]]
-    neumann_elements: Optional[np.ndarray]
-    neumann_dir_type: Optional[str | list[str]]
-    neumann_values: Optional[float | list[float]]
+    dirichlet_nodes: np.ndarray | list[np.ndarray] | None
+    dirichlet_dofs: np.ndarray | list[np.ndarray] | None  # The nessecity is not high?
+    dirichlet_elements: np.ndarray | None
+    dirichlet_values: float | list[float] | None
+
+    neumann_nodes: np.ndarray | list[np.ndarray] | None
+    neumann_elements: np.ndarray | None
+    neumann_dir_type: str | list[str] | None
+    neumann_values: float | list[float] | None
     # neumann_vector: Optional[np.ndarray | list[np.ndarray]]
 
-    robin_nodes: Optional[np.ndarray | list[np.ndarray]]
-    robin_elements: Optional[np.ndarray]
-    robin_coefficient: Optional[float | list[float]]
-    robin_bc_value: Optional[float | list[float]]
+    robin_nodes: np.ndarray | list[np.ndarray] | None
+    robin_elements: np.ndarray | None
+    robin_coefficient: float | list[float] | None
+    robin_bc_value: float | list[float]| None
     # robin_vector: Optional[np.ndarray | list[np.ndarray]]
 
     design_elements: np.ndarray
@@ -101,47 +103,66 @@ class FEMDomain():
     def from_nodes(
         cls,
         basis: skfem.Basis,
-        dirichlet_nodes: np.ndarray | list[np.ndarray],
-        dirichlet_dir: _lit_bc | list[_lit_bc],
-        neumann_nodes: Optional[np.ndarray | list[np.ndarray]],
-        neumann_dir_type: Optional[str | list[str]],
-        neumann_values: Optional[np.ndarray | list[np.ndarray]],
-        robin_nodes: Optional[np.ndarray | list[np.ndarray]],
-        robin_coefficient: Optional[float | list[float]],
-        robin_bc_value: Optional[float | list[float]],
+        dirichlet_nodes: np.ndarray | list[np.ndarray] | None,
+        dirichlet_dir: _lit_bc | list[_lit_bc] | None,
+        dirichlet_values: float | list[float] | None,
+        neumann_nodes: np.ndarray | list[np.ndarray] | None,
+        neumann_dir_type: str | list[str] | None,
+        neumann_values: np.ndarray | list[np.ndarray] | None,
+        robin_nodes: np.ndarray | list[np.ndarray] | None,
+        robin_coefficient: float | list[float] | None,
[…]
