+++ scikit-topt/sktopt/fea/__init__.py
@@ -1,0 +1,8 @@
+
+from sktopt.fea.solver_elastic import FEM_SimpLinearElastisicity
+from sktopt.fea.solver_heat import FEM_SimpLinearHeatConduction
+
+__all__ = [
+    "FEM_SimpLinearElastisicity",
+    "FEM_SimpLinearHeatConduction"
+]
