+++ scikit-topt/sktopt/mesh/__init__.py
@@ -1,12 +1,14 @@
 
 from sktopt.mesh.task_common import FEMDomain
 from sktopt.mesh.task_elastic import LinearElastisicity
+from sktopt.mesh.task_heat import LinearHeatConduction
 from sktopt.mesh import toy_problem
 from sktopt.mesh import loader
 
 __all__ = [
     "FEMDomain",
     "LinearElastisicity",
+    "LinearHeatConduction",
     "toy_problem",
     "loader"
 ]

[…]
