### 2025-11-09 82295e3
bug fix, multiple robin conditions
