### 2025-11-09 5c5aaf8
add example for heat conduction
