+++ scikit-topt/sktopt/mesh/task_elastic.py
@@ -211,6 +211,7 @@ class LinearElastisicity(FEMDomain):
             basis,
             dirichlet_facets_ids,
             dirichlet_dir,
+            None,
             force_facets_ids,
             force_dir_type,
             force_value,
@@ -234,6 +235,7 @@ class LinearElastisicity(FEMDomain):
             base.dirichlet_nodes,
             base.dirichlet_dofs,
             base.dirichlet_elements,
+            base.dirichlet_values,
             base.neumann_nodes,
             base.neumann_elements,
             base.neumann_dir_type,
