### 2025-11-09 398cc82
add averaged-temperature objective
