+++ scikit-topt/sktopt/fea/solver_heat.py
@@ -15,6 +15,44 @@ from sktopt.tools.logconf import mylogger
 logger = mylogger(__name__)
 
 
+def solve_scipy(
+    K_csr, emit,
+    dirichlet_dofs_list, dirichlet_values_list,
+    u_all,
+    n_joblib: int = 1
+):
+    from joblib import Parallel, delayed, parallel_backend
+
+    def solve_system(
+        dirichlet_dofs_loop, dirichlet_values_loop
+    ):
+        T_sol = skfem.solve(
+            *skfem.condense(
+                K_csr, emit, D=dirichlet_dofs_loop,
+                x=np.full(dirichlet_dofs_loop.N, dirichlet_values_loop)
+            )
+        )
+        return T_sol
+    if n_joblib > 1:
+        with parallel_backend("threading"):
+            u_all[:, :] = np.column_stack(
+                Parallel(n_jobs=n_joblib)(
+                    delayed(solve_system)(
+                        dirichlet_dofs_list[i], dirichlet_values_list[i]
+                    ) for i in range(len(dirichlet_values_list))
+                )
+            )
+
+    else:
+        u_all[:, :] = np.column_stack(
+            [
+                solve_system(
+                    dirichlet_dofs_list[i], dirichlet_values_list[i]
+                ) for i in range(len(dirichlet_values_list))
+            ]
+        )
+
+
 def solve_multi_load(
     basis: skfem.Basis,
     free_dofs: np.ndarray,
@@ -31,7 +69,7 @@ def solve_multi_load(
     maxiter: int = None,
     n_joblib: int = 1,
     objective: Literal["compliance", "averaged_temp"] = "averaged_temp"
-) -> float:
+) -> list:
     solver = 'spsolve' if solver == 'auto' else solver
     n_dof = basis.N
     # assert u_all.shape == (n_dof, len(dirichlet_values))
@@ -57,53 +95,39 @@ def solve_multi_load(
         basis.get_dofs(nodes=loop) for loop in dirichlet_nodes_list
     ]
 
-    def solve_system(
-        dirichlet_dofs_loop, dirichlet_values_loop
-    ):
-        T_sol = skfem.solve(
-            *skfem.condense(
-                K_csr, emit, D=dirichlet_dofs_loop,
-                x=np.full(dirichlet_dofs_loop.N, dirichlet_values_loop)
-            )
+    if solver == "spsolve":
+        solve_scipy(
+            K_csr, emit,
+            dirichlet_dofs_list, dirichlet_values_list,
+            u_all, n_joblib=n_joblib
         )
-        return T_sol
-
-    if solver == 'spsolve':
-        if n_joblib > 1:
-            from joblib import Parallel, delayed, parallel_backend
-            with parallel_backend("threading"):
-                u_all[:, :] = np.column_stack(
-                    Parallel(n_jobs=n_joblib)(
-                        delayed(solve_system)(
-                            dirichlet_dofs_list[i], dirichlet_values_list[i]
-                        ) for i in range(len(dirichlet_values_list))
-                    )
-                )
-
-        else:
-            u_all[:, :] = np.column_stack(
-                [
-                    solve_system(
-                        dirichlet_dofs_list[i], dirichlet_values_list[i]
-                    ) for i in range(len(dirichlet_values_list))
-                ]
-            )
+    else:
+        raise NotImplementedError("")
 
     objective_list = list()
     if objective == "compliance":
         for i in range(u_all.shape[1]):
             T_i = u_all[:, i]
             objective_list.append(float(T_i @ (K_csr @ T_i)))
+
+        # λ_all = None
+        λ_all = - 2 * u_all
[…]
