### 2025-11-09 4ba4840
add func in readme
