+++ scikit-topt/sktopt/mesh/task_heat.py
@@ -61,20 +61,42 @@ def assemble_surface_robin(
     basis,
     robin_facets_ids: np.ndarray | List[np.ndarray],
     robin_coefficient: float | List[float],
-    robin_bc_value: float | List[float]
+    robin_bc_value: float | List[float],
+    rho: Optional[np.ndarray] = None,
+    p: float = 3.0
 ):
     def _to_list(x):
         return x if isinstance(x, list) else [x]
 
     facets_list = _to_list(robin_facets_ids)
-    robin_h_list = _to_list(robin_coefficient)
-    robin_Tenv_list = _to_list(robin_bc_value)
+    robin_h_list = [
+        robin_coefficient for _ in range(len(facets_list))
+    ] if isinstance(robin_coefficient, float) else robin_coefficient
+    robin_Tenv_list = [
+        robin_bc_value for _ in range(len(facets_list))
+    ] if isinstance(robin_bc_value, float) else robin_bc_value
 
     if not (len(facets_list) == len(robin_h_list) == len(robin_Tenv_list)):
         raise ValueError(
             "Lengths of robin_facets_ids and robin_value\
                 must match when lists."
         )
+
+    rho_field = None
+    if rho is not None:
+        # nodal field if given as element-wise array
+        if rho.shape[0] == basis.mesh.nelements:
+            # element平均ρ → 節点に平均化
+            rho_nodal = np.zeros(basis.mesh.nvertices)
+            count = np.zeros(basis.mesh.nvertices)
+            t = basis.mesh.t
+            np.add.at(rho_nodal, t.ravel(), np.repeat(rho, t.shape[0]))
+            np.add.at(count, t.ravel(), 1)
+            rho_nodal /= np.maximum(count, 1)
+        else:
+            rho_nodal = rho
+
+        # rho_field = basis.interpolate(rho_nodal)
 
     linear_list = list()
     bilinear_list = list()
@@ -85,32 +107,54 @@ def assemble_surface_robin(
             basis.mesh, basis.elem,
             facets=np.asarray(facets, dtype=int)
         )
+        if rho_field is not None:
+            rho_facet = fb.interpolate(rho_nodal)
 
         @skfem.BilinearForm
         def robin_form(u, v, w):
-            # Surface integral (heat radiation term)
-            return w.h * u * v
+            if rho_field is None:
+                h_eff = w.h
+            else:
+                h_eff = w.h * w.rho**p
+            return h_eff * u * v
 
         @skfem.LinearForm
         def robin_load(v, w):
-            # Temperature difference from outside temperature
-            return w.h * w.Tenv * v
+            if rho_field is None:
+                h_eff = w.h
+            else:
+                h_eff = w.h * w.rho**p
+            return h_eff * w.Tenv * v
 
-        bilinear_list.append(asm(robin_form, fb, h=h))
-        linear_list.append(asm(robin_load, fb, h=h, Tenv=Tenv))
+        w_dict = {'h': h, 'Tenv': Tenv}
+        if rho_field is not None:
+            w_dict['rho'] = rho_facet
 
-    if len(bilinear_list) == 1:
-        bilinear_list = bilinear_list[0]
-        linear_list = linear_list[0]
+        bilinear_list.append(skfem.asm(robin_form, fb, **w_dict))
+        linear_list.append(skfem.asm(robin_load, fb, **w_dict))
+
+    # if len(bilinear_list) == 1:
+    #     bilinear_list = bilinear_list[0]
+    #     linear_list = linear_list[0]
     return bilinear_list, linear_list
 
 
 @dataclass
 class LinearHeatConduction(FEMDomain):
     k: float  # thermal conductivity
-    robin_bilinear: Optional[np.array] = None
-    robin_linear: Optional[np.array] = None
[…]
