+++ examples/tutorial/json_moc.py
@@ -48,12 +48,12 @@ def get_task():
     neumann_value = 100
     # Define it as a task
     tsk = sktopt.mesh.LinearElastisicity.from_mesh_tags(
-        210e9,
-        0.30,
         basis,
         dirichlet_dir,
         neumann_dir_type,
-        neumann_value
+        neumann_value,
+        210e3,
+        0.30,
     )
     return tsk
 

[…]
