+++ examples/tutorial/json_moc.py
@@ -47,7 +47,7 @@ def get_task():
     neumann_dir_type = "u^3"
     neumann_value = 100
     # Define it as a task
-    tsk = sktopt.mesh.LinearElastisicity.from_mesh_tags(
+    tsk = sktopt.mesh.LinearElasticity.from_mesh_tags(
         basis,
         dirichlet_dir,
         neumann_dir_type,
