+++ scikit-topt/tests/test_global_flow.py
@@ -2,7 +2,7 @@ import sktopt
 # import pytest
 
 
-def define_task() -> sktopt.mesh.LinearElasticity:
+def define_task_elasticity() -> sktopt.mesh.LinearElasticity:
 
     import skfem
     x_len, y_len, z_len = 1.0, 1.0, 1.0
@@ -45,6 +45,59 @@ def define_task() -> sktopt.mesh.LinearElasticity:
         210e3,
         0.30,
     )
+
+
+def define_task_heatconduction() -> sktopt.mesh.LinearHeatConduction:
+    import numpy as np
+    import skfem
+    x_len = 8.0
+    y_len = 8.0
+    z_len = 1.0
+    # z_len = 8.0
+    mesh_size = 0.2
+
+    mesh = sktopt.mesh.toy_problem.create_box_hex(
+        x_len, y_len, z_len, mesh_size
+    )
+    robin_in_range_0 = sktopt.mesh.utils.get_points_in_range(
+        (0.0, 0.0), (0.0, y_len), (0.0, z_len)
+    )
+    robin_in_range_1 = sktopt.mesh.utils.get_points_in_range(
+        (0.0, x_len), (y_len, y_len), (0.0, z_len)
+    )
+
+    robin_coefficient = 4.0e-5
+    robin_bc_value = 300.0
+
+    eps = mesh_size
+    dirichlet_in_range = sktopt.mesh.utils.get_points_in_range(
+        (x_len-1.0*x_len/20, x_len), (0.0, 1.0*y_len/20), (0.0, z_len)
+    )
+    dirichlet_value = 600.0
+    boundaries = {
+        "robin_0": robin_in_range_0,
+        "robin_1": robin_in_range_1,
+        "dirichlet_0": dirichlet_in_range
+    }
+
+    mesh = mesh.with_boundaries(boundaries)
+    subdomains = {"design": np.array(range(mesh.nelements))}
+    mesh = mesh.with_subdomains(subdomains)
+
+    e = skfem.ElementHex1()
+    basis = skfem.Basis(mesh, e, intorder=1)
+    k = 10.0
+    objective = "compliance"
+    design_robin_boundary = True
+    mytask = sktopt.mesh.LinearHeatConduction.from_mesh_tags(
+        basis,
+        dirichlet_value,
+        robin_coefficient,
+        robin_bc_value,
+        design_robin_boundary,
+        k, objective
+    )
+    return mytask
 
 
 def oc_optimize(tsk) -> sktopt.core.OC_Optimizer:
@@ -91,6 +144,13 @@ def test_optimizers():
     result2 = moc_optimizer.recorder.as_object()
     assert moc_optimizer is not None, "LogMOC optimizer returned None"
     assert np.isfinite(result2.compliance), "LogMOC compliance must be finite"
+    
+    tsk3 = define_task_heatconduction()
+    oc_optimizer = oc_optimize(tsk3)
+    result1 = oc_optimizer.recorder.as_object_latest()
+    result1 = oc_optimizer.recorder.as_object()
+    assert oc_optimizer is not None, "OC optimizer returned None"
+    assert np.isfinite(result1.compliance), "OC compliance must be finite"
 
 
 if __name__ == "__main__":
