+++ scikit-topt/sktopt/mesh/__init__.py
@@ -1,13 +1,13 @@
 
 from sktopt.mesh.task_common import FEMDomain
-from sktopt.mesh.task_elastic import LinearElastisicity
+from sktopt.mesh.task_elastic import LinearElasticity
 from sktopt.mesh.task_heat import LinearHeatConduction
 from sktopt.mesh import toy_problem
 from sktopt.mesh import loader
 
 __all__ = [
     "FEMDomain",
-    "LinearElastisicity",
+    "LinearElasticity",
     "LinearHeatConduction",
     "toy_problem",
     "loader"
