+++ README.md
@@ -28,6 +28,12 @@
   <img src="https://raw.githubusercontent.com/kevin-tofu/scikit-topt/refs/heads/master/assets/ex-multi-load-v-50.jpg" alt="multi-load-condition-distribution" width="400">
 </p>
 
+### Example 3 : Heat Conduction
+<p align="center">
+  <img src="https://raw.githubusercontent.com/kevin-tofu/scikit-topt/refs/heads/feature-heat-equation/assets/ex-heat-conduction-0.jpg" alt="heat-conduction" width="400" style="margin-right: 20px;">
+  <img src="https://media.githubusercontent.com/media/kevin-tofu/scikit-topt/feature-heat-equation/assets/ex-heat-conduction-0.gif" alt="heat-conduction" width="400" style="margin-right: 20px;">
+</p>
+
 ### Progress Report
 <p align="center">
   <img src="https://media.githubusercontent.com/media/kevin-tofu/scikit-topt/master/assets/ex-progress-report.jpg" alt="multi-load-condition-progress" width="600">
@@ -48,7 +54,7 @@ The currently supported features are as follows:
 - High-performance computation using sparse matrices with Scipy and PyAMG  
 - has a function to monitor the transition of parameters.
 
-## Usage
+## SetUp
 
 You can install **Scikit-Topt** either via **pip** or **Poetry**.
 
