+++ docs/mesh.task_heat.rst
@@ -1,0 +1,8 @@
+sktopt.mesh.task_elastic
+==============================
+
+.. automodule:: sktopt.mesh.task_elastic
+   :members:
+   :no-index:
+   :undoc-members:
+   :show-inheritance:
