+++ docs/fea.solver_heat.rst
@@ -1,0 +1,8 @@
+sktopt.fea.solver_heat
+==============================
+
+.. automodule:: sktopt.fea.solver_heat
+   :members:
+   :no-index:
+   :undoc-members:
+   :show-inheritance:
