+++ docs/fea.solver_elastic.rst
@@ -1,0 +1,8 @@
+sktopt.mesh.solver_elastic
+==============================
+
+.. automodule:: sktopt.fea.solver_elastic
+   :members:
+   :no-index:
+   :undoc-members:
+   :show-inheritance:
