+++ docs/api.rst
@@ -9,6 +9,8 @@ This page provides detailed documentation for modules, classes and functions.
    core.optimizers.common
    core.optimizers.oc
    core.optimizers.logmoc
+   fea.solver_elastic
+   fea.solver_heat
    mesh.task_common
    mesh.task_elastic
    mesh.task_heat
