+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -7,7 +7,6 @@ import inspect
 import shutil
 import json
 
-
 import numpy as np
 import sktopt
 from sktopt import tools
@@ -29,82 +28,89 @@ class DensityMethodConfig():
     """
     Configuration for density-based topology optimization (SIMP/RAMP).
 
-    This configuration controls interpolation, continuation schedules via
-    `SchedulerConfig`, filtering/projection thresholds, and numerical options.
-    It targets sensitivity-based methods such as SIMP and RAMP.
+    This configuration controls interpolation models, continuation schedules
+    (via :class:`~sktopt.tools.SchedulerConfig`), filtering/projection thresholds,
+    and numerical options. It targets sensitivity-based topology optimization
+    methods such as SIMP and RAMP.
 
     Notes on continuation
     ---------------------
     Parameters that change during optimization are represented by
-    :class:`tools.SchedulerConfig`:
-      - ``p``: penalization power (e.g., 1.0 → 3.0)
-      - ``vol_frac``: volume fraction constraint (e.g., 0.8 → 0.4)
-      - ``beta``: Heaviside projection sharpness (e.g., 1.0 → 2.0)
-      - ``filter_radius``: filter radius (often constant)
+    :class:`~sktopt.tools.SchedulerConfig`:
 
-    Each scheduler defines how a value evolves (e.g., "Step", "StepAccelerating",
-    "Constant") and may include extra shape parameters (e.g., ``curvature`` for
-    accelerating schedules).
+    - ``p``: Penalization power (e.g., 1.0 → 3.0)
+    - ``vol_frac``: Volume fraction constraint (e.g., 0.8 → 0.4)
+    - ``beta``: Heaviside projection sharpness (e.g., 1.0 → 2.0)
+    - ``filter_radius``: Filter radius (often constant)
+
+    Each scheduler defines how its value evolves (e.g., ``"Step"``,
+    ``"StepAccelerating"``, ``"Constant"``) and may include shape parameters
+    such as ``curvature`` for accelerating schedules.
 
     Attributes
     ----------
     dst_path : str
         Output directory for results and logs.
     interpolation : {"SIMP", "RAMP"}
-        Material interpolation model for penalization.
+        Material interpolation model used for penalization.
     record_times : int
         Number of snapshots recorded during optimization.
     max_iters : int
         Maximum number of optimization iterations.
 
     beta_eta : float
-        Threshold parameter (η) for Heaviside projection; the density at which
-        the projection transitions sharply between 0 and 1.
+        Threshold parameter (η) for the Heaviside projection. Represents the
+        density at which the projection transitions sharply between 0 and 1.
     eta : float
         Threshold density used in projection or post-processing.
 
-    p : tools.SchedulerConfig
+    p : :class:`~sktopt.tools.SchedulerConfig`
         Continuation schedule for penalization power.
-        Default: Step from 1.0 to 3.0.
-    vol_frac : tools.SchedulerConfig
-        Continuation schedule for volume fraction constraint.
-        Default: Step from 0.8 to 0.4.
-    beta : tools.SchedulerConfig
+        Default: ``Step(init_value=1.0, target_value=3.0, num_steps=3)``.
+    vol_frac : :class:`~sktopt.tools.SchedulerConfig`
+        Continuation schedule for the volume fraction constraint.
+        Default: ``Constant(target_value=0.8)``.
+    beta : :class:`~sktopt.tools.SchedulerConfig`
         Continuation schedule for Heaviside sharpness (β).
-        Default: StepAccelerating from 1.0 to 2.0 with ``curvature=2.0``.
-    filter_radius : tools.SchedulerConfig
[…]
