+++ docs/api.rst
@@ -1,16 +1,26 @@
-
 API Reference
-===============
+=============
 This page provides detailed documentation for modules, classes and functions.
 
 .. toctree::
    :maxdepth: 1
+   :caption: Core Optimizers
 
    core.optimizers.common_density
    core.optimizers.oc
    core.optimizers.logmoc
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Finite Element Solvers
+
    fea.solver_elastic
    fea.solver_heat
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Mesh and Task Utilities
+
    mesh.task_common
    mesh.task_elastic
    mesh.task_heat
