+++ scikit-topt/sktopt/mesh/__init__.py
@@ -5,6 +5,10 @@ from sktopt.mesh.task_heat import LinearHeatConduction
 from sktopt.mesh import toy_problem
 from sktopt.mesh import loader
 
+FEMDomain.__module__ = "sktopt.mesh"
+LinearElasticity.__module__ = "sktopt.mesh"
+LinearHeatConduction.__module__ = "sktopt.mesh"
+
 __all__ = [
     "FEMDomain",
     "LinearElasticity",
