+++ scikit-topt/sktopt/tools/__init__.py
@@ -5,6 +5,13 @@ from sktopt.tools.scheduler import SchedulerStepAccelerating
 from sktopt.tools.scheduler import SchedulerStep
 from sktopt.tools.scheduler import SchedulerSawtoothDecay
 
+HistoriesLogger.__module__ = "sktopt.tools"
+SchedulerConfig.__module__ = "sktopt.tools"
+Schedulers.__module__ = "sktopt.tools"
+SchedulerStepAccelerating.__module__ = "sktopt.tools"
+SchedulerStep.__module__ = "sktopt.tools"
+SchedulerSawtoothDecay.__module__ = "sktopt.tools"
+
 __all__ = [
     "HistoriesLogger",
     "SchedulerConfig",
