+++ docs/fea.solver_elastic.rst
@@ -1,8 +1,0 @@
-sktopt.fea.solver_elastic
-==============================
-
-.. automodule:: sktopt.fea.solver_elastic
-   :members:
-   :no-index:
-   :undoc-members:
-   :show-inheritance:
