+++ docs/mesh.task_common.rst
@@ -1,8 +1,0 @@
-sktopt.mesh.task_common
-==============================
-
-.. automodule:: sktopt.mesh.task_common
-   :members:
-   :no-index:
-   :undoc-members:
-   :show-inheritance:
