+++ docs/core.optimizers.common_density.rst
@@ -1,8 +1,0 @@
-sktopt.core.optimizers.common_density
-======================================
-
-.. automodule:: sktopt.core.optimizers.common_density
-   :members:
-   :noindex:
-   :undoc-members:
-   :show-inheritance:
