+++ docs/mesh.task_heat.rst
@@ -1,8 +1,0 @@
-sktopt.mesh.task_elastic
-==============================
-
-.. automodule:: sktopt.mesh.task_elastic
-   :members:
-   :no-index:
-   :undoc-members:
-   :show-inheritance:
