+++ docs/api.rst
@@ -6,22 +6,17 @@ This page provides detailed documentation for modules, classes and functions.
    :maxdepth: 1
    :caption: Core Optimizers
 
-   core.optimizers.common_density
-   core.optimizers.oc
-   core.optimizers.logmoc
+   core
 
 .. toctree::
    :maxdepth: 1
    :caption: Finite Element Solvers
 
-   fea.solver_elastic
-   fea.solver_heat
+   fea
 
 .. toctree::
    :maxdepth: 1
    :caption: Mesh and Task Utilities
 
-   mesh.task_common
-   mesh.task_elastic
-   mesh.task_heat
+   mesh
    mesh.utils
