+++ docs/core.optimizers.logmoc.rst
@@ -1,8 +1,0 @@
-sktopt.core.optimizers.logmoc
-==============================
-
-.. automodule:: sktopt.core.optimizers.logmoc
-   :members:
-   :no-index:
-   :undoc-members:
-   :show-inheritance:
