### 2025-11-13 d9a5d7d
add trycatch for joblib module import
