+++ _sources/core.optimizers.common.rst.txt
@@ -1,8 +1,0 @@
-sktopt.core.optimizers.common
-==============================
-
-.. automodule:: sktopt.core.optimizers.common
-   :members:
-   :no-index:
-   :undoc-members:
-   :show-inheritance:
