+++ _sources/core.optimizers.oc.rst.txt
@@ -1,8 +1,0 @@
-sktopt.core.optimizers.oc
-==============================
-
-.. automodule:: sktopt.core.optimizers.oc
-   :members:
-   :no-index:
-   :undoc-members:
-   :show-inheritance:
