+++ _sources/api.rst.txt
@@ -1,13 +1,22 @@
-
 API Reference
-===============
+=============
 This page provides detailed documentation for modules, classes and functions.
 
 .. toctree::
    :maxdepth: 1
+   :caption: Core Optimizers
 
-   core.optimizers.common
-   core.optimizers.oc
-   core.optimizers.logmoc
-   mesh.task
+   core
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Finite Element Solvers
+
+   fea
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Mesh and Task Utilities
+
+   mesh
    mesh.utils
