### 2025-11-13 b86e9fa
fix error handling
