### 2025-11-14 561189c
add docstring
