### 2025-11-17 c81209a
Expand and refine bullet points in the JOSS paper for greater clarity and detail.
