### 2025-11-19 8345ede
add base class to doc
