+++ docs/tools.rst
@@ -1,0 +1,8 @@
+sktopt.tools
+============
+
+.. automodule:: sktopt.tools
+   :members:
+   :imported-members:
+   :no-index:
+   :show-inheritance:
