### 2025-11-21 e0201c0
adding objective, heat-exchange
