### 2025-11-21 4673e00
change class name of history, and add its docstring
