+++ scikit-topt/sktopt/tools/__init__.py
@@ -1,11 +1,11 @@
-from sktopt.tools.history import HistoriesLogger
+from sktopt.tools.history import HistoryCollection
 from sktopt.tools.scheduler import SchedulerConfig
 from sktopt.tools.scheduler import Schedulers
 from sktopt.tools.scheduler import SchedulerStepAccelerating
 from sktopt.tools.scheduler import SchedulerStep
 from sktopt.tools.scheduler import SchedulerSawtoothDecay
 
-HistoriesLogger.__module__ = "sktopt.tools"
+HistoryCollection.__module__ = "sktopt.tools"
 SchedulerConfig.__module__ = "sktopt.tools"
 Schedulers.__module__ = "sktopt.tools"
 SchedulerStepAccelerating.__module__ = "sktopt.tools"
@@ -13,7 +13,7 @@ SchedulerStep.__module__ = "sktopt.tools"
 SchedulerSawtoothDecay.__module__ = "sktopt.tools"
 
 __all__ = [
-    "HistoriesLogger",
+    "HistoryCollection",
     "SchedulerConfig",
     "Schedulers",
     "SchedulerStep",
