+++ scikit-topt/sktopt/tools/history.py
@@ -7,7 +7,49 @@ from sktopt.tools.logconf import mylogger
 logger = mylogger(__name__)
 
 
-class HistoryLogger():
+class HistorySeries():
+    """
+    Container for logging the evolution of a single scalar quantity
+    (or its summary statistics) over iterations.
+
+    This class stores a sequence of values or aggregated statistics
+    (min / mean / max / std) and provides helpers to:
+
+    * Append new data from scalars or arrays
+    * Query the logged data as a NumPy array
+    * Pretty-print the latest entry
+    * Export the data in a consistent array + header format
+
+    Parameters
+    ----------
+    name : str
+        Name of the history (e.g. ``"compliance"``, ``"volume"``).
+    constants : list of float, optional
+        Optional constant values associated with this history
+        (e.g. target volume fraction, penalty parameters).
+        Currently stored but not used in the internal logic.
+    constant_names : list of str, optional
+        Names corresponding to ``constants``. Must have the same
+        length as ``constants`` if both are provided.
+    plot_type : {"min-max-mean", "min-max-mean-std"}, optional
+        How array-valued inputs should be aggregated when passed to
+        :meth:`add`:
+
+        * ``"min-max-mean"``:
+          store ``[min(x), mean(x), max(x)]``
+        * ``"min-max-mean-std"``:
+          store ``[min(x), mean(x), max(x), std(x)]``
+
+        Default is ``"min-max-mean"``.
+    ylog : bool, optional
+        If ``True``, the history is intended to be plotted on a
+        logarithmic y-axis. Used by :meth:`HistoryCollection.export_progress`.
+    data : array_like, optional
+        Initial data for this history. May be a 1D/2D NumPy array or a
+        Python list. Internally it is stored as a Python list.
+        If omitted, the history starts empty.
+    """
+
     def __init__(
         self,
         name: str,
@@ -35,11 +77,41 @@ class HistoryLogger():
             self.data = []
 
     def exists(self):
+        """
+        Return whether this history has at least one logged entry.
+
+        Returns
+        -------
+        bool
+            ``True`` if at least one value has been stored,
+            ``False`` otherwise.
+        """
+
         ret = True if len(self.data) > 0 else False
         return ret
 
     @property
     def data_np_array(self) -> np.ndarray:
+        """
+        Return the logged data as a NumPy array.
+
+        Returns
+        -------
+        numpy.ndarray
+            Array view of the internal data. The shape depends on how
+            data has been added:
+
+            * Scalar inputs → 1D array of shape ``(n_steps,)``
+            * Aggregated lists (min/mean/max[/std]) → 2D array of shape
+              ``(n_steps, n_stats)``
+
+        Raises
+        ------
+        ValueError
+            If no data is available or if the internal buffer cannot be
+            converted to a NumPy array.
+        """
+
         try:
             value = self.data[0]
             if isinstance(value, float):
@@ -51,6 +123,23 @@ class HistoryLogger():
             raise ValueError(f"data not exit {e}")
 
     def add(self, data_input: np.ndarray | float):
+        """
+        Append a new data point to the history.
+
+        Parameters
+        ----------
+        data_input : float or numpy.ndarray
+            If a scalar (Python float or 0-dimensional array), the raw
+            value is appended.
+
+            If an array with at least one dimension:
+
+            * For ``plot_type="min-max-mean"``:
+              append ``[min(x), mean(x), max(x)]``
+            * For ``plot_type="min-max-mean-std"``:
+              append ``[min(x), mean(x), max(x), std(x)]``
+        """
+
         if isinstance(data_input, np.ndarray):
             if data_input.shape == ():
                 self.data.append(float(data_input))
@@ -65,6 +154,13 @@ class HistoryLogger():
             self.data.append(float(data_input))
 
[…]
