+++ scikit-topt/sktopt/fea/solver_heat.py
@@ -74,29 +74,29 @@ def solve_multi_load(
     rtol: float = 1e-5,
     maxiter: int = None,
     n_joblib: int = 1,
-    objective: Literal["compliance", "averaged_temp"] = "averaged_temp"
 ) -> list:
     solver = 'spsolve' if solver == 'auto' else solver
-    # n_dof = basis.N
-    # assert u_all.shape == (n_dof, len(dirichlet_values))
+
     K = composer.assemble_conduction_matrix(
         basis, rho, k0, kmin, p, elem_func
     )
+
     if isinstance(robin_bilinear, scipy.sparse.csr_matrix):
         K = K + robin_bilinear
     elif isinstance(robin_bilinear, list):
         for loop in robin_bilinear:
             K += loop
 
-    emit = np.zeros([K.shape[0]])
+    emit = np.zeros(K.shape[0])
     if isinstance(robin_linear, np.ndarray):
         emit = robin_linear
     elif isinstance(robin_linear, list):
         for loop in robin_linear:
             emit += loop
+
     K_csr = K.tocsr()
-    # compliance_total = 0.0
     u_all[:, :] = 0.0
+
     dirichlet_dofs_list = [
         basis.get_dofs(nodes=loop) for loop in dirichlet_nodes_list
     ]
@@ -110,138 +110,7 @@ def solve_multi_load(
     else:
         raise NotImplementedError("")
 
-    objective_list = list()
-    if objective == "compliance":
-        for i in range(u_all.shape[1]):
-            T_i = u_all[:, i]
-            objective_list.append(float(T_i @ (K_csr @ T_i)))
-
-        # λ_all = None
-        λ_all = - 2 * u_all
-    elif objective == "averaged_temp":
-        objective_list = [
-            float(np.sum(u_all[:, i])) for i in range(u_all.shape[1])
-        ]
-        λ_all = np.zeros_like(u_all)
-        solve_scipy(
-            K_csr, np.ones_like(emit),
-            dirichlet_dofs_list, dirichlet_values_list,
-            λ_all, n_joblib=n_joblib
-        )
-    else:
-        raise ValueError("")
-    return objective_list, λ_all
-
-
-def solve_multi_load(
-    basis: skfem.Basis,
-    free_dofs: np.ndarray,
-    dirichlet_nodes_list: list[np.ndarray],
-    dirichlet_values_list: list[float],
-    robin_bilinear: scipy.sparse.csr_matrix | list[scipy.sparse.csr_matrix],
-    robin_linear: np.ndarray | list[np.ndarray],
-    k0: float, kmin: float, p: float,
-    rho: np.ndarray,
-    u_all: np.ndarray,
-    solver: Literal['auto', 'spsolve'] = 'auto',
-    elem_func: Callable = composer.simp_interpolation,
-    rtol: float = 1e-5,
-    maxiter: int = None,
-    n_joblib: int = 1,
-    objective: Literal["compliance", "averaged_temp"] = "averaged_temp"
-) -> list:
-    solver = 'spsolve' if solver == 'auto' else solver
-    # n_dof = basis.N
-    # assert u_all.shape == (n_dof, len(dirichlet_values))
-    K = composer.assemble_conduction_matrix(
-        basis, rho, k0, kmin, p, elem_func
-    )
-    if isinstance(robin_bilinear, scipy.sparse.csr_matrix):
-        K = K + robin_bilinear
-    elif isinstance(robin_bilinear, list):
-        for loop in robin_bilinear:
-            K += loop
-
-    emit = np.zeros([K.shape[0]])
-    if isinstance(robin_linear, np.ndarray):
-        emit = robin_linear
[…]
