+++ README.md
@@ -137,7 +137,7 @@ If you use Scikit Topt in your research or software, please cite it as:
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.15441499},
   url          = {https://doi.org/10.5281/zenodo.15441499},
-  note         = {Version 0.2.18}
+  note         = {Version 0.3.1}
 }
 ```
 
