### 2025-11-22 #9092 @HaoLI-KU
Dear Kohei (@kevin-tofu)
My detailed comments are as follows:

1.     In https://kevin-tofu.github.io/scikit-topt/project.html

- The project name appears inconsistently as “scikit-topt” and “Scikit-Topt”; please standardise the capitalization throughout.
- “Minimizing compliance under a volume constraint” is presented as the only objective. Given that the heat conduction problem has been implemented, I suggest stating the physics and objectives supported explicitly.

2.     https://kevin-tofu.github.io/scikit-topt/installation.html

- The phrase “sktopt is a Python package…” uses the wrong package name, please match the actual library name.
- The compatibility statement “Python 3.11-3.12” is narrow. If earlier or newer versions are supported (e.g., 3.9, 3.10, 3.13, 3.14), please clarify.

3.     https://kevin-tofu.github.io/scikit-topt/tutorial.html

- The text states “compliance minimization with OC method”, but the optimiser class used is LogMOC_Optimizer, this inconsistency should be addressed.
- vol_frac=vol_frac= appears duplicated (with equal signs repeated)
- What .msh formats are supported (Gmsh v2/v4); this may confuse users. The call to task.LinearElasticity.from_mesh_tags relies on tags that seems not have been described. Please provide guidance on how users should tag mesh regions.
- The example uses skfem.ElementHex1() without explaining what element type or order this represents. For clarity, please mention that this is a first-order hexahedral element, and what element types are supported so far?
- The two overlayed image captions (multi-load-condition and Multi-Load-condition-Density-Distribution) appear broken. Please fix these images and labels.
- The subsection “Results and Visualization” is duplicated as a heading, leading to redundancy.

4.     https://kevin-tofu.github.io/scikit-topt/physics.html

- The title contains a major typo: “Gourvan Equation” should be corrected to “Governing Equation.”

5.     https://kevin-tofu.github.io/scikit-topt/physics-elasticity.html

- The document states that Dirichlet conditions are enforced by “zeroing rows” and setting diagonals to 1, but does not mention zeroing columns, which is required for symmetry-preserving enforcement. Clarify whether skfem handles this internally. And how the right hand side is corrected after the elimination.

6.     https://kevin-tofu.github.io/scikit-topt/physics-heat.html

- The notation does not state explicitly whether the conductivity may be anisotropic or tensor valued. This is important in topology optimisation contexts and please clarify.

7.     https://kevin-tofu.github.io/scikit-topt/formulation.html

- The interpolation formulas are presented without explaining that Emin must be strictly positive to avoid singular stiffness matrices; And it should be at least three orders of magnitude smaller than the E_0.
- The sensitivity expression does not explicitly explain how sensitivities propagate through filtering and projection. Adding this would strengthen the explanation.
- Only iteration count and convergence are mentioned. Typical criteria include max density change or KKT residual. Adding them would make the method description more complete.

8.     https://kevin-tofu.github.io/scikit-topt/optimization.html

- It is mentioned that OC relies on heuristic parameters but does not give recommended ranges (e.g., move limits of 0.2–0.5, typical \eta values), which would be useful to practitioners.
- The OC update implicitly requires a volume constraint bisection loop, but the document does not outline the bisection procedure; this should be briefly included.

9.     https://kevin-tofu.github.io/scikit-topt/api.html

- As written, “This page provides detailed documentation,” consider removing this for now, and adding it back when the detailed documentation is ready.
