### 2025-11-23 93472e2
modified installation/project
