+++ .github/workflows/python-tests.yml
@@ -16,7 +16,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ["3.11", "3.12"]
+        python-version: ["3.10", "3.11", "3.12", "3.13", "3.14"]
         # python-version: ["3.11", "3.12", "3.13"]
 
     steps:
