+++ pyproject.toml
@@ -7,11 +7,13 @@ authors = [
 ]
 license = { text = "Apache-2.0" }
 readme = "README.md"
-requires-python = ">=3.11,<3.13"
+requires-python = ">=3.10,<=3.14"
 classifiers = [
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
 ]
 dependencies = [
     "scikit-fem (>=11.0,<12.0)",
@@ -38,7 +40,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.11,<3.13"
+python = ">=3.10,<=3.14"
 scikit-fem = ">=10.0.2,<11.0.0"
 matplotlib = ">=3.10.1,<4.0.0"
 meshio = ">=5.3.5,<6.0.0"
