### 2025-11-23 280c0dc
add python version on sphinx
