+++ pyproject.toml
@@ -56,7 +56,7 @@ pytest-cov = "^5.0.0"
 line-profiler = ">=4.2.0,<5.0.0"
 memory-profiler = ">=0.61.0,<0.62.0"
 flake8 = "^7.2.0"
-sphinx = "^8.2.3"
+sphinx = { version = "^8.2.3", python = ">=3.11" }
 recommonmark = "^0.7.1"
 sphinx-rtd-theme = "^3.0.2"
 myst-parser = "^4.0.1"
