### 2025-11-23 3e99f1a
modified python version in sphyinx-autodoc-typehints
