### 2025-11-23 a6b9374
modified python version in sphyinx-sitemap
