### 2025-11-23 9ddb316
modified python version in sphyinx-sitemap
