### 2025-11-23 cc4ef7f
add VisualizationUnavailableWarning
