+++ pyproject.toml
@@ -20,7 +20,7 @@ dependencies = [
     "meshio (>=5.3.5,<6.0.0)",
     "numba (>=0.61.0,<0.62.0)",
     "pyamg (>=5.2.1,<6.0.0)",
-    "pyvista (>=0.44.2,<0.45.0)",
+    "pyvista (>=0.44.2,<0.45.0); python_version < '3.13'",
     "pyproject-toml (>=0.1.0,<0.2.0)",
     "imageio (>=2.37.0,<3.0.0)",
     "scipy (>=1.15.2,<2.0.0)",
@@ -40,13 +40,13 @@ packages = [
 
 [tool.poetry.dependencies]
 python = ">=3.10,<3.14"
-scikit-fem = ">=10.0.2,<11.0.0"
+scikit-fem = ">=11.0,<12.0"
 matplotlib = ">=3.10.1,<4.0.0"
 meshio = ">=5.3.5,<6.0.0"
 numba = ">=0.61.0,<0.62.0"
 pyamg = ">=5.2.1,<6.0.0"
 pyproject-toml = ">=0.1.0,<0.2.0"
-pyvista = ">=0.44.2,<0.45.0"
+pyvista = { version = ">=0.44.2,<0.45.0", python = ">=3.10,<3.13" }
 imageio = ">=2.37.0,<3.0.0"
 
 [tool.poetry.group.dev.dependencies]
