+++ _sources/fea.rst.txt
@@ -1,11 +1,0 @@
-sktopt.fea
-===========
-
-This package provides functions to handle finite element analysis
-
-.. automodule:: sktopt.fea
-   :members:
-   :imported-members:
-   :no-index:
-   :undoc-members:
-   :show-inheritance:
