+++ _sources/core.rst.txt
@@ -1,8 +1,0 @@
-sktopt.core
-===========
-
-.. automodule:: sktopt.core
-   :members:
-   :imported-members:
-   :no-index:
-   :show-inheritance:
