+++ _sources/project.rst.txt
@@ -1,30 +1,0 @@
-Project
-============
-
-
-**Purpose of this Project**
-
- Scikit-Topt is a lightweight and flexible Python library for topology optimization
-  - a computational approach for determining the optimal shape and material layout of structures and components.
-
-Using the finite element method (FEM), topology optimization automatically generates efficient and 
-high-performance designs by redistributing material to achieve a given objective, such as minimizing compliance 
-under a volume constraint.
-
-The library aims to make topology optimization accessible, transparent, and practical for both academic research 
-and real-world engineering. It provides modular tools for assembling finite element models, defining optimization 
-problems, and visualizing results, enabling users to experiment with algorithms, benchmark methods, and integrate 
-topology optimization into their existing simulation workflows.
-
-
-**Contributing**
-
-
-Scikit-Topt is an open-source project, and we warmly welcome community contributions!
-
-If you’d like to contribute:
-- Open an issue to discuss ideas, report bugs, or suggest improvements.
-- Submit a pull request for fixes, enhancements, or new examples.
-
-Please feel free to join the discussion on [GitHub Issues](https://github.com/Kevin-Tofu/scikit-topt/issues).
-We appreciate all forms of contributions, from small documentation improvements to larger feature proposals.
