+++ _sources/optimization.rst.txt
@@ -1,129 +1,0 @@
-Optimization Algorithms in Density-Based Topology Optimization
-================================================================
-
-Density-based topology optimization relies on continuous design variables to represent material distribution, and various optimization algorithms have been developed to update these densities effectively.
-
-One of the most widely used methods is the **Optimality Criteria (OC) method**, which updates the density based on a closed-form expression derived from the Karush-Kuhn-Tucker (KKT) conditions. The OC method is simple to implement and provides fast convergence in many structural problems. It typically uses a multiplicative update rule that enforces a volume constraint while improving the compliance.
-
-Another approach is the **Modified Optimality Criteria (MOC)** method, which extends the OC framework by incorporating additional strategies for handling constraints, improving stability, or enhancing convergence. For example, some MOC variants integrate projection and filtering directly into the update step or apply move limits and continuation schemes to control intermediate densities more robustly.
-
-In addition to these, gradient-based methods using standard optimization libraries (e.g., MMA: Method of Moving Asymptotes) are also popular in density-based formulations, particularly when dealing with multiple constraints or noncompliance objectives.
-
-Each of these methods offers different trade-offs in terms of implementation complexity, convergence speed, and robustness. The choice of optimization algorithm can significantly affect the quality and performance of the final design.
-
-Optimality Criteria
------------------------------------
-
-Optimality Criteria Method (OC)
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-The Optimality Criteria (OC) method is a widely used algorithm in density-based topology optimization due to its simplicity and effectiveness. 
-It provides a closed-form multiplicative update rule for the design variables, derived from the Karush–Kuhn–Tucker (KKT) optimality conditions of the compliance minimization problem.
-
-Assume the goal is to minimize compliance :math:`C(\rho)` subject to a volume constraint:
-
-.. math::
-
-   \min_{\rho} \quad C(\rho) = \mathbf{f}^T \mathbf{u}(\rho) \\
-   \text{subject to} \quad \sum_e \rho_e v_e \leq V_{\text{max}}, 
-   \quad \rho_{\text{min}} \leq \rho_e \leq 1.
-
-The KKT conditions lead to an element-wise update rule involving the compliance sensitivity and a Lagrange multiplier :math:`\lambda`:
-
-.. math::
-
-   \rho_e^{(t+1)} = \text{clip}\left(
-   \rho_e^{(t)} \cdot 
-   \left( \frac{-\partial C / \partial \rho_e}{\lambda} \right)^{\eta},\ 
-   \rho_{\text{min}},\ 1
-   \right).
-
-Here:
-
-- :math:`\partial C / \partial \rho_e` is the sensitivity of compliance with respect to the design variable,
-- :math:`\lambda` is a Lagrange multiplier chosen (typically by bisection) to satisfy the volume constraint,
-- :math:`\eta` is a numerical damping or scaling factor (commonly :math:`\eta = 1`),
-- ``clip`` ensures the updated density stays within prescribed bounds,
-- optional move limits may also be enforced to restrict the maximum per-iteration change in :math:`\rho_e`.
-
-The OC method requires only element-wise sensitivities and a scalar multiplier update, 
-and therefore avoids reliance on general-purpose gradient-based optimizers such as MMA or SQP.
-
-Advantages
-^^^^^^^^^^
-
-- **Simplicity**: Easy to implement and understand.
-- **Efficiency**: Fast convergence in practice for compliance minimization.
[…]
