+++ _sources/mesh.rst.txt
@@ -1,16 +1,0 @@
-sktopt.mesh
-===========
-
-This package provides mesh utilities and task domain definitions.
-
-.. automodule:: sktopt.mesh
-   :members:
-   :imported-members:
-   :no-index:
-   :show-inheritance:
-
-.. toctree::
-   :maxdepth: 1
-
-   mesh.utils
-
