+++ _sources/api.rst.txt
@@ -1,28 +1,0 @@
-API Reference
-=============
-This page provides detailed documentation for modules, classes and functions.
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Core Optimizers
-
-   core
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Finite Element Solvers
-
-   fea
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Mesh and Task Utilities
-
-   mesh
-   mesh.utils
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Tools to manage optimization
-
-   tools

[…]
