+++ .github/workflows/sphinx.yml
@@ -28,20 +28,31 @@ jobs:
         with:
           poetry-version: "1.8.3"
 
+      - name: Install dependencies
+        run: |
+          poetry install --no-interaction --no-root
+          poetry run pip install sphinx
+
       - name: Copy assets to _static
         run: |
           mkdir -p docs/_static
-          cp -r assets/* docs/_static/
-          cp -r paper/figures/* docs/_static/
+          cp -r assets/* docs/_static/ || true
+          cp -r paper/figures/* docs/_static/ || true
 
-      - name: Install dependencies and Build Sphinx documentation
-        working-directory: docs
+      - name: Build Sphinx documentation
         run: |
-          poetry install --no-interaction
-          poetry run sphinx-build -b html . _build/html
+          poetry run sphinx-build -b html docs docs/_build/html
+
+      - name: Prepare dev folder when necessary
+        if: github.ref == 'refs/heads/dev'
+        run: |
+          mkdir -p docs/_build/dev
+          cp -r docs/_build/html/* docs/_build/dev/
 
       - name: Add .nojekyll
-        run: touch docs/_build/html/.nojekyll
+        run: |
+          touch docs/_build/html/.nojekyll
+          touch docs/_build/dev/.nojekyll || true
 
       - name: Deploy to GitHub Pages
         uses: peaceiris/actions-gh-pages@v3
@@ -50,6 +61,6 @@ jobs:
           publish_branch: gh-pages
           publish_dir: |
             ${{
-              github.ref == 'refs/heads/master' && './docs/_build/html' ||
-              format('./docs/_build/html/../dev')
+              github.ref == 'refs/heads/master' && 'docs/_build/html' ||
+              'docs/_build/dev'
             }}
