+++ _sources/installation.rst.txt
@@ -1,0 +1,21 @@
+Installation
+============
+
+sktopt is a Python package that can be installed using pip or poetry.  
+It is compatible with Python 3.11 - 3.12. To install sktopt, follow these steps:
+
+**Install sktopt**:  
+Open your terminal or command prompt and run one of the following commands:
+
+- Using pip::
+
+      pip install scikit-topt
+
+- Using poetry::
+
+      poetry add scikit-topt
+
+**Verify the installation**:  
+After the installation is complete, you can verify that sktopt is installed correctly by running the following command::
+
+    python -c "import sktopt; print(sktopt.__version__)"
