+++ robots.txt
@@ -1,0 +1,4 @@
+User-agent: *
+Allow: /
+
+Sitemap: https://kevin-tofu.github.io/scikit-topt/sitemap.xml
