+++ _sources/index.rst.txt
@@ -1,0 +1,21 @@
+.. scikit-topt documentation master file, created by
+   sphinx-quickstart on Wed May 21 02:20:24 2025.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Scikit Topology Optimization (scikit-topt) documentation
+================================================================
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Content
+
+   project
+   installation
+   tutorial
+   physics
+   formulation
+   optimization
+   api
+
