+++ _sources/core.rst.txt
@@ -1,0 +1,8 @@
+sktopt.core
+===========
+
+.. automodule:: sktopt.core
+   :members:
+   :imported-members:
+   :no-index:
+   :show-inheritance:
