### 2025-11-23 6ab943d
update sphinxyaml
