### 2025-11-24 2c0fb78
update docstring
