+++ docs/index.rst
@@ -1,9 +1,9 @@
-.. scikit-topt documentation master file, created by
+.. Scikit-Topt documentation master file, created by
    sphinx-quickstart on Wed May 21 02:20:24 2025.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Scikit Topology Optimization (scikit-topt) documentation
+Scikit Topology Optimization (Scikit-Topt) documentation
 ================================================================
 
 
