### 2025-11-24 712ef13
Fix: Address all JOSS review comments except item 7
