### 2025-11-24 fac1b99
add  condition for convrgence check
