+++ scikit-topt/sktopt/core/optimizers/logmoc.py
@@ -11,42 +11,50 @@ logger = mylogger(__name__)
 @dataclass
 class LogMOC_Config(common_density.DensityMethod_OC_Config):
     """
-    Configuration for log-space gradient update with optional mean-centering.
+    Configuration for a log-space, OC-like multiplicative density update.
 
-    This configuration supports a variant of log-space Lagrangian descent used in topology
-    optimization. The update is performed in the logarithmic domain to ensure strictly positive
-    densities and to simulate multiplicative behavior. Optionally, the computed update direction
-    can be centered (mean-subtracted) to improve numerical stability.
+    This configuration supports a variant of the Optimality Criteria (OC)
+    method implemented in :math:`\\log \\rho`-space. The update remains
+    multiplicative in :math:`\\rho`, but it is applied by adding a scaled
+    log-factor to :math:`\\log \\rho`. The dual variable associated with
+    the volume constraint is updated by an exponential moving average (EMA)
+    driven by the volume error.
 
-    This method differs from standard OC-based updates: it adds the volume constraint penalty
-    directly to the compliance sensitivity and applies the gradient step in log(ρ)-space:
+    Conceptually, the update has the form
 
-        log(ρ_new) = log(ρ) - η · (∂C/∂ρ + λ)
+    .. math::
 
-    Optionally (when enabled), the update direction is centered as:
+        \\rho^{\\text{new}}
+        = \\rho^{\\text{old}}
+          \\left(-\\frac{\\partial C / \\partial \\rho}{\\lambda_v}
+          \\right)^{\\eta},
 
-        Δ = (∂C/∂ρ + λ) - mean(∂C/∂ρ + λ)
+    where the ratio :math:`-\\partial C/\\partial \\rho / \\lambda_v`
+    is clamped to avoid numerical instabilities and move limits are
+    enforced in log-space.
 
     Attributes
     ----------
     interpolation : Literal["SIMP"]
-        Interpolation scheme used for material penalization. Currently only "SIMP" is supported.
-
+        Interpolation scheme used for material penalization. Currently
+        fixed to ``"SIMP"``.
     mu_p : float
-        Penalty weight for the volume constraint. This controls how strongly the constraint
-        influences the descent direction.
-
+        Gain used to convert the volume error into a dual update. Larger
+        values make the dual variable :attr:`lambda_v` react more strongly
+        to constraint violations.
     lambda_v : float
-        Initial Lagrange multiplier for the volume constraint.
-
+        Initial value for the volume-related dual variable. This is used
+        in the OC-like scaling term :math:`-\\partial C/\\partial \\rho
+        / \\lambda_v`.
     lambda_decay : float
-        Factor by which lambda_v decays over iterations. Smaller values cause λ to adapt more rapidly.
-
+        Decay factor in the EMA update of :attr:`lambda_v`. Values closer
+        to 1.0 yield slower, smoother updates; smaller values react more
+        quickly to changes in the volume error.
     """
 
     interpolation: Literal["SIMP"] = "SIMP"
-    mu_p: float = 3000.0
-    lambda_v: float = 10.0
+    mu_p: float = 1e-1
+    lambda_v: float = 1e+2
     lambda_decay: float = 0.70
 
 
@@ -108,48 +116,46 @@ def moc_log_update_logspace(
 # Lagrangian Dual MOC
 class LogMOC_Optimizer(common_density.DensityMethod):
     """
-    Topology optimization solver using log-space Lagrangian gradient descent.
+    Topology optimization solver using a log-space, OC-like multiplicative update.
 
-    This optimizer performs sensitivity-based topology optimization by applying
[…]
