+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -26,7 +26,7 @@ logging.getLogger("skfem").setLevel(logging.WARNING)
 @dataclass
 class DensityMethodConfig():
     """
-    Configuration for density-based topology optimization (SIMP/RAMP).
+    Configuration for density-based topology optimization (SIMP).
 
     This configuration collects the main numerical settings for
     density-based topology optimization, including interpolation models,
@@ -120,7 +120,38 @@ class DensityMethodConfig():
         and loads for improved numerical conditioning.
     n_joblib : int
         Number of parallel workers used in joblib-enabled sections.
+    check_convergence : bool
+        If ``True``, enable automatic convergence checking during optimization.
+        Convergence is determined by two criteria:
 
+        1. Maximum density change (``tol_rho_change``)
+        2. KKT-like residual or Lagrangian gradient norm (``tol_kkt_residual``)
+
+        Optimization terminates early only when both are satisfied.
+
+    tol_rho_change : float
+        Tolerance for the maximum per-iteration density change.
+        Convergence is considered achieved when::
+
+            max(|rho_new - rho_old|) < tol_rho_change
+
+        Typical values range from ``1e-2`` to ``1e-4`` depending on mesh resolution.
+
+    tol_kkt_residual : float
+        Tolerance for the KKT-style stationarity residual.
+
+        For OC:
+            The residual is computed as the infinity norm of the Lagrangian
+            gradient ``dC/drho + λ * dV/drho`` over interior design elements.
+
+        For MOC / LogMOC:
+            A Lagrangian-like gradient norm based on the current volume-penalty
+            multiplier is used instead, serving as an approximate stationarity
+            measure.
+
+        Convergence is considered achieved when::
+
+            kkt_residual < tol_kkt_residual
     """
 
     dst_path: str = "./result/pytests"
@@ -166,6 +197,10 @@ class DensityMethodConfig():
     solver_option: Literal["spsolve", "cg_pyamg"] = "spsolve"
     scaling: bool = False
     n_joblib: int = 1
+
+    check_convergence: bool = False
+    tol_rho_change: float = 2e-1
+    tol_kkt_residual: float = 5e-3
 
     @classmethod
     def from_defaults(cls, **args) -> 'DensityMethodConfig':
@@ -332,7 +367,7 @@ class DensityMethod(DensityMethodBase):
     Core driver for sensitivity-based density topology optimization.
 
     This class implements the common workflow shared by multiple
-    optimization strategies (e.g., OC, LogMOC). It handles
+    optimization strategies (e.g., OC, MOC). It handles
 
     - problem setup (scaling, restart, output directories),
     - finite element analysis (FEA) for elasticity / heat conduction,
@@ -437,6 +472,14 @@ class DensityMethod(DensityMethodBase):
         # self.recorder = self.add_recorder(tsk)
         self.schedulers = tools.Schedulers(self.cfg.dst_path)
 
+        # params for convergence check
+        self._rho_e_buffer: np.ndarray | None = None
+        self._dC_raw_buffer: np.ndarray | None = None
+        ele_vol_design = tsk.elements_volume[tsk.design_elements]
+        ele_vol_design_sum = np.sum(ele_vol_design)
+        self._dV_drho_design = ele_vol_design / ele_vol_design_sum
+        self.kkt_residual = None
+
     def add_recorder(
         self, tsk: sktopt.mesh.FEMDomain
     ) -> tools.HistoryCollection:
@@ -451,6 +494,9 @@ class DensityMethod(DensityMethodBase):
             recorder.add("u_max")
         recorder.add("compliance", ylog=True)
         recorder.add("scaling_rate", plot_type="min-max-mean-std")
+
+        recorder.add("rho_change_max")
+        recorder.add("kkt_residual")
         return recorder
 
     def params_latest(self):
[…]

[…]
