### 2025-11-24 677daa7
changed representation of filter operator
