### 2025-11-24 b16a7ed
add explanation of lagrangian
