### 2025-11-24 a6266d0
fix broken sphinx math description
