### 2025-11-25 102343e
add members in api.rst
