+++ docs/api.rst
@@ -1,6 +1,20 @@
 API Reference
 =============
-The following sections contain the API documentation for the core modules, classes, and functions.
+
+This section summarizes the main public API of Scikit-Topt.
+For full details, see the module-specific pages below.
+
+Core Optimizers
+---------------
+
+Key classes:
+
+- :class:`sktopt.core.OC_Optimizer`
+- :class:`sktopt.core.LogMOC_Optimizer`
+- :class:`sktopt.core.DensityMethodConfig`
+- :class:`sktopt.core.OptimizationHistory`
+
+Full API documentation: :mod:`sktopt.core`
 
 .. toctree::
    :maxdepth: 1
@@ -8,11 +22,48 @@ The following sections contain the API documentation for the core modules, class
 
    core
 
+
+Finite Element Solvers
+----------------------
+
+Key classes:
+
+- :class:`sktopt.fea.composer.LinearElasticity`
+- :class:`sktopt.fea.composer.SteadyStateHeatConduction`
+
+Full API documentation: :mod:`sktopt.fea`
+
 .. toctree::
    :maxdepth: 1
    :caption: Finite Element Solvers
 
    fea
+
+
+Tools
+-----
+
+Key utilities:
+
+- :class:`sktopt.tools.SchedulerConfig`
+- :class:`sktopt.tools.HistoryLogger`
+- :mod:`sktopt.tools.logconf`
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Tools to manage optimization
+
+   tools
+
+
+Mesh and Task Utilities
+-----------------------
+
+Key utilities:
+
+- :mod:`sktopt.mesh.toy_problem`
+- :mod:`sktopt.mesh.loader`
+- :mod:`sktopt.mesh.utils`
 
 .. toctree::
    :maxdepth: 1
@@ -20,9 +71,3 @@ The following sections contain the API documentation for the core modules, class
 
    mesh
    mesh.utils
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Tools to manage optimization
-
-   tools
