### 2025-11-25 96bc48a
add core
