### 2025-11-25 d1974e3
add docstring for schedulers
