+++ docs/core.rst
@@ -6,5 +6,4 @@ sktopt.core
 .. automodule:: sktopt.core
    :members:
    :imported-members:
-   :no-index:
    :show-inheritance:
