+++ docs/mesh.rst
@@ -6,11 +6,4 @@ This package provides mesh utilities and task domain definitions.
 .. automodule:: sktopt.mesh
    :members:
    :imported-members:
-   :no-index:
    :show-inheritance:
-
-.. toctree::
-   :maxdepth: 1
-
-   mesh.utils
-
