+++ docs/mesh.utils.rst
@@ -3,6 +3,5 @@ sktopt.mesh.utils
 
 .. automodule:: sktopt.mesh.utils
    :members:
-   :no-index:
    :undoc-members:
    :show-inheritance:
