+++ _sources/optimization.rst.txt
@@ -5,7 +5,7 @@ Density-based topology optimization relies on continuous design variables to rep
 
 One of the most widely used methods is the **Optimality Criteria (OC) method**, which updates the density based on a closed-form expression derived from the Karush-Kuhn-Tucker (KKT) conditions. The OC method is simple to implement and provides fast convergence in many structural problems. It typically uses a multiplicative update rule that enforces a volume constraint while improving the compliance.
 
-Another approach is the **Modified Optimality Criteria (MOC)** method, which extends the OC framework by incorporating additional strategies for handling constraints, improving stability, or enhancing convergence. For example, some MOC variants integrate projection and filtering directly into the update step or apply move limits and continuation schemes to control intermediate densities more robustly.
+Another approach is the **Modified Optimality Criteria (MOC)** method, which extends the OC framework by incorporating additional strategies for handling constraints, improving stability, or enhancing convergence. For example, MOC integrate projection and filtering directly into the update step or apply move limits and continuation schemes to control intermediate densities more robustly.
 
 In addition to these, gradient-based methods using standard optimization libraries (e.g., MMA: Method of Moving Asymptotes) are also popular in density-based formulations, particularly when dealing with multiple constraints or noncompliance objectives.
 
@@ -46,8 +46,34 @@ Here:
 - ``clip`` ensures the updated density stays within prescribed bounds,
 - optional move limits may also be enforced to restrict the maximum per-iteration change in :math:`\rho_e`.
 
-The OC method requires only element-wise sensitivities and a scalar multiplier update, 
-and therefore avoids reliance on general-purpose gradient-based optimizers such as MMA or SQP.
+Bisection for the Volume Constraint
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+The Lagrange multiplier :math:\lambda enforcing the volume constraint is determined by a scalar bisection procedure.
+This procedure is the same for both OC and MOC formulations, since the volume constraint enters only through this scalar multiplier.
+For each trial value of :math:\lambda, the densities are tentatively updated using the OC/MOC rule, the resulting volume is evaluated, and the bisection bounds are adjusted accordingly.
+Because this step requires only vectorized density updates—without any additional finite element analyses—it is computationally inexpensive.
+
+A typical implementation is shown in the following pseudo-code:
+
+.. code-block:: none
+
+   λ_low  = 1e−9
+   λ_high = 1e+9
+
+   while (λ_high − λ_low) / (λ_high + λ_low) > tol:
+
+       λ = (λ_low + λ_high) / 2
+       ρ_trial = OC_Update(ρ, ∂C/∂ρ, λ)
+       V = Volume(ρ_trial)
+
+       if V > V_max:
+           λ_low = λ        # too much material
+       else:
+           λ_high = λ       # too little material
+
+   λ_final = λ
+
 
 Advantages
 ^^^^^^^^^^
@@ -68,8 +94,8 @@ Despite these limitations, the OC method remains one of the most popular approac
 especially when a robust and lightweight heuristic is sufficient for prototyping or academic demonstration.
 
 
-Modified Optimality Criteria (MOC) Variants
--------------------------------------------
+Modified Optimality Criteria (MOC)
+-----------------------------------
 
 In density-based topology optimization, the Modified Optimality Criteria (MOC) method can be implemented in several ways. 
[…]
