+++ _sources/core.rst.txt
@@ -1,8 +1,9 @@
+.. _core_api:
+
 sktopt.core
 ===========
 
 .. automodule:: sktopt.core
    :members:
    :imported-members:
-   :no-index:
    :show-inheritance:
