+++ _sources/formulation.rst.txt
@@ -17,10 +17,12 @@ Where:
 - :math:`E_0` and :math:`E_\text{min}` are the maximum and minimum Young's moduli.
 - :math:`p` and :math:`q` are penalization parameters.
 
+Note: \(E_\text{min}\) must be strictly positive to avoid singular stiffness
+matrices. A common choice is \(E_\text{min} \approx 10^{-3} E_0\).
+
 Initialization
 ~~~~~~~~~~~~~~
 
-- Design variable :math:`\rho` is initialized using the ``vol_frac_init`` or a fixed value depending on the interpolation scheme.
 - Restart logic loads a previous design if specified.
 - Dirichlet and force boundary elements are set to full material.
 
@@ -105,37 +107,67 @@ Compliance Evaluation
 
      \frac{\partial C}{\partial \rho} = -p \rho^{p-1} (E_0 - E_\text{min}) \epsilon^T \mathbf{D} \epsilon
 
-Sensitivity Analysis with Backpropagation
---------------------------------------------
+Sensitivity Propagation Through Filtering and Projection
+--------------------------------------------------------
 
-In topology optimization, the sensitivity of the objective function (e.g., compliance) with respect to the design variable :math:`\rho` is computed by applying the chain rule through each computational step. This is conceptually similar to backpropagation in machine learning.
-
-Assume the objective function :math:`C` depends on the displacement field :math:`\mathbf{u}`, which in turn depends on the projected density :math:`\tilde{\rho}`, which is computed from the filtered density :math:`\rho_{\text{filtered}}`, and ultimately from the design variable :math:`\rho`.
-
-The total derivative is written as:
+In density-based topology optimization, the compliance sensitivity computed
+from FEM does not act directly on the design density. Instead, the design
+variable passes through two transformations:
 
 .. math::
+   \rho
+   \quad\xrightarrow{\text{filter}}\quad
+   \tilde{\rho}
+   \quad\xrightarrow{\text{projection}}\quad
+   \hat{\rho}.
 
+The total derivative uses the chain rule.
+
+We denote by :math:`\mathcal{F}` the (linear) density–filter operator such that
+
+.. math::
+   \tilde{\rho} = \mathcal{F}(\rho).
+
+Because :math:`\mathcal{F}` is linear, its derivative with respect to
+:math:`\rho` is the operator itself, and the adjoint filter required for
+backpropagation is the transpose operator :math:`\mathcal{F}^T`,
+implemented via ``filter.gradient()`` in Scikit-Topt.
+
+Applying the chain rule gives
+
+.. math::
    \frac{\partial C}{\partial \rho}
    =
-   \frac{\partial C}{\partial \mathbf{u}}
+   \frac{\partial C}{\partial \hat{\rho}}
    \cdot
-   \frac{\partial \mathbf{u}}{\partial \tilde{\rho}}
+   \frac{\partial \hat{\rho}}{\partial \tilde{\rho}}
    \cdot
-   \frac{\partial \tilde{\rho}}{\partial \rho_{\text{filtered}}}
-   \cdot
-   \frac{\partial \rho_{\text{filtered}}}{\partial \rho}
+   \frac{\partial \tilde{\rho}}{\partial \rho}.
 
-Each term corresponds to:
+Meaning of each term:
 
-- :math:`\frac{\partial C}{\partial \mathbf{u}}`: the derivative of compliance with respect to displacement, typically equal to the load vector :math:`\mathbf{f}`.
-- :math:`\frac{\partial \mathbf{u}}{\partial \tilde{\rho}}`: the derivative of the displacement field with respect to material stiffness, obtained via the adjoint method or differentiating the FEM equilibrium equation.
[…]
