+++ _sources/physics-heat.rst.txt
@@ -23,6 +23,11 @@ where
 - :math:`k(\mathbf{x})` is the thermal conductivity  
 - :math:`T` is the unknown temperature
 
+
+Here, the thermal conductivity :math:`k(\mathbf{x})` is treated as a scalar
+(isotropic conductivity). Anisotropic or tensor-valued conductivity is not
+considered in the current implementation.
+
 Boundary Conditions
 -------------------
 

[…]
