+++ scikit-topt/sktopt/fea/solver_elastic.py
@@ -145,86 +145,37 @@ def solve_multi_load(
     elem_func: Callable = composer.simp_interpolation,
     rtol: float = 1e-5,
     maxiter: int = None,
-    n_joblib: int = 1
-) -> float:
-    solver = 'spsolve' if solver == 'auto' else solver
-    n_dof = basis.N
-    assert u_all.shape == (n_dof, len(force_list))
+) -> np.ndarray:
+    """
+    Solve multi-load linear system using one LU factorization and
+    solving all RHS in a single call: lu.solve(F_stack).
 
+    This is the fastest and safest way for shared-stiffness multi-load problems.
+    """
+
+    # Assemble stiffness matrix
     K = composer.assemble_stiffness_matrix(
         basis, rho, E0, Emin, p, nu0, elem_func
     )
-    _maxiter = min(1000, max(300, n_dof // 5)) if maxiter is None else maxiter
     K_csr = K.tocsr()
 
-    # all_dofs = np.arange(n_dof)
-    # free_dofs = np.setdiff1d(all_dofs, dirichlet_dofs, assume_unique=True)
-    # K_e = K_csr[free_dofs, :][:, free_dofs]
-    # K_e, _ = skfem.enforce(
-    #     K_csr[free_dofs, :][:, free_dofs], force_list[0],
-    #     D=dirichlet_dofs
-    # )
-    # K_c, _, _, _ = skfem.condense(K_csr, force_list[0], D=dirichlet_dofs)
+    # Enforced K for first load (pattern only — K_e is reused)
     K_e, _ = skfem.enforce(K_csr, force_list[0], D=dirichlet_dofs)
+
+    # Build all RHS after enforcement
     F_stack = np.column_stack([
         skfem.enforce(K_csr, f, D=dirichlet_dofs)[1] for f in force_list
     ])
-    compliance_total = 0.0
-    u_all[:, :] = 0.0
-    if solver == 'spsolve':
-        try:
-            from joblib import Parallel, delayed, parallel_backend
-            if n_joblib > 1:
-                lu = scipy.sparse.linalg.splu(K_e.tocsc())
 
-                def solve_system(F_stack):
-                    return lu.solve(F_stack)
+    # LU factorization only once
+    lu = scipy.sparse.linalg.splu(K_e.tocsc())
 
-                with parallel_backend("threading"):
-                    u_all[:, :] = np.column_stack(
-                        Parallel(n_jobs=n_joblib)(
-                            delayed(solve_system)(F_stack[:, i]) for i in range(
-                                F_stack.shape[1]
-                            )
-                        )
-                    )
-                return F_stack
-        except ModuleNotFoundError as e:
-            logger.info(f"ModuleNotFoundError: {e}")
-            n_joblib = -1
+    # Solve all RHS at once (fastest)
+    u_sol = lu.solve(F_stack)      # shape = (ndof, nloads)
 
-        lu = scipy.sparse.linalg.splu(K_e.tocsc())
-        u_all[:, :] = np.column_stack(
-            [lu.solve(F_stack[:, i]) for i in range(F_stack.shape[1])]
-        )
+    # Store to user-provided array
+    u_all[:, :] = u_sol
 
-    else:
-        # choose preconditioner if needed
-        if solver == 'cg_jacobi':
-            M_diag = K_e.diagonal()
-            M_inv = 1.0 / M_diag
-            M = LinearOperator(K_e.shape, matvec=lambda x: M_inv * x)
-        elif solver == 'cg_pyamg':
-            ml = pyamg.smoothed_aggregation_solver(K_e)
-            M = ml.aspreconditioner()
-        else:
-            raise ValueError(f"Unknown solver: {solver}")
-
-        for i, _ in enumerate(force_list):
-            F_e = F_stack[:, i]
[…]
