+++ scikit-topt/sktopt/fea/solver_elastic.py
@@ -18,27 +18,10 @@ from sktopt.tools.logconf import mylogger
 logger = mylogger(__name__)
 
 
-def compute_compliance_simp_basis(
-    basis, free_dofs, dirichlet_dofs, force,
-    E0, Emin, p, nu0,
-    rho,
-) -> tuple:
-    K = composer.assemble_stiffness_matrix(
-        basis, rho, E0,
-        Emin, p, nu0
-    )
-    K_e, F_e = skfem.enforce(K, force, D=dirichlet_dofs)
-    # u = scipy.sparse.linalg.spsolve(K_e, F_e)
-    u = skfem.solve(K_e, F_e)
-    f_free = force[free_dofs]
-    compliance = f_free @ u[free_dofs]
-    return (compliance, u)
-
-
 def solve_u(
-    K_cond: scipy.sparse.csc_matrix,
+    K_cond: scipy.sparse.spmatrix,
     F_cond: np.ndarray,
-    chosen_solver: Literal['cg_jacobi', 'spsolve', 'cg_pyamg'] = 'auto',
+    chosen_solver: Literal['cg_jacobi', 'spsolve', 'cg_pyamg'] = 'spsolve',
     rtol: float = 1e-8,
     maxiter: int = None,
 ) -> np.ndarray:
@@ -74,7 +57,7 @@ def solve_u(
             raise ValueError(f"Unknown solver: {chosen_solver}")
 
     except Exception as e:
-        print(f"Solver exception - {e}, falling back to spsolve.")
+        logger.warning(f"Solver exception - {e}, falling back to spsolve.")
         u_c = scipy.sparse.linalg.spsolve(K_cond, F_cond)
 
     return u_c
@@ -144,36 +127,92 @@ def solve_multi_load(
     solver: Literal['auto', 'cg_jacobi', 'spsolve', 'cg_pyamg'] = 'auto',
     elem_func: Callable = composer.simp_interpolation,
     rtol: float = 1e-5,
-    maxiter: int = None,
+    maxiter: int | None = None,
 ) -> np.ndarray:
     """
-    Solve multi-load linear system using one LU factorization and
-    solving all RHS in a single call: lu.solve(F_stack).
+    Solve a shared-stiffness linear elasticity problem for multiple load cases.
 
-    This is the fastest and safest way for shared-stiffness multi-load problems.
+    The stiffness matrix K(ρ) is assembled once, and Dirichlet boundary
+    conditions are enforced via `skfem.enforce`. All load cases share the
+    same Dirichlet DOF set; only the right-hand side (Neumann loads) differs.
+
+    This function uses a single LU factorization of the enforced matrix K_e
+    and solves all right-hand sides in one call:
+
+        u_all = lu.solve(F_stack)
+
+    which is typically the fastest and most robust approach for
+    shared-stiffness multi-load problems.
+
+    Parameters
+    ----------
+    basis : skfem.CellBasis
+        Finite element basis for the displacement field.
+    free_dofs : np.ndarray
+        Array of free DOF indices (currently unused; kept for API compatibility).
+    dirichlet_dofs : np.ndarray
+        DOF indices with Dirichlet boundary conditions, shared across all loads.
+    force_list : list of (n_dof,) ndarray
+        List of global load vectors, one per load case.
+    E0, Emin : float
+        Maximum and minimum Young's modulus values for the SIMP interpolation.
+    p : float
+        SIMP penalization exponent.
+    nu0 : float
+        Poisson's ratio (assumed constant).
+    rho : (n_elem,) ndarray
+        Element-wise density field.
+    u_all : (n_dof, n_loads) ndarray
+        Output array to store displacement solutions; each column is one load case.
+    solver : {'auto', 'cg_jacobi', 'spsolve', 'cg_pyamg'}, optional
[…]
