+++ examples/tutorial/box_oc_heat.py
@@ -145,8 +145,7 @@ def get_cfg():
         max_iters=40,
         record_times=40,
         # filter_type="spacial"
-        filter_type="helmholtz",
-        check_convergence=True
+        filter_type="helmholtz"
     )
     return cfg
 
