+++ scikit-topt/sktopt/fea/solver_heat.py
@@ -15,46 +15,113 @@ from sktopt.tools.logconf import mylogger
 logger = mylogger(__name__)
 
 
-def solve_scipy(
-    K_csr, emit,
-    dirichlet_dofs_list, dirichlet_values_list,
-    u_all
-):
-    def solve_system(
-        dirichlet_dofs_loop, dirichlet_values_loop
-    ):
-        T_sol = skfem.solve(
-            *skfem.condense(
-                K_csr, emit, D=dirichlet_dofs_loop,
-                x=np.full(dirichlet_dofs_loop.N, dirichlet_values_loop)
-            )
+def solve_scipy_heat_multi_enforce(
+    K_csr: scipy.sparse.csr_matrix,
+    emit: np.ndarray,
+    dirichlet_dofs_list: list[skfem.Dofs],
+    dirichlet_values_list: list[float],
+    u_all: np.ndarray,
+) -> None:
+    """
+    General multi-load heat solver, no joblib.
+
+    - For each load i, applies enforce(K, emit, D_i, x_i) and solves.
+    - Does NOT assume all Dirichlet DOFs are identical across loads.
+    """
+
+    n_loads = len(dirichlet_values_list)
+    assert u_all.shape[1] == n_loads
+
+    for i in range(n_loads):
+        D_i = dirichlet_dofs_list[i]
+        val_i = dirichlet_values_list[i]
+
+        K_e_i, f_e_i = skfem.enforce(
+            K_csr, emit,
+            D=D_i,
+            x=np.full(D_i.N, val_i),
         )
-        return T_sol
+        lu_i = scipy.sparse.linalg.splu(K_e_i.tocsc())
+        u_all[:, i] = lu_i.solve(f_e_i)
 
-    try:
-        from joblib import Parallel, delayed, parallel_backend
-        if n_joblib > 1:
-            with parallel_backend("threading"):
-                u_all[:, :] = np.column_stack(
-                    Parallel(n_jobs=n_joblib)(
-                        delayed(solve_system)(
-                            dirichlet_dofs_list[i], dirichlet_values_list[i]
-                        ) for i in range(len(dirichlet_values_list))
-                    )
-                )
-            return
 
-    except ModuleNotFoundError as e:
-        logger.info(f"ModuleNotFoundError: {e}")
-        n_joblib = -1
+def solve_scipy(
+    K_csr,
+    emit,
+    dirichlet_dofs_list,
+    dirichlet_values_list,
+    u_all,
+):
+    """
+    Multi-load heat solver (no joblib, enforce-based).
 
-    u_all[:, :] = np.column_stack(
-        [
-            solve_system(
-                dirichlet_dofs_list[i], dirichlet_values_list[i]
-            ) for i in range(len(dirichlet_values_list))
-        ]
+    Assumptions
+    -----------
+    - All load cases share the *same* Dirichlet DOF set.
+      (i.e., dirichlet_dofs_list[i] are identical objects)
+    - Only the Dirichlet *values* differ across load cases.
+    - The RHS contribution "emit" is common to all load cases.
+
+    Method
+    ------
+    1. Build the enforced stiffness matrix K_e using the first load case.
+       (K_e depends on the Dirichlet DOF set D, not on the actual values)
+    2. Factorize K_e only once (LU).
+    3. For each load case, build only the enforced RHS f_e
+       using skfem.enforce(K, emit, D, x=value).
+    4. Stack all RHS vectors column-wise into F_stack.
+    5. Solve all RHS simultaneously using lu.solve(F_stack).
+       This is significantly faster than solving each load separately.
+
+    Parameters
+    ----------
+    K_csr : csr_matrix
+        Global conduction matrix.
+    emit : (n_dof,) ndarray
[…]
