+++ README.md
@@ -134,7 +134,6 @@ Scikit-Topt is standing on the shoulders of proverbial giants. In particular, I 
  - MeshIO
  - Matplotlib
  - PyVista
- - Joblib
  - Topology Optimization Community
 
 
