+++ pyproject.toml
@@ -25,7 +25,6 @@ dependencies = [
     "pyproject-toml (>=0.1.0,<0.2.0)",
     "imageio (>=2.37.0,<3.0.0)",
     "scipy (>=1.15.2,<2.0.0)",
-    "joblib (>=1.5.1,<2.0.0)",
 ]
 
 [tool.poetry]
