### 2025-11-30 5a4f0ed
add State in Optimizer and can be run with optimize_step
