+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -24,6 +24,30 @@ logging.getLogger("skfem").setLevel(logging.WARNING)
 
 
 @dataclass
+class DensityIterState:
+    rho: np.ndarray
+    rho_prev: np.ndarray
+    rho_filtered: np.ndarray
+    rho_projected: np.ndarray
+    dH_drho: np.ndarray
+    grad_filtered: np.ndarray
+    dC_drho_projected: np.ndarray
+    energy_mean: np.ndarray
+    dC_drho_full: np.ndarray
+    dC_drho_design_eles: np.ndarray
+    scaling_rate: np.ndarray
+    rho_design_eles: np.ndarray
+    rho_clip_lower: np.ndarray
+    rho_clip_upper: np.ndarray
+    u_dofs: np.ndarray
+    filter_radius: float
+    elements_volume_design: np.ndarray
+    elements_volume_design_sum: float
+    iter_begin: int
+    iter_end: int
+
+
+@dataclass
 class DensityMethodConfig():
     """
     Configuration for density-based topology optimization (SIMP).
@@ -466,6 +490,7 @@ class DensityMethod(DensityMethodBase):
             )
         # self.recorder = self.add_recorder(tsk)
         self.schedulers = tools.Schedulers(self.cfg.dst_path)
+        self._schedulers_initialized = False
 
         # params for convergence check
         self._rho_e_buffer: np.ndarray | None = None
@@ -474,6 +499,11 @@ class DensityMethod(DensityMethodBase):
         ele_vol_design_sum = np.sum(ele_vol_design)
         self._dV_drho_design = ele_vol_design / ele_vol_design_sum
         self.kkt_residual = None
+        # iterative state for partial runs
+        self._state: DensityIterState | None = None
+        self._iter_next: int | None = None
+        self._iter_end: int | None = None
+        self._completed: bool = False
 
     def add_recorder(
         self, tsk: sktopt.mesh.FEMDomain
@@ -541,6 +571,7 @@ class DensityMethod(DensityMethodBase):
 
         if export:
             self.schedulers.export()
+        self._schedulers_initialized = True
 
     def parameterize(self):
 
@@ -707,41 +738,155 @@ class DensityMethod(DensityMethodBase):
         )
 
     def optimize(self):
+        """Run optimization until ``cfg.max_iters`` (default behavior)."""
+        self._optimize_impl()
+
+    def optimize_steps(self, num_steps: int):
+        """
+        Run only ``num_steps`` iterations of the optimizer and record progress.
+
+        This is useful when an external driver wants to probe or checkpoint
+        intermediate states without waiting for the full ``max_iters`` loop.
+        """
+        if num_steps <= 0:
+            logger.info("optimize_steps called with non-positive num_steps; skipping.")
+            return
+        self._optimize_impl(num_steps)
+
+    def _ensure_state_initialized(self):
+        """Initialize schedulers and iteration state if not already done."""
+        if self._completed:
+            logger.info("Optimization already completed; skipping.")
+            return False
+
+        if not self._schedulers_initialized:
+            self.init_schedulers()
+
+        if self._state is None:
+            (
+                iter_begin,
+                iter_end,
+                rho,
+                rho_prev,
+                rho_filtered,
+                rho_projected,
+                dH_drho,
+                grad_filtered,
+                dC_drho_projected,
+                energy_mean,
+                dC_drho_full,
+                dC_drho_design_eles,
+                scaling_rate,
+                rho_design_eles,
+                rho_clip_lower,
+                rho_clip_upper,
+                u_dofs,
+                filter_radius
+            ) = self.initialize_params()
+            elements_volume_design = self.tsk.elements_volume[self.tsk.design_elements]
[…]
