### 2025-11-30 db0e4cd
make DensityState public
