### 2025-11-30 13e44af
add variables for evaluation
