+++ scikit-topt/sktopt/tools/timer.py
@@ -1,0 +1,528 @@
+import time
+from collections import defaultdict
+from contextlib import contextmanager
+from dataclasses import dataclass
+from typing import Callable, DefaultDict, Iterator, List, Optional
+
+from sktopt.tools.logconf import mylogger
+
+logger = mylogger(__name__)
+
+
+@dataclass
+class SectionStats:
+    name: str
+    count: int
+    total: float
+    avg: float
+    max: float
+
+
+class SectionTimer:
+    """
+    Lightweight helper to measure how long named sections take.
+
+    Examples
+    --------
+    >>> timer = SectionTimer()
+    >>> with timer.section("assemble"):
+    ...     ...
+    >>> with timer.section("solve"):
+    ...     ...
+    >>> timer.report()
+
+    To track nested calls, enable ``hierarchical`` so section names
+    are recorded with their call stack (e.g., ``outer>inner``)::
+
+        timer = SectionTimer(hierarchical=True)
+        with timer.section("outer"):
+            with timer.section("inner"):
+                ...
+    """
+
+    def __init__(
+        self,
+        clock: Optional[Callable[[], float]] = None,
+        hierarchical: bool = False,
+        sep: str = ">",
+    ):
+        self._clock: Callable[[], float] = clock or time.perf_counter
+        self._records: DefaultDict[str, List[float]] = defaultdict(list)
+        self._hierarchical = hierarchical
+        self._sep = sep
+        self._stack: List[str] = []
+
+    @contextmanager
+    def section(self, name: str) -> Iterator[None]:
+        if self._hierarchical and self._stack:
+            full_name = self._sep.join([*self._stack, name])
+        else:
+            full_name = name
+        self._stack.append(name)
+        start = self._clock()
+        try:
+            yield
+        finally:
+            duration = self._clock() - start
+            self._records[full_name].append(duration)
+            self._stack.pop()
+
+    def wrap(self, name: str):
+        """
+        Decorator form of :meth:`section`.
+
+        This is convenient for quickly instrumenting functions without
+        rewriting call sites::
+
+            @timer.wrap("my_step")
+            def my_step(...):
+                ...
+        """
+
+        def _decorator(func):
+            def _wrapper(*args, **kwargs):
+                with self.section(name):
+                    return func(*args, **kwargs)
+            return _wrapper
+
+        return _decorator
+
+    def add(self, name: str, duration: float) -> None:
+        self._records[name].append(duration)
+
+    def reset(self, name: Optional[str] = None) -> None:
+        if name is None:
+            self._records.clear()
+        else:
+            self._records.pop(name, None)
+
+    def stats(self) -> List[SectionStats]:
+        stats: List[SectionStats] = []
+        for name, values in self._records.items():
+            if not values:
+                continue
+            total = sum(values)
+            stats.append(
+                SectionStats(
+                    name=name,
+                    count=len(values),
+                    total=total,
+                    avg=total / len(values),
+                    max=max(values),
+                )
+            )
+        return stats
+
+    def summary(
+        self, sort_by: str = "total", descending: bool = True
+    ) -> List[SectionStats]:
+        key_map = {
+            "total": lambda s: s.total,
+            "avg": lambda s: s.avg,
+            "max": lambda s: s.max,
+            "count": lambda s: s.count,
+            "name": lambda s: s.name,
+        }
+        try:
+            key_func = key_map[sort_by]
+        except KeyError as exc:
+            raise ValueError(
+                'sort_by must be one of {"total", "avg", "max", "count", "name"}'
+            ) from exc
+
+        return sorted(self.stats(), key=key_func, reverse=descending)
+
+    def report(
+        self,
+        sort_by: str = "total",
[…]
