+++ scikit-topt/sktopt/tools/__init__.py
@@ -4,6 +4,7 @@ from sktopt.tools.scheduler import Schedulers
 from sktopt.tools.scheduler import SchedulerStepAccelerating
 from sktopt.tools.scheduler import SchedulerStep
 from sktopt.tools.scheduler import SchedulerSawtoothDecay
+from sktopt.tools.timer import SectionTimer
 
 HistoryCollection.__module__ = __name__
 SchedulerConfig.__module__ = __name__
@@ -11,6 +12,7 @@ Schedulers.__module__ = __name__
 SchedulerStepAccelerating.__module__ = __name__
 SchedulerStep.__module__ = __name__
 SchedulerSawtoothDecay.__module__ = __name__
+SectionTimer.__module__ = __name__
 
 __all__ = [
     "HistoryCollection",
@@ -19,4 +21,5 @@ __all__ = [
     "SchedulerStep",
     "SchedulerStepAccelerating",
     "SchedulerSawtoothDecay",
+    "SectionTimer",
 ]
