+++ scikit-topt/sktopt/core/optimizers/common_density.py
@@ -241,6 +241,8 @@ class DensityMethodConfig():
         import json
         with open(f"{path}/cfg.json", "r") as f:
             data = json.load(f)
+        # Drop legacy keys that may linger in old configs.
+        data.pop("record_timing", None)
         return cls(**data)
 
     def export(self, path: str):
@@ -459,6 +461,7 @@ class DensityMethod(DensityMethodBase):
     ):
         self.cfg = cfg
         self.tsk = tsk
+        self.timer: tools.SectionTimer = tools.SectionTimer(hierarchical=True)
         if cfg.scaling is True:
             self.scale()
 
@@ -818,6 +821,23 @@ class DensityMethod(DensityMethodBase):
             self._iter_end = iter_end
         return True
 
+    def _timed_section(self, name: str):
+        return self.timer.section(name)
+
+    def _report_timing(self):
+        logger.info("--- Timing summary ---")
+        self.timer.report(logger_instance=logger)
+
+    def _save_timing_plot(self, filename: str | None = None):
+        if not self.timer.stats():
+            return
+        path = filename or os.path.join(self.cfg.dst_path, "timing.png")
+        try:
+            self.timer.save_plot(path, format_nested=True, stacked_nested=True)
+        except ValueError:
+            # Raised when no data; skip plot
+            pass
+
     def _finalize(self):
         """Finalize outputs after completing all iterations."""
         if self._completed:
@@ -915,130 +935,132 @@ class DensityMethod(DensityMethodBase):
                 self.filter.update_radius(filter_radius)
 
             logger.info("--- project and filter ---")
-            rho_prev[:] = rho[:]
-            rho_filtered[:] = self.filter.forward(rho)
-            projection.heaviside_projection_inplace(
-                rho_filtered, beta=beta, eta=cfg.beta_eta, out=rho_projected
-            )
+            with self._timed_section("filter_and_project"):
+                rho_prev[:] = rho[:]
+                rho_filtered[:] = self.filter.forward(rho)
+                projection.heaviside_projection_inplace(
+                    rho_filtered, beta=beta, eta=cfg.beta_eta, out=rho_projected
+                )
             logger.info("--- compute compliance ---")
             dC_drho_design_eles[:] = 0.0
             dC_drho_full[:] = 0.0
             energy_mean[:] = 0.0
             u_max = list()
 
-            compliance_avg = self.fem.objectives_multi_load(
-                rho_projected, p, u_dofs
-            ).mean()
-            energy = self.fem.energy_multi_load(
-                rho_projected, p, u_dofs,
-            )
-            energy_mean[:] = energy.mean(axis=1)
-            state.compliance = float(compliance_avg)
-            for task_loop in range(self.tsk.n_tasks):
-
-                u_max.append(np.abs(u_dofs[:, task_loop]).max())
-                dH_drho[:] = 0.0
-                np.copyto(
-                    dC_drho_projected,
-                    dC_drho_func(
-                        rho_projected,
-                        energy[:, task_loop],
-                        self.tsk.material_coef,
-                        self.tsk.material_coef*cfg.E_min_coeff,
-                        p
+            with self._timed_section("objective_and_energy"):
+                with self._timed_section("objective"):
+                    compliance_avg = self.fem.objectives_multi_load(
+                        rho_projected, p, u_dofs
+                    ).mean()
+                with self._timed_section("energy"):
+                    energy = self.fem.energy_multi_load(
+                        rho_projected, p, u_dofs,
                     )
-                )
-                projection.heaviside_projection_derivative_inplace(
-                    rho_filtered,
-                    beta=beta, eta=cfg.beta_eta, out=dH_drho
-                )
[…]
