+++ scikit-topt/sktopt/fea/solver_elastic.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 from typing import Callable, Literal
+from contextlib import contextmanager
 
 import numpy as np
 import scipy
@@ -71,10 +72,19 @@ def compute_compliance_basis(
     solver: Literal['auto', 'cg_jacobi', 'spsolve', 'cg_pyamg'] = 'auto',
     rtol: float = 1e-5,
     maxiter: int = None,
+    timer=None,
 ) -> tuple:
-    K = composer.assemble_stiffness_matrix(
-        basis, rho, E0, Emin, p, nu0, elem_func
-    )
+    def section(name: str):
+        if timer:
+            return timer.section(name)
+        @contextmanager
+        def _noop():
+            yield
+        return _noop()
+    with section("assemble"):
+        K = composer.assemble_stiffness_matrix(
+            basis, rho, E0, Emin, p, nu0, elem_func
+        )
     n_dof = K.shape[0]
     # Solver auto-selection
     if solver == 'auto':
@@ -95,11 +105,13 @@ def compute_compliance_basis(
     free_dofs = np.setdiff1d(all_dofs, dirichlet_dofs, assume_unique=True)
 
     # enforce
-    K_e, F_e = skfem.enforce(K_csr, force, D=dirichlet_dofs)
-    u = solve_u(
-        K_e, F_e, chosen_solver=chosen_solver,
-        rtol=rtol, maxiter=_maxiter
-    )
+    with section("enforce_bc"):
+        K_e, F_e = skfem.enforce(K_csr, force, D=dirichlet_dofs)
+    with section("solve"):
+        u = solve_u(
+            K_e, F_e, chosen_solver=chosen_solver,
+            rtol=rtol, maxiter=_maxiter
+        )
 
     # condense
     # K_c, F_c, U_c, I = skfem.condense(K, F, D=fixed_dofs)
@@ -128,6 +140,7 @@ def solve_multi_load(
     elem_func: Callable = composer.simp_interpolation,
     rtol: float = 1e-5,
     maxiter: int | None = None,
+    timer=None,
 ) -> np.ndarray:
     """
     Solve a shared-stiffness linear elasticity problem for multiple load cases.
@@ -181,6 +194,21 @@ def solve_multi_load(
         Stack of enforced right-hand sides for each load case.
     """
 
+    # fall back to single-load path to honor chosen solver/timer
+    if len(force_list) == 1:
+        compliance, u = compute_compliance_basis(
+            basis, free_dofs, dirichlet_dofs, force_list[0],
+            E0, Emin, p, nu0,
+            rho,
+            elem_func=elem_func,
+            solver=solver,
+            rtol=rtol,
+            maxiter=maxiter,
+            timer=timer,
+        )
+        u_all[:, 0] = u
+        return np.array([compliance])
+
     # Multi-load currently supports only direct LU factorization.
     if solver not in ('auto', 'spsolve'):
         raise NotImplementedError(
@@ -189,28 +217,41 @@ def solve_multi_load(
             "not implemented."
         )
 
+    def section(name: str):
+        if timer:
+            return timer.section(name)
+        @contextmanager
+        def _noop():
+            yield
+        return _noop()
+
     # Assemble the global stiffness matrix with SIMP interpolation
-    K = composer.assemble_stiffness_matrix(
-        basis, rho, E0, Emin, p, nu0, elem_func
-    )
+    with section("assemble"):
+        K = composer.assemble_stiffness_matrix(
+            basis, rho, E0, Emin, p, nu0, elem_func
+        )
     K_csr = K.tocsr()
 
     # Enforce Dirichlet BCs once to obtain the enforced matrix K_e.
[…]
