### 2025-12-02 5829bbb
change elapsed time in legend, total -> average
