+++ scikit-topt/sktopt/tools/timer.py
@@ -505,10 +505,10 @@ class SectionTimer:
         )
         ax.axis("equal")
         if show_total:
-            suffix = f" (total {total:.3f}{unit})" if unit else f" (total {total:.0f})"
+            suffix = f"(sum of slices {total:.3f}{unit})" if unit else f"(sum of slices {total:.0f})"
         else:
             suffix = ""
-        ax.set_title(f"Section timing share{suffix}\n(legend shows each: seconds)")
+        ax.set_title(f"Section timing share {suffix}\n(legend shows each: {unit or 'count'})")
         if show_legend:
             legend_opts = {
                 "bbox_to_anchor": (1.05, 0.5),
