### 2025-12-03 048c6b5
add deaccelerarating scheduler
