+++ scikit-topt/sktopt/tools/__init__.py
@@ -2,6 +2,7 @@ from sktopt.tools.history import HistoryCollection
 from sktopt.tools.scheduler import SchedulerConfig
 from sktopt.tools.scheduler import Schedulers
 from sktopt.tools.scheduler import SchedulerStepAccelerating
+from sktopt.tools.scheduler import SchedulerStepDecelerating
 from sktopt.tools.scheduler import SchedulerStep
 from sktopt.tools.scheduler import SchedulerSawtoothDecay
 from sktopt.tools.timer import SectionTimer
@@ -10,6 +11,7 @@ HistoryCollection.__module__ = __name__
 SchedulerConfig.__module__ = __name__
 Schedulers.__module__ = __name__
 SchedulerStepAccelerating.__module__ = __name__
+SchedulerStepDecelerating.__module__ = __name__
 SchedulerStep.__module__ = __name__
 SchedulerSawtoothDecay.__module__ = __name__
 SectionTimer.__module__ = __name__
@@ -20,6 +22,7 @@ __all__ = [
     "Schedulers",
     "SchedulerStep",
     "SchedulerStepAccelerating",
+    "SchedulerStepDecelerating",
     "SchedulerSawtoothDecay",
     "SectionTimer",
 ]
