### 2025-12-04 3ae3363
Fix compliance sensitivities to scale physical strain energy and add regression tests
