### 2025-12-04 1ee87b2
Fix heat compliance energy scaling and add sensitivity regression tests
