### 2025-12-04 4c5e428
add test for straing energy computation
