### 2025-12-04 8605097
Fix OC KKT residual to use unscaled sensitivities before percentile scaling
