### 2025-12-04 1319a38
Add configurable scaling-rate clipping to OC/LogMOC and refine tutorial box OC settings
