+++ examples/tutorial/box_oc.py
@@ -64,11 +64,14 @@ def get_cfg():
         ),
         # keep volume fixed for stability check
         vol_frac=sktopt.tools.SchedulerConfig.constant(
-            target_value=0.4
-        ),
-        # softer projection
-        beta=sktopt.tools.SchedulerConfig.constant(
             target_value=0.6
+        ),
+        # tighten projection only (others stay constant)
+        beta=sktopt.tools.SchedulerConfig.step_accelerating(
+            init_value=0.6,
+            target_value=4.0,
+            num_steps=4,
+            curvature=2.0
         ),
         filter_radius=sktopt.tools.SchedulerConfig.constant(
             target_value=0.6
