### 2025-12-05 47979ff
remove redundant test
