+++ scikit-topt/sktopt/tools/scheduler.py
@@ -264,6 +264,7 @@ def schedule_sawtooth_decay(
 _lit_schedulers = Literal[
     'Constant',
     'Step', 'StepAccelerating', 'StepDecelerating', 'SawtoothDecay',
+    'StepToOne',
     'None'
 ]
 
@@ -283,15 +284,17 @@ class SchedulerConfig:
         Identifier for the schedule (e.g., "p", "vol_frac").
     init_value : float, optional
         Starting value of the parameter.
-        - Required for Step, StepAccelerating, SawtoothDecay.
+        - Required for Step, StepAccelerating, StepDecelerating, SawtoothDecay.
+        - Automatically inferred for StepToOne.
         - Ignored if Constant.
     target_value : float, optional
         Final value of the parameter.
-        - Required for Step, StepAccelerating, SawtoothDecay.
+        - Required for Step, StepAccelerating, StepDecelerating, SawtoothDecay.
+        - Fixed to 1.0 for StepToOne.
         - Used as fixed value if Constant.
     num_steps : int, optional
         Number of continuation steps or cycles.
-        - Step / StepAccelerating: number of increments between init and target.
+        - Step / StepAccelerating / StepDecelerating / StepToOne: number of increments between init and target.
         - SawtoothDecay: number of sawtooth cycles (restarts).
         - Ignored for Constant.
     iters_max : int, optional
@@ -302,7 +305,7 @@ class SchedulerConfig:
         Shape parameter used in StepAccelerating / StepDecelerating.
         - Example: curvature=2.0 accelerates change near the end.
         - Ignored in other schedulers.
-    scheduler_type : {"Constant", "Step", "StepAccelerating", "StepDecelerating", "SawtoothDecay"}
+    scheduler_type : {"Constant", "Step", "StepAccelerating", "StepDecelerating", "SawtoothDecay", "StepToOne"}
         Scheduling strategy:
           - **Constant**: fixed at ``target_value``.
           - **Step**: discrete continuation from ``init_value`` → ``target_value`` in
@@ -312,6 +315,8 @@ class SchedulerConfig:
           - **SawtoothDecay**: parameter decays linearly from ``init_value`` → ``target_value``
             within each cycle (cycle length = iters_max/num_steps), then resets to
             ``init_value`` at the start of the next cycle.
+          - **StepToOne**: like Step but automatically fixes ``target_value`` to ``1.0`` and
+            infers ``init_value`` from ``num_steps`` (``1.0 / num_steps``).
     """
 
     name: Optional[str] = None
@@ -345,15 +350,18 @@ class SchedulerConfig:
             Identifier for the schedule (e.g., "p", "vol_frac").
         init_value : float, optional
             Starting value of the parameter.
-            - Required for Step, StepAccelerating, SawtoothDecay.
+            - Required for Step, StepAccelerating, StepDecelerating, SawtoothDecay.
+            - Computed automatically for StepToOne.
             - Ignored if Constant.
         target_value : float, optional
             Final value of the parameter.
-            - Required for Step, StepAccelerating, SawtoothDecay.
+            - Required for Step, StepAccelerating, StepDecelerating, SawtoothDecay.
+            - Forced to 1.0 for StepToOne.
             - Used as fixed value if Constant.
         num_steps : int, optional
             Number of continuation steps or cycles.
-            - Step / StepAccelerating: number of increments from init → target.
+            - Step / StepAccelerating / StepDecelerating / StepToOne:
+              number of increments from init → target.
             - SawtoothDecay: number of sawtooth cycles.
             - Ignored if Constant.
         iters_max : int, optional
@@ -364,14 +372,17 @@ class SchedulerConfig:
             Shape parameter for StepAccelerating (default ≈ 2.0).
             Controls acceleration of the step transition.
[…]
