+++ scikit-topt/sktopt/tools/__init__.py
@@ -3,6 +3,7 @@ from sktopt.tools.scheduler import SchedulerConfig
 from sktopt.tools.scheduler import Schedulers
 from sktopt.tools.scheduler import SchedulerStepAccelerating
 from sktopt.tools.scheduler import SchedulerStepDecelerating
+from sktopt.tools.scheduler import SchedulerStepToOne
 from sktopt.tools.scheduler import SchedulerStep
 from sktopt.tools.scheduler import SchedulerSawtoothDecay
 from sktopt.tools.timer import SectionTimer
@@ -12,6 +13,7 @@ SchedulerConfig.__module__ = __name__
 Schedulers.__module__ = __name__
 SchedulerStepAccelerating.__module__ = __name__
 SchedulerStepDecelerating.__module__ = __name__
+SchedulerStepToOne.__module__ = __name__
 SchedulerStep.__module__ = __name__
 SchedulerSawtoothDecay.__module__ = __name__
 SectionTimer.__module__ = __name__
@@ -21,6 +23,7 @@ __all__ = [
     "SchedulerConfig",
     "Schedulers",
     "SchedulerStep",
+    "SchedulerStepToOne",
     "SchedulerStepAccelerating",
     "SchedulerStepDecelerating",
     "SchedulerSawtoothDecay",
