### 2025-12-08 5a1e20c
add scale scheduler for neumann
