+++ scikit-topt/sktopt/tools/__init__.py
@@ -2,7 +2,9 @@ from sktopt.tools.history import HistoryCollection
 from sktopt.tools.scheduler import SchedulerConfig
 from sktopt.tools.scheduler import Schedulers
 from sktopt.tools.scheduler import SchedulerStepAccelerating
+from sktopt.tools.scheduler import SchedulerStepAcceleratingToOne
 from sktopt.tools.scheduler import SchedulerStepDecelerating
+from sktopt.tools.scheduler import SchedulerStepDeceleratingToOne
 from sktopt.tools.scheduler import SchedulerStepToOne
 from sktopt.tools.scheduler import SchedulerStep
 from sktopt.tools.scheduler import SchedulerSawtoothDecay
@@ -12,7 +14,9 @@ HistoryCollection.__module__ = __name__
 SchedulerConfig.__module__ = __name__
 Schedulers.__module__ = __name__
 SchedulerStepAccelerating.__module__ = __name__
+SchedulerStepAcceleratingToOne.__module__ = __name__
 SchedulerStepDecelerating.__module__ = __name__
+SchedulerStepDeceleratingToOne.__module__ = __name__
 SchedulerStepToOne.__module__ = __name__
 SchedulerStep.__module__ = __name__
 SchedulerSawtoothDecay.__module__ = __name__
@@ -25,7 +29,9 @@ __all__ = [
     "SchedulerStep",
     "SchedulerStepToOne",
     "SchedulerStepAccelerating",
+    "SchedulerStepAcceleratingToOne",
     "SchedulerStepDecelerating",
+    "SchedulerStepDeceleratingToOne",
     "SchedulerSawtoothDecay",
     "SectionTimer",
 ]
