+++ scikit-topt/tests/test_scheduler.py
@@ -8,6 +8,8 @@ from sktopt.tools.scheduler import (
     schedule_step_decelerating,
     Scheduler,
     SchedulerConfig,
+    SchedulerStepAcceleratingToOne,
+    SchedulerStepDeceleratingToOne,
 )
 
 
@@ -88,68 +90,111 @@ def test_step_to_one_rejects_non_one_target():
         )
 
 
-def test_plot_schedules_to_local_png():
-    import matplotlib
+def test_step_accel_to_one_sets_target_and_init():
+    cfg = SchedulerConfig.step_accelerating_to_one(
+        name="x", num_steps=5, iters_max=10, curvature=2.0
+    )
+    sched = Scheduler.from_config(cfg)
 
-    matplotlib.use("Agg")  # ensure headless-friendly backend
-    import matplotlib.pyplot as plt
+    assert np.isclose(cfg.target_value, 1.0)
+    assert np.isclose(cfg.init_value, 0.2)
+    assert np.isclose(sched.value(1), 0.2)
+    assert sched.value(10) <= 1.0
 
-    total = 20
-    num_steps = 5
-    xs = np.arange(1, total + 1)
 
-    scheds = {
-        "Step": [
-            schedule_step(
-                it=i,
-                total=total,
-                initial_value=0.0,
-                target_value=1.0,
-                num_steps=num_steps,
-            )
-            for i in xs
-        ],
-        "Accelerating": [
-            schedule_step_accelerating(
-                it=i,
-                total=total,
-                initial_value=0.0,
-                target_value=1.0,
-                num_steps=num_steps,
-                curvature=2.0,
-            )
-            for i in xs
-        ],
-        "Decelerating": [
-            schedule_step_decelerating(
-                it=i,
-                total=total,
-                initial_value=0.0,
-                target_value=1.0,
-                num_steps=num_steps,
-                curvature=2.0,
-            )
-            for i in xs
-        ],
-    }
+def test_step_decel_to_one_sets_target_and_init():
+    cfg = SchedulerConfig.step_decelerating_to_one(
+        name="x", num_steps=5, iters_max=10, curvature=2.0
+    )
+    sched = Scheduler.from_config(cfg)
 
-    # Include the StepToOne scheduler via the factory to mirror real usage.
-    sto_cfg = SchedulerConfig.step_to_one(name="sto", num_steps=num_steps, iters_max=total)
-    sto_sched = Scheduler.from_config(sto_cfg)
-    scheds["StepToOne"] = [sto_sched.value(i) for i in xs]
+    assert np.isclose(cfg.target_value, 1.0)
+    assert np.isclose(cfg.init_value, 0.2)
+    assert np.isclose(sched.value(1), 0.2)
+    assert sched.value(10) <= 1.0
 
-    fig, axes = plt.subplots(2, 2, figsize=(10, 8), sharex=True, sharey=True)
-    axes = axes.ravel()
-    for ax, (name, ys) in zip(axes, scheds.items()):
-        ax.plot(xs, ys, marker="o", linestyle="-")
-        ax.set_title(name)
-        ax.set_xlabel("Iteration")
-        ax.set_ylabel("Value")
-        ax.grid(True)
 
-    outfile = Path(__file__).parent / "scheduler_plots.png"
-    fig.tight_layout()
-    fig.savefig(outfile)
-    plt.close(fig)
+# def test_plot_schedules_to_local_png():
+#     import matplotlib
 
-    assert outfile.exists()
+#     matplotlib.use("Agg")  # ensure headless-friendly backend
+#     import matplotlib.pyplot as plt
+
+#     total = 20
+#     num_steps = 5
[…]
