+++ scikit-topt/sktopt/tools/scheduler.py
@@ -262,9 +262,9 @@ def schedule_sawtooth_decay(
 
 
 _lit_schedulers = Literal[
-    'Constant',
+    'Constant', 'ConstantOne',
     'Step', 'StepAccelerating', 'StepDecelerating', 'SawtoothDecay',
-    'StepToOne',
+    'StepToOne', 'StepAcceleratingToOne', 'StepDeceleratingToOne',
     'None'
 ]
 
@@ -305,9 +305,10 @@ class SchedulerConfig:
         Shape parameter used in StepAccelerating / StepDecelerating.
         - Example: curvature=2.0 accelerates change near the end.
         - Ignored in other schedulers.
-    scheduler_type : {"Constant", "Step", "StepAccelerating", "StepDecelerating", "SawtoothDecay", "StepToOne"}
+    scheduler_type : {"Constant", "ConstantOne", "Step", "StepAccelerating", "StepDecelerating", "SawtoothDecay", "StepToOne", "StepAcceleratingToOne", "StepDeceleratingToOne"}
         Scheduling strategy:
           - **Constant**: fixed at ``target_value``.
+          - **ConstantOne**: fixed at ``1.0`` (alias of Constant with target 1).
           - **Step**: discrete continuation from ``init_value`` → ``target_value`` in
             ``num_steps`` increments.
           - **StepAccelerating**: like Step but transition rate controlled by ``curvature``.
@@ -317,6 +318,10 @@ class SchedulerConfig:
             ``init_value`` at the start of the next cycle.
           - **StepToOne**: like Step but automatically fixes ``target_value`` to ``1.0`` and
             infers ``init_value`` from ``num_steps`` (``1.0 / num_steps``).
+          - **StepAcceleratingToOne**: accelerating variant with fixed target 1.0 and
+            inferred ``init_value`` (``1.0 / num_steps``).
+          - **StepDeceleratingToOne**: decelerating variant with fixed target 1.0 and
+            inferred ``init_value`` (``1.0 / num_steps``).
     """
 
     name: Optional[str] = None
@@ -372,9 +377,10 @@ class SchedulerConfig:
             Shape parameter for StepAccelerating (default ≈ 2.0).
             Controls acceleration of the step transition.
             Ignored in other schedulers.
-        scheduler_type : {"Constant", "Step", "StepAccelerating", "StepDecelerating", "SawtoothDecay", "StepToOne"}, default="Constant"
+        scheduler_type : {"Constant", "ConstantOne", "Step", "StepAccelerating", "StepDecelerating", "SawtoothDecay", "StepToOne", "StepAcceleratingToOne", "StepDeceleratingToOne"}, default="Constant"
             Which scheduling strategy to use:
               - **Constant**: fixed value at ``target_value``.
+              - **ConstantOne**: fixed to 1.0 regardless of provided values.
               - **Step**: discrete continuation from ``init_value`` → ``target_value``
                 over ``num_steps`` stages.
               - **StepAccelerating**: like Step, but interpolation biased by ``curvature``.
@@ -383,6 +389,8 @@ class SchedulerConfig:
                 within each cycle, resetting each time; ``iters_max`` defines total length.
               - **StepToOne**: fixed target of 1.0 with the starting point inferred
                 from the number of steps (``1.0 / num_steps``).
+              - **StepAcceleratingToOne** / **StepDeceleratingToOne**:
+                accelerating / decelerating variants with fixed target 1.0.
 
         Returns
         -------
@@ -409,6 +417,9 @@ class SchedulerConfig:
                 init_value = target_value
             if target_value is None:
                 raise ValueError("Should set target_value")
+        elif scheduler_type == "ConstantOne":
+            init_value = 1.0
+            target_value = 1.0
         elif scheduler_type == "Step":
             if init_value is None:
[…]

[…]
