### 2025-12-12 d8162bb
minor edits on paper.md
