### 2025-12-19 #9092 @kevin-tofu
@mbarzegary

Thanks for checking!
I have updated the Zenodo archive so that:
(1) the title matches the paper exactly, and
(2) you are removed from the author list.
