### 2025-12-19 #9092 @kyleniemeyer
Hi @kevin-tofu, just a few minor things in the paper to fix before publishing:
- "scikit-fem" appears to be the preferred formatting (rather than "Scikit-fem").
- In the references, please wrap words in paper titles using `{` and `}` to preserve capitalization (e.g., `{P}ython`)
- There is an extraneous linebreak between two sentences, can you fix?
> During the optimization process, intermediate density fields are automatically visualized and
saved using PyVista and Matplotlib.
This enables users to monitor the convergence behavior and spatial evolution of the material
distribution.
