### 2025-12-20 876c2a0
Paper: fix scikit-fem naming and reference capitalization
