+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Virtanen2020,
-  title = {SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title = {{SciPy} 1.0: fundamental algorithms for scientific computing in {P}ython},
   volume = {17},
   ISSN = {1548-7105},
   url = {http://dx.doi.org/10.1038/s41592-019-0686-2},
@@ -23,7 +23,7 @@
 }
 
 @article{Bell2023,
-  title = {PyAMG: Algebraic Multigrid Solvers in Python},
+  title = {{PyAMG: Algebraic Multigrid Solvers in Python}},
   volume = {8},
   ISSN = {2475-9066},
   url = {http://dx.doi.org/10.21105/joss.05495},
@@ -46,7 +46,7 @@
 }
 
 @article{Sullivan2019,
-  title = {PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK)},
+  title = {{PyVista}: {3D} plotting and mesh analysis through a streamlined interface for the {Visualization Toolkit} ({VTK})},
   volume = {4},
   ISSN = {2475-9066},
   url = {http://dx.doi.org/10.21105/joss.01450},
@@ -61,7 +61,7 @@
 }
 
 @article{Hunter2007,
-  title = {Matplotlib: A 2D Graphics Environment},
+  title = {{Matplotlib: A 2D Graphics Environment}},
   volume = {9},
   ISSN = {1521-9615},
   url = {http://dx.doi.org/10.1109/MCSE.2007.55},
@@ -76,7 +76,7 @@
 
 @misc{Hunter2007william,
   author = {Hunter, William and others},
-  title = {ToPy - Topology optimization with Python},
+  title = {{ToPy - Topology optimization with Python}},
   year = {2017},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -84,7 +84,7 @@
 }
 
 @article{Jia2024,
-  title = {FEniTop: a simple FEniCSx implementation for 2D and 3D topology optimization supporting parallel computing},
+  title = {{FEniTop}: a simple {FEniCSx} implementation for {2D} and {3D} topology optimization supporting parallel computing},
   volume = {67},
   ISSN = {1615-1488},
   url = {http://dx.doi.org/10.1007/s00158-024-03818-7},
@@ -98,21 +98,21 @@
 }
 
 @inproceedings{tarek2019topoptjl,
-  title={TopOpt.jl: An efficient and high-performance package for topology optimization of continuum structures in the Julia programming language},
+  title={{TopOpt.jl}: {A}n efficient and high-performance package for topology optimization of continuum structures in the {Julia} programming language},
   author={Tarek, Mohamed},
   booktitle={Proceedings of the 13th World Congress of Structural and Multidisciplinary Optimization},
   year={2019}
 }
 
 @inproceedings{huang2021topoptjl,
-  title={TopOpt.jl: Truss and Continuum Topology Optimization, Interactive Visualization, Automatic Differentiation and More},
+  title={{TopOpt.jl: Truss and Continuum Topology Optimization, Interactive Visualization, Automatic Differentiation and More}},
   author={Huang, Yijiang and Tarek, Mohamed},
   booktitle={Proceedings of the 14th World Congress of Structural and Multidisciplinary Optimization},
   year={2021}
 }
 
 @article{Smit2021,
-  title = {Topology optimization using PETSc: a Python wrapper and extended functionality},
+  title = {{T}opology optimization using {PETSc}: a {Python} wrapper and extended functionality},
   volume = {64},
[…]
