### 2025-12-27 b2eb26a
update bibtex in readme
