### 2026-01-19 f3dcb9f
install petsc
